### 2024-07-13 #6979 @editorialbot
[PRE REVIEW]: TurboFlow: Meanline Modelling of Turbomachinery
**Submitting author:** @lasseband (<a href="http://orcid.org/0009-0001-2293-9954">Lasse Borg Anderson</a>)
**Repository:** https://github.com/turbo-sim/TurboFlow
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.11
