### 2024-07-19 523da2b
setup test manager to generate regression data for evaluate_cascade_throat test for linux
