### 2024-07-19 0d97fa4
add kinetic energy loss coefficient
