### 2024-07-19 4dde12a
modify evaluate_cascade_throat choking model
