+++ _sources/source/tutorials/configuration_options_short.rst.txt
@@ -34,6 +34,7 @@ This section describes the different configuration options available in TurboFlo
     These are the valid loss coefficient options:
 
     - `stagnation_pressure`: see :ref:`loss_coefficients` for details.
+    - `kinetic_energy`: see :ref:`loss_coefficients` for details.
 
 .. dropdown:: Deviation Models
     :animate: fade-in-slide-down
