### 2024-07-20 a20bd5e
organize dev_scripts and add script with critical Mach calculations for perfect gas
