+++ dev projects/dev_scripts/deviation_model.py
@@ -1,335 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-import CoolProp.CoolProp as cp
-from scipy.optimize._numdiff import approx_derivative
-import turboflow as tf
-
-from scipy import optimize
-
-fac = 1.00
-
-
-def find_critical_mass_flux(Ma_crit, p0, T0, fluid):
-
-    # Define residual function
-    def get_residual(v):
-        h = h0 - v**2 / 2
-        props = fluid.compute_properties(cp.HmassSmass_INPUTS, h, s0 * fac)
-        res = Ma_crit - v / props["a"]
-        return res
-
-    # Compute the velocity corresponding to critical Mach
-    props_in = fluid.compute_properties(cp.PT_INPUTS, p0, T0)
-    s0 = props_in["s"]
-    h0 = props_in["h"]
-    sol = optimize.root_scalar(get_residual, bracket=[0, 500], method="brentq")
-    v_crit = sol.root
-
-    # Compute the critical mass flux
-    h = h0 - v_crit**2 / 2
-    props = fluid.compute_properties(cp.HmassSmass_INPUTS, h, s0 * fac)
-    mass_flux_crit = v_crit * props["rho"]
-    return mass_flux_crit, v_crit
-
-
-def get_beta_subsonic(v_out, p0, T0, fluid, opening, pitch, radius_curvature, Ma_crit):
-
-    Ma = get_mach(v_out, p0, T0, fluid)
-    beta_g = np.arccos(opening / pitch) * 180 / np.pi
-    beta_0 = 7 / 6 * (beta_g - 10) + 4 * pitch / radius_curvature
-    if Ma < 0.50:
-        beta = beta_0
-
-    elif Ma > 0.50:
-        beta = beta_0 + (beta_g - beta_0) * (2 * Ma - 1) / (2 * Ma_crit - 1)
-
-    return beta
-
-
-def get_Aungier_beta(v_out, p0, T0, fluid, opening, pitch, radius_curvature, Ma_crit):
-
-    beta_g = np.arcsin(opening / pitch) * 180 / np.pi
-    delta_0 = (
-        np.arcsin(opening / pitch * (1 + (1 - opening / pitch) * (beta_g / 90) ** 2))
-        * 180
-        / np.pi
-        - beta_g
-    )
-    Ma = get_mach(v_out, p0, T0, fluid)
-
-    if Ma < 0.50:
-        delta = delta_0
-
-    elif Ma > 0.50 and Ma < Ma_crit:
-        X = (2 * Ma - 1) / (2 * Ma_crit - 1)
-        delta = delta_0 * (1 - 10 * X**3 + 15 * X**4 - 6 * X**5)
-
-    else:
-        delta = 0
-
-    beta = np.arccos(opening / pitch) * 180 / np.pi - delta
-
-    return beta
-
-
-def get_beta_supersonic(v_out, p0, T0, fluid, phi_crit):
-
-    if isinstance(v_out, np.ndarray):
-        v_out = v_out[0]
-
-    props_in = fluid.compute_properties(cp.PT_INPUTS, p0, T0)
-    s_out = props_in["s"]
-    h_out = props_in["h"] - v_out**2 / 2
-    props_out = fluid.compute_properties(cp.HmassSmass_INPUTS, h_out, s_out * fac)
-    rho_out = props_out["rho"]
-    beta = np.arccos(np.min([1, phi_crit / rho_out / v_out])) * 180 / np.pi
-    # beta = np.arccos(phi_crit/rho_out/v_out)*180/np.pi
-
-    return beta
-
-
-def get_mach(v_out, p0, T0, fluid):
[…]
