+++ dev_projects/dev_scripts/deviation_model/parabolic_fitting.py
@@ -24,3 +24,4 @@ x = np.linspace(0.8, 1, 20)
 fig, ax = plt.subplots()
 ax.plot(x, y(x))
 ax.scatter([x1, x2], [y1, y2])
+plt.show()

[…]
