### 2024-07-20 32a78df
add dev scripts about conversion between loss coefficient definitions
