+++ dev_projects/dev_scripts/critical_mach/critical_mach_perfect_gas.py
@@ -95,50 +95,51 @@ def get_mach_crit(gamma, eta):
     return Ma_crit
 
 
+if __name__ == "__main__":
 
-# Plot the critical Mach number against the kinetic energy loss coefficient
-fig = plt.figure(figsize=(6.4, 4.8))
-ax = fig.gca()
-ax.set_xlabel(r"$\Delta \phi^2$ - Kinetic energy loss coefficient")
-ax.set_ylabel(r"$\text{Ma}_\text{crit}$ - Critical Mach number")
-gamma_array = [1.1, 1.2, 1.3, 1.4, 1.5, 1.6]
-colormap = cm.Blues(np.linspace(0.5, 1.0, len(gamma_array)))
-for i, gamma in enumerate(gamma_array):
-    eta = np.logspace(np.log10(1e-9), np.log10(1.00), 100)
+    # Plot the critical Mach number against the kinetic energy loss coefficient
+    fig = plt.figure(figsize=(6.4, 4.8))
+    ax = fig.gca()
+    ax.set_xlabel(r"$\Delta \phi^2$ - Kinetic energy loss coefficient")
+    ax.set_ylabel(r"$\text{Ma}_\text{crit}$ - Critical Mach number")
+    gamma_array = [1.1, 1.2, 1.3, 1.4, 1.5, 1.6]
+    colormap = cm.Blues(np.linspace(0.5, 1.0, len(gamma_array)))
+    for i, gamma in enumerate(gamma_array):
+        eta = np.logspace(np.log10(1e-9), np.log10(1.00), 100)
+        Ma_crit = get_mach_crit(gamma, eta)
+        phi_crit = get_flow_capacity(Ma_crit, gamma, eta)
+        ax.plot(eta, Ma_crit, linestyle='-', color=colormap[i], label=f"$\gamma={gamma}$")
+    ax.legend(loc="lower right", ncol=1, fontsize=11)
+    plt.tight_layout(pad=1)
+    tf.savefig_in_formats(fig, os.path.join(out_dir, "mach_crit_vs_loss_coefficient"))
+
+    # Plot the dimensionless mass function against the Mach number
+    fig = plt.figure(figsize=(6.4, 4.8))
+    ax = fig.gca()
+    ax.set_xlabel(r"$\text{Ma}$ - Mach number")
+    ax.set_ylabel(r"$\Phi = \left(\frac{\dot{m} \sqrt{R T_{0}}}{p_{0}  A}\right)$ - Dimensionless mass flow function")
+    ax.set_xscale("linear")
+    ax.set_yscale("linear")
+
+    gamma = 1.4
+    eta = np.logspace(np.log10(1e-4), np.log10(0.99), 100)
     Ma_crit = get_mach_crit(gamma, eta)
     phi_crit = get_flow_capacity(Ma_crit, gamma, eta)
-    ax.plot(eta, Ma_crit, linestyle='-', color=colormap[i], label=f"$\gamma={gamma}$")
-ax.legend(loc="lower right", ncol=1, fontsize=11)
-plt.tight_layout(pad=1)
-tf.savefig_in_formats(fig, os.path.join(out_dir, "mach_crit_vs_loss_coefficient"))
+    ax.plot(Ma_crit, phi_crit, linestyle='--', color="black")
 
-# Plot the dimensionless mass function against the Mach number
-fig = plt.figure(figsize=(6.4, 4.8))
-ax = fig.gca()
-ax.set_xlabel(r"$\text{Ma}$ - Mach number")
[…]
