### 2024-07-20 c2470ba
update dev projects
