+++ dev projects/Kofskey1974/plot_kofskey1974.py
@@ -21,12 +21,12 @@ if isinstance(cascades_data["operation_points"], list):
 else:
     design_point = cascades_data["operation_points"]
 
-Case = "pressure_line"  # performance_map/error_plot
+Case = "test"  # performance_map/error_plot
 # Get the name of the latest results file
-# filename = ml.utils.find_latest_results_file(RESULTS_PATH)
-filename = (
-    "output\performance_analysis_2024-03-16_20-10-25.xlsx"  # performance map, 1963 REVS
-)
+filename = tf.utils.find_latest_results_file(RESULTS_PATH)
+# filename = 'output\performance_analysis_2024-07-11_11-43-46.xlsx' # New loss coefficient, performance_map
+# filename = 'output\performance_analysis_2024-07-11_20-00-05.xlsx' # New loss coefficient, error_plot
+# filename = "output\performance_analysis_2024-03-16_20-10-25.xlsx"  # performance map, 1963 REVS
 # filename = 'output\performance_analysis_2024-03-16_20-20-51.xlsx' # experimental points 1963 REVS
 
 save_figs = False
@@ -98,24 +98,42 @@ if Case == "pressure_line":
     outdir = "figures"
 
     # Plot mass flow rate
-    # title = "Mass flow rate"
-    # filename = title.lower().replace(" ", "_") + '_' + timestamp
-    # fig1, ax1 = ml.plot_functions.plot_lines(
-    #     data,
-    #     x_key="PR_ts",
-    #     y_keys=["mass_flow_rate"],
-    #     xlabel="Total-to-static pressure ratio",
-    #     ylabel="Mass flow rate [kg/s]",
-    #     title=title,
-    #     filename=filename,
-    #     outdir=outdir,
-    #     color_map=color_map,
-    #     save_figs=save_figs,
-    # )
+    title = "Mass flow rate"
+    filename = title.lower().replace(" ", "_") + '_' + timestamp
+    fig1, ax1 = tf.plot_functions.plot_lines(
+        data,
+        x_key="PR_ts",
+        y_keys=["mass_flow_rate"],
+        xlabel="Total-to-static pressure ratio",
+        ylabel="Mass flow rate [kg/s]",
+        title=title,
+        filename=filename,
+        outdir=outdir,
+        color_map=color_map,
+        save_figs=save_figs,
+    )
+
+    title = "Entropy"
+    filename = title.lower().replace(" ", "_") + '_' + timestamp
+    fig1, ax1 = tf.plot_functions.plot_lines(
+        data,
+        x_key="PR_ts",
+        y_keys=["s_1", "s_throat_1", "s_2"],
+        xlabel="Total-to-static pressure ratio",
+        ylabel="Entropy",
+        title=title,
+        filename=filename,
+        outdir=outdir,
+        color_map=color_map,
+        save_figs=save_figs,
+    )
 
     # Plot mach
     title = "Mach number"
-    linstyles = ["-", "--"]
+    linstyles = ["-", "--"]*2
+    color_map_1 = plt.get_cmap("Reds")(np.linspace(0.2, 1.0, 4))
+    color_map_2 = plt.get_cmap("Blues")(np.linspace(0.2, 1.0, 4))
+    colors = np.concatenate((color_map_1[1:3], color_map_2[1:3]))
     filename = title.lower().replace(" ", "_") + "_" + timestamp
     fig1, ax1 = tf.plot_functions.plot_lines(
         data,
@@ -125,24 +143,7 @@ if Case == "pressure_line":
         ylabel="Cascade exit relative mach number [-]",
         filename=filename,
         outdir=outdir,
-        color_map=color_map,
[…]
