### 2024-07-22 6c0f86c
rename loss coefficient conversion demo directory
