### 2024-07-22 #6979 @lasseband
Hi again @kyleniemeyer

I have reviewed the DOI warnings and corrected several issues. Most references lacking DOIs do not have an associated DOI, but I have added the ones that do.

Additionally, we have discussed potential reviewers and have a few suggestions. Although none of our suggestions are on the list of people who have already agreed to review, we still believe they would be suitable candidates.

- Alfonso del Carre, Skydweller Aero Arbea Campus, Madrid, Spain (Note: Alfonso del Carre is the author of one of the most similar publications.)
- Lasse Mueller, Von Karman Institute for Fluid Dynamics, Rhode-Saint-Genese, Belgium
- Tim Albring, Rheinland-Pfälzische Technische Universität Kaiserslautern-Landau, Kaiserslautern, Germany

Hope this helps in finding suitable reviewers.
