+++ docs/source/todo_list.md
@@ -1,117 +1,0 @@
-
-
-
-
-
-
-## To-do list
-- [x] Verify torque and efficiency deviation match
-- [x] Check if the correlation for incidence losses from Moustapha 1990 is better
-- [x] Check if baseline kacker-okapuu + moustapha 1990 is better
-- [x] Create clean dataset including torque/efficiency/flow angle for each pressure ration/speed
-- [x] Extend validation Kofskey1972 to exit flow angle
-- [x] Add x=y comparison plots for the validation
-- [ ] Verify the displacement thickness value for kofskey1974
-- [ ] Try to extract the shock loss at the inlet fromm the profile loss to have a more clear split of the losses
-- [x] Add better smoother for Mach number constraint
-- [x] Add generic smoothing functions for min/max/abs/piecewise
-- [ ] Replace all non-differentiable functions of the code (specially in the loss models)
-- [x] Improve initial guess computation for supersonic case (call the function `initialize()`)
-- [ ] Validate model with experimental data
-- [x] Add automatic testing (pytest) to ensure that new code pushes do not introduce breaking changes
-- [x] Make function to plot performance maps more general
-- [x] Add functionality to export/import all the parameters corresponding to a calculated operating point
-- [x] Make the plotting of performance maps more general and add functionality to save figures
-- [x] Add environment.yaml file to install all dependencies at once
-- [x] Add Sphinx documentation structure
-- [ ] Implement design optimization
-  - [x] Single-point
-  - [ ] Multi-point
-- [x] Add CI/CD pipeline to run tests and update documentation on new pushes to main branch
-
-
-## To-do list 10.11.2023
-- [x] Residuals and independent variables should be dictionaries
-- [x] Improve initial guess functionality (Lasse)
-  - [x] Move initial guess generation into the CascadesNonlinearSystemProblem() class
-  - [x] Move the solution scaling into the CascadesNonlinearSystemProblem() class [how to re-use scaling for optimization?]
-  - [x] Generation within CascadesNonlinearSystemProblem() class
-  - [x] Initial guess should be a dictionary with keys
-  - [x] Initial guess specification in YAML file (this will be used for the first operation point)
-  - [x] Improve initial guess calculation and extend to multistage
-  - [ ] 1D correlation for Ma_crit as a function of the loss coefficient of the cascade
-- [x] Geometry processing (Roberto)
-  - [x] Clean up code
-  - [x] Improve docstrings
-  - [x] Discuss and improve the throat area calculation (Roberto/Lasse)
-- [ ] Update PySolverView to solve root finding problems with the optimization solver.
-- [x] Add computation time in solver report
-- [ ] Improve robustness with random generation of initial guesses
-  - [ ] Specify the number of initial guesses that will be tried, e.g., 50
-  - [ ] Specify the ranges of variables to be sampled from (enthalpy frac distribution, efficiency ts and tt)
-  - [ ] Use a sampling techniques like latin hypercube, montecarlo, or orthogonal sampling'
-  - [ ] https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.qmc.LatinHypercube.html
-- [ ] Increase robustness allowing CoolProp failures during convergence? but not in the final iteration
-- [x] Add regression tests
-
-## TODO 17.11.2023
-- [x] Add validation of configuration file
-- [x] Add configuration file options to documentation (automated process)
-- [ ] Add logger object to the performance analysis and optimization functions
-- [ ] Log errors and warnings during model evaluation (print after convergence history?)
-- [ ] Improve geometry generation functionality to create full geometry from design optimization variables
[…]
