### 2024-07-22 c15fe38
add option to plot only solutions that improve objective function (gradient-free optimization)
