+++ _sources/source/api/turboflow.pysolver_view.install_local.rst.txt
@@ -1,0 +1,7 @@
+turboflow.pysolver\_view.install\_local module
+==============================================
+
+.. automodule:: turboflow.pysolver_view.install_local
+   :members:
+   :undoc-members:
+   :show-inheritance:
