+++ _sources/source/developer_guide.rst.txt
@@ -1,0 +1,270 @@
+.. _developer_guide:
+
+Developer guide
+=======================
+
+Thank you for considering contributing to this project! Here are some guidelines to help you get started:
+
+Developer Installation Guide
+----------------------------
+
+This installation guide is intended for developers who wish to contribute to or modify the Turboflow source code. It assumes that the developer is using a Linux distribution or Windows with Git Bash terminal to have access to Git and Linux-like commands.
+
+1. **Fork the repository:**
+
+   - Navigate to the `project's GitHub page <https://github.com/turbo-sim/TurboFlow>`_.
+   - Click the "Fork" button in the upper right corner of the repository page to create a copy of the repository under your own GitHub account.
+
+
+2. **Clone the forked repository:**
+
+   - Open your terminal.
+   - Run the following command, replacing `<your-username>` with your GitHub username:
+
+   .. code-block:: bash
+
+      git clone https://github.com/<your-username>/<repository-name>.git
+
+   - Navigate into the cloned repository:
+
+   .. code-block:: bash
+
+      cd <repository-name>
+
+3. **Create a dedicated Conda virtual environment for TurboFlow development**:
+
+   - Check that conda is installed:
+
+   .. code-block:: bash
+
+      conda list
+
+   - If not conda is installed, `install conda <https://conda.io/projects/conda/en/latest/user-guide/install/index.html>`_.
+   - Create dedicated virtual environment for turboflow package:
+
+   .. code-block:: bash
+
+      conda create --name turboflow_env python=3.11
+
+4. **Activate the newly created virtual environment**:
+
+   .. code-block:: bash
+
+      conda activate turboflow_env
+
+5. **Install Poetry to manage dependencies**:
+
+   .. code-block:: bash
+
+      conda install poetry
+
+   Poetry is a powerful dependency manager that offers separation of user and developer dependencies, ensuring that only the necessary packages are installed based on the user's intent. Additionally, it simplifies the process of adding, updating, and removing dependencies, making it easier to maintain the project's requirements.
+
+6. **Use Poetry to install the required dependencies for TurboFlow development**:
+
+   .. code-block:: bash
+
+      poetry install
+
+7. **Verify the installation by running the following command**:
+
+   .. code-block:: bash
+
+      python -c "import turboflow; turboflow.print_package_info()"
+
+   If the installation was successful, you should see the Turboflow banner and package information displayed in the console output.
+
+
+Pull request guidelines
+-------------------------
+
+Please follow these steps to submit a pull request.
+
+1. **Create a branch in your forked repository**:
+
+   - Open your terminal in the projects root.
+   - Create branch:
+
+   .. code-block:: bash
+
+      git checkout -b <feature-name>
+
+2. **Make your changes**:
+
+   - Implement your feature or bugfix.
+
+
+3. **Commit your changes**:
+
+   .. code-block:: bash 
+
+      git commit -m "Description of changes"
+
+4. **Push to your fork**: 
+
+   .. code-block:: bash
+
+      git push origin feature-name
+
+5. **Open a pull request**: 
+
+   - Go to your fork on GitHub and click the "New pull request" button.
+
+Testing guidlines
+-------------------
+
+When implementing new features or adding new submodels, you should aim to create unit tests that verifies the functionality of what you are adding. 
+This testing framework is designed to streamline the process of testing and validating the project using a centralized script to manage tests. 
+The framework includes functions for running tests and generating regression data, and are contained in the `tests` folder:
+
+.. code-block:: text
+
+    tests/
+    ├── config_files/       # Configuration files for tests
+    │   ├── performance_analysis_ainley_mathieson.yaml
+    │   ├── design_optimization.yaml
[…]
