### 2024-07-22 801134d
remove version from index.rst and add it to docs/conf.py instead
