+++ .bumpversion.cfg
@@ -11,6 +11,6 @@ replace = version = "{new_version}"
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
-[bumpversion:file:docs/index.rst]
-search = **Version:** {current_version}
-replace = **Version:** {new_version}
+[bumpversion:file:docs/conf.py]
+search = release = "v{current_version}"
+replace = release = "v{new_version}"
