### 2024-07-22 de3eaa3
Script to add package to PYTHONPATH refactored as function
