+++ turboflow/__init__.py
@@ -1,3 +1,6 @@
+import os
+import sys
+
 # Highlight exception messages
 # https://stackoverflow.com/questions/25109105/how-to-colorize-the-output-of-python-errors-in-the-gnome-terminal/52797444#52797444
 try:
@@ -6,11 +9,8 @@ except ImportError:
     # No IPython. Use default exception printing.
     pass
 else:
-    import sys
-
     sys.excepthook = IPython.core.ultratb.ColorTB()
 
-import toml
 
 # Import submodules
 from .math import *
@@ -31,11 +31,11 @@ from .axial_turbine.design_optimization import *
 set_plot_options()
 
 
-import os
-import toml
-
+__version__ = "0.1.12"
+PACKAGE_NAME = "turboflow"
+URL_GITHUB = "https://github.com/turbo-sim/turboflow"
+URL_DOCS = "https://turbo-sim.github.io/turboflow/"
 BREAKLINE = 80 * "-"
-
 
 def print_banner():
     """Prints a banner."""
@@ -49,9 +49,6 @@ def print_banner():
     print(BREAKLINE)
     print(banner)
 
-__version__ = "0.1.12"
-URL_GITHUB = "https://github.com/turbo-sim/TurboFlow"
-URL_DOCS = "https://turbo-sim.github.io/TurboFlow/"
 
 def print_package_info():
     """Prints package information with predefined values."""
@@ -59,11 +56,47 @@ def print_package_info():
     info = f""" Version:       {__version__}
  Repository:    {URL_GITHUB}
  Documentation: {URL_DOCS}"""
-    
-    # Assuming print_banner and BREAKLINE are defined elsewhere in your code
     print_banner()
     print(BREAKLINE)
     print(info)
+    print(BREAKLINE)
 
 
+def add_package_to_pythonpath():
+    """Append package containing this __file__ to the PYTHONPATH environment variable"""
 
+    # Define bash command to append cwd to PYTHONPATH
+    # PACKAGE_PATH = os.getcwd()
+    # PACKAGE_PATH = os.path.dirname(os.getcwd())
+    PACKAGE_PATH = os.path.dirname(__file__)
+    bashrc_header = f"# Append {PACKAGE_NAME} package to PYTHONPATH"
+    if sys.platform == "win32":  # Windows
+        bashrc_line = f'export PYTHONPATH=$PYTHONPATH\;"{PACKAGE_PATH}"'
+    else:  # Linux or MacOS
+        bashrc_line = f'export PYTHONPATH=$PYTHONPATH:"{PACKAGE_PATH}"'
+
+    # Locate the .bashrc file
+    bashrc_path = os.path.expanduser("~/.bashrc")
+
+    # Ask for user confirmation with default set to no
+    response = input(
+        f"Do you want to add the {PACKAGE_NAME} path to your .bashrc? [yes/NO]: "
+    )
+    if response.lower() in ["y", "yes"]:
+        try:
+            # Check if the line already exists in the .bashrc
+            with open(bashrc_path, "r") as file:
+                if bashrc_line in file.read():
+                    print(f".bashrc already contains the {PACKAGE_NAME} path.")
+                else:
+                    with open(bashrc_path, "a") as file_append:
+                        file_append.write(f"\n{bashrc_header}")
+                        file_append.write(f"\n{bashrc_line}\n")
+                    print(f"Path to {PACKAGE_NAME} package added to .bashrc")
+                    print(
+                        f"Restart the terminal or run 'source ~/.bashrc' for the changes to take effect."
+                    )
+        except Exception as e:
+            print(f"An error occurred: {e}")
+    else:
+        print("Operation aborted by user.")
