+++ dev_projects/dev_scripts/critical_mach/critical_mach_perfect_gas.py
@@ -11,90 +11,6 @@ out_dir = "figures"
 if not os.path.exists(out_dir):
     os.makedirs(out_dir)
 
-
-def get_flow_capacity(Ma, gamma, eta):
-    """
-    Compute dimensionless mass flow rate for non-isentropic flow in a nozzle.
-
-    This function calculates the dimensionless mass flow rate ($\Phi$) using the 
-    given Mach number ($\text{Ma}$), specific heat ratio ($\gamma$), and efficiency ($\eta$).
-    
-    .. math::
-        \Phi = \left(\frac{\dot{m} \sqrt{R T_{0}}}{p_{0}  A}\right) = \left(\frac{2 \gamma}{\gamma-1}\right)^{1 / 2} \cdot \frac{\sqrt{1-\hat{T}}}{\hat{T}}
-        \left[1-\frac{1}{\eta}(1-\hat{T})\right]^{\frac{\gamma}{\gamma-1}}
-
-    where:
-
-    .. math::
-        \hat{T} = \left(\frac{T}{T_{0}}\right)
-
-    and
-
-    .. math::
-        \eta = 1 - \Delta \phi^2
-
-    Here, $\Delta \phi^2$ is the kinetic energy loss coefficient.
-    
-    Parameters
-    ----------
-    Ma : float
-        Mach number of the flow.
-    gamma : float
-        Specific heat ratio of the gas.
-    eta : float
-        Nozzle efficiency, related to the kinetic energy loss coefficient.
-
-    Returns
-    -------
-    float
-        Dimensionless mass flow rate (\Phi).
-    """
-    T_hat = (1 + (gamma - 1) / 2 * Ma**2) ** (-1)
-    Phi = np.sqrt(2 * gamma / (gamma - 1)) * np.sqrt(1 - T_hat) / T_hat * (1 - 1 / eta * (1 - T_hat)) ** (gamma / (gamma - 1))
-    return Phi
-
-
-def get_mach_crit(gamma, eta):
-    """
-    Compute the critical Mach number for non-isentropic flow in a nozzle.
-
-    This function calculates the critical Mach number ($\text{Ma}_\text{crit}$) using the 
-    given specific heat ratio (\$gamma$) and nozzle efficiency ($\eta$).
-    
-    .. math::
-        \text{Ma}_\text{crit}^{2} = \left(\frac{2}{\gamma-1}\right)\left(\frac{1}{\hat{T}_\text{crit}}-1\right) 
-        = \left(\frac{2}{\gamma-1}\right)\left[\frac{4 \alpha-2}{(2 \alpha+\eta-3) + \sqrt{(1+\eta)^{2}+4 \alpha(1+\alpha-3 \eta)}} - 1\right]
-
-    where:
-
-    .. math::
-        \alpha = \frac{\gamma}{\gamma-1}
-
-    and
-
-    .. math::
-        \eta = 1 - \Delta \phi^2
-
-    Here, $\Delta \phi^2$ is the kinetic energy loss coefficient.
-        
-    Parameters
-    ----------
-    gamma : float
-        Specific heat ratio of the gas.
-    eta : float
-        Nozzle efficiency, related to the kinetic energy loss coefficient.
-
-    Returns
-    -------
-    float
-        Critical Mach number (\text{Ma}_\text{crit}).
-    """
-    alpha = gamma / (gamma - 1)
-    T_hat_crit = (2*alpha + eta - 3 + np.sqrt((1+eta)**2 + 4*alpha*(1+alpha-3*eta))) / (4*alpha - 2)
[…]
