+++ docs/source/model_description.rst
@@ -19,3 +19,5 @@ The TurboFlow meanline-model is designed to be flexible and offers a variety of 
    model_description/deviation_models
    model_description/choking_models
 
+
+   .. model_description/equation_formulation
