+++ docs/source/api_reference.rst
@@ -1,23 +1,0 @@
-API Reference
-=================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   api/turboflow.axial_turbine
-   api/turboflow.properties
-   api/turboflow.pysolver_view
-   api/turboflow.utilities
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 4
-
-   api/turboflow.config_validation
-   api/turboflow.math
-   api/turboflow.plot_functions
