### 2024-07-23 b941cb7
update the smooth minimum and smooth maximum functions
