### 2024-07-23 86c84c3
update critical Mach number docstrings
