### 2024-07-25 38c8e16
update pysolverview docstring
