### 2024-07-25 e099169
add initial guess for perfomance analysis to configurations
