+++ tests/config_files/performance_analysis_evaluate_cascade_isentropic_throat.yaml
@@ -23,6 +23,10 @@ performance_analysis:
     derivative_abs_step: 1e-6
     print_convergence: True
     plot_convergence: False
+  initial_guess :
+    eta_tt : [0.9, 0.8]
+    eta_ts : [0.8, 0.7]
+    ma : [0.8, 0.8]
 
 simulation_options:
   deviation_model : aungier
