+++ tests/config_files/performance_analysis_zero_deviation.yaml
@@ -22,7 +22,11 @@ performance_analysis:
     derivative_method: "2-point"
     derivative_abs_step: 1e-6
     print_convergence: True
-    plot_convergence: False 
+    plot_convergence: False
+  initial_guess :
+    eta_tt : [0.9, 0.8]
+    eta_ts : [0.8, 0.7]
+    ma : [0.8, 0.8]
 
 simulation_options:
   deviation_model : zero_deviation
