### 2024-07-25 c216971
avoid printing configuration summary during testing
