### 2024-07-25 8f1cca5
update test manager to generate new regression data
