+++ tests/tests_manager.py
@@ -30,7 +30,15 @@ TEST_CONFIGS = {
 
 # Dictionary of configuration files for regression data generation
 REGRESSION_CONFIGS = {
-    'performance_analysis': ["performance_analysis_evaluate_cascade_throat.yaml"],
+    'performance_analysis': [
+                                "performance_analysis_evaluate_cascade_throat.yaml",
+                                "performance_analysis_evaluate_cascade_critical.yaml",
+                                "performance_analysis_evaluate_cascade_isentropic_throat.yaml",
+                                "performance_analysis_kacker_okapuu.yaml",
+                                "performance_analysis_moustapha.yaml",
+                                "performance_analysis_zero_deviation.yaml",
+                                "performance_analysis_ainley_mathieson.yaml",
+                            ],
     'design_optimization': None,
 }
 
