### 2024-07-25 2b83db9
new initial guess strategy
