+++ turboflow/axial_turbine/performance_analysis.py
@@ -15,6 +15,11 @@ from .. import utilities as utils
 from .. import properties as props
 from . import geometry_model as geom
 from . import flow_model as flow
+from . import choking_model as ch
+from . import deviation_model as dm
+from scipy.stats import qmc
+from scipy import optimize
+
 
 SOLVER_MAP = {"lm": "Lavenberg-Marquardt", "hybr": "Powell's hybrid"}
 """
@@ -22,56 +27,56 @@ Available solvers for performance analysis.
 """
 
 
-def get_heuristic_guess_input(n):
-    r"""
+# def get_heuristic_guess_input(n):
+#     r"""
 
-    Generate a list of `n` number of different sets of dictionaries used to generate initial guesses.
+#     Generate a list of `n` number of different sets of dictionaries used to generate initial guesses.
 
-    Total-to-total efficiency, total-to-static efficiency, enthalpy loss fractions for each cascade and critical mach can be used to generate
-    an initial guess for performance analysis. This function generate a list of such different sets used to generate a full initial guess through the function `compute_heuristic_initial_guess`.
+#     Total-to-total efficiency, total-to-static efficiency, enthalpy loss fractions for each cascade and critical mach can be used to generate
+#     an initial guess for performance analysis. This function generate a list of such different sets used to generate a full initial guess through the function `compute_heuristic_initial_guess`.
 
-    The total-to-static efficiency, varies between  0.6, 0.7, or 0.8, while the total-to-total efficiency can be 0.7, 0.8 or 0.9. The enthalpy loss fractions can either split equally between the cascade,
-    or with equal increments for neighbouring values (see `utils.fill_array_with_increment` for more information). The critical mach is assumed to be 0.95.
+#     The total-to-static efficiency, varies between  0.6, 0.7, or 0.8, while the total-to-total efficiency can be 0.7, 0.8 or 0.9. The enthalpy loss fractions can either split equally between the cascade,
+#     or with equal increments for neighbouring values (see `utils.fill_array_with_increment` for more information). The critical mach is assumed to be 0.95.
 
-    Parameters
-     ----------
-     n : int
-         Number of cascades.
+#     Parameters
+#      ----------
+#      n : int
+#          Number of cascades.
 
-     Returns
-     -------
-     list
-         List of sets of variables used to generate an initial guess.
+#      Returns
+#      -------
+#      list
+#          List of sets of variables used to generate an initial guess.
 
-    """
+#     """
 
-    eta_ts_vec = [0.8, 0.7, 0.6]
+#     eta_ts_vec = [0.8, 0.7, 0.6]
 
-    array = utils.fill_array_with_increment(n)
-    enthalpy_distributions = []
-    enthalpy_distributions.append(np.ones(n) * 1 / n)
-    enthalpy_distributions.append(array)
-    enthalpy_distributions.append(np.flip(array))
+#     array = utils.fill_array_with_increment(n)
+#     enthalpy_distributions = []
+#     enthalpy_distributions.append(np.ones(n) * 1 / n)
+#     enthalpy_distributions.append(array)
+#     enthalpy_distributions.append(np.flip(array))
 
-    initial_guesses = []
-    for eta_ts in eta_ts_vec:
-        for enthalpy_distribution in enthalpy_distributions:
-            initial_guesses.append(
-                {
-                    "enthalpy_loss_fractions": enthalpy_distribution,
-                    "eta_ts": eta_ts,
-                    "eta_tt": eta_ts + 0.1,
-                    "Ma_crit": 0.95,
-                }
-            )
+#     initial_guesses = []
+#     for eta_ts in eta_ts_vec:
+#         for enthalpy_distribution in enthalpy_distributions:
[…]
