### 2024-07-25 59141aa
dev script for new initial guess strategy
