+++ dev_projects/dev_scripts/initial_guess/kofskey1972_1stage.yaml
@@ -1,0 +1,240 @@
+turbomachinery: axial_turbine
+operation_points: 
+  fluid_name: air
+  T0_in: 295.6
+  p0_in: 13.8e4
+  p_out : 13.8e4/2.298 
+  omega: 1627
+  alpha_in: 0
+
+simulation_options: # Optional
+  deviation_model : aungier  # Optional
+  choking_model : evaluate_cascade_throat # evaluate_cascade_critical, evaluate_cascade_throat, evaluate_cascade_isentropic_throat  # Optional
+  rel_step_fd: 1e-4  # Optional
+  loss_model:  # Optional
+    model: benner  # Optional
+    loss_coefficient: stagnation_pressure  # Optional
+  
+performance_analysis :
+  performance_map: # Optional
+    fluid_name: air 
+    T0_in: 295.6
+    p0_in: 13.8e4
+    p_out: 13.8e4/np.linspace(1.6, 4.5, 40)
+    omega: 1627
+    alpha_in: 0 
+  solver_options: # Optional
+    method: hybr  # Optional
+    tolerance: 1e-8  # Optional
+    max_iterations: 100  # Optional
+    derivative_method: 2-point  # Optional
+    derivative_abs_step: 1e-6  # Optional
+    print_convergence: False
+    plot_convergence: False
+
+geometry:
+  cascade_type: ["stator", "rotor"]
+  radius_hub_in: [0.084785, 0.084785]
+  radius_hub_out: [0.084785, 0.081875]
+  radius_tip_in: [0.118415, 0.118415]
+  radius_tip_out: [0.118415, 0.121325]
+  pitch: [1.8294e-2, 1.524e-2]
+  chord: [2.616e-2, 2.606e-2]
+  stagger_angle: [+43.03, -31.05]
+  opening: [0.747503242e-2, 0.735223377e-2]
+  leading_edge_angle : [0.00, 29.60]
+  leading_edge_wedge_angle : [50.00, 50.00]
+  leading_edge_diameter : [2*0.127e-2, 2*0.081e-2]
+  trailing_edge_thickness : [0.050e-2, 0.050e-2]
+  maximum_thickness : [0.505e-2, 0.447e-2]
+  tip_clearance: [0.00, 0.030e-2]
+  throat_location_fraction: [1, 1]
+  
+
+design_optimization:
+  objective_function: 
+    variable : overall.efficiency_ts
+    scale : 1
+    type : maximize
+  variables :
+    specific_speed :
+      value : 0.660834348
+      lower_bound : 0.01
+      upper_bound : 10
+    blade_jet_ratio :
+      value : 0.416813303
+      lower_bound : 0.1
+      upper_bound : 0.9
+    hub_tip_ratio_in :
+      value : [0.715998818, 0.715998818]
+      lower_bound : [0.6, 0.6]
+      upper_bound : [0.9, 0.9]
+    hub_tip_ratio_out :
+      value : [0.715998818, 0.674840305]
+      lower_bound : [0.6, 0.6]
+      upper_bound : [0.9, 0.9]
+    aspect_ratio :
+      value : [1.2855504587156, 1.40214888718342]
+      lower_bound : [1.0, 1.0]
+      upper_bound : [2.0, 2.0]
+    pitch_chord_ratio :
+      value : [0.699311926605505, 0.584804298]
+      lower_bound : [0.5, 0.5]
+      upper_bound : [1.10, 1.10]
+    trailing_edge_thickness_opening_ratio :
+      value : [0.066889342, 0.068006543]
+      lower_bound : [0.05, 0.05]
+      upper_bound : [0.4, 0.4]
[…]
