### 2024-07-25 d39e306
dev script for genetic algorthim
