+++ dev_projects/Kofskey1972_one_stage/kofskey1972_1stage.yaml
@@ -3,15 +3,15 @@ operation_points:
   - fluid_name: air
     T0_in: 295.6
     p0_in: 13.8e4
-    # p_out : 13.8e4/2.298 #, 13.8e4/2.5]
-    p_out : 52252.42718
+    p_out : 13.8e4/2.298 #, 13.8e4/2.5]
+    # p_out : 52252.42718
     omega: 1627
     alpha_in: 0
 
 simulation_options: # Optional
   deviation_model : aungier  # Optional
   # blockage_model : 0.00  #flat_plate_turbulent  # Optional
-  choking_model : evaluate_cascade_throat # evaluate_cascade_critical, evaluate_cascade_throat, evaluate_cascade_isentropic_throat  # Optional
+  choking_model : evaluate_cascade_isentropic_throat # evaluate_cascade_critical, evaluate_cascade_throat, evaluate_cascade_isentropic_throat  # Optional
   rel_step_fd: 1e-4  # Optional
   loss_model:  # Optional
     model: benner  # Optional
@@ -29,7 +29,8 @@ performance_analysis :
     fluid_name: air 
     T0_in: 295.6
     p0_in: 13.8e4
-    p_out: 13.8e4/np.linspace(1.6, 4.5, 40)
+    # p_out: 13.8e4/np.linspace(1.6, 4.5, 40)
+    p_out: 13.8e4/np.linspace(1.6, 1.8, 2)
     omega: 1627
     alpha_in: 0 
   solver_options: # Optional
@@ -40,6 +41,18 @@ performance_analysis :
     derivative_abs_step: 1e-6  # Optional
     print_convergence: True
     plot_convergence: False
+  # initial_guess :
+  #   n_samples : 100
+  #   eta_tt : (0.9, 0.99)
+  #   eta_ts : (0.7, 0.89)
+  #   ma : (0.5, 1.2)
+  initial_guess :
+    eta_ts : [0.9, 0.8] 
+    eta_tt : [0.8, 0.7]
+    ma : [0.8, 0.8]
+
+  
+
 
 geometry:
   cascade_type: ["stator", "rotor"]
@@ -67,39 +80,39 @@ design_optimization:
     type : maximize
   variables :
     specific_speed :
-      value : 1.2
+      value : 0.660834348
       lower_bound : 0.01
       upper_bound : 10
     blade_jet_ratio :
-      value : 0.5
+      value : 0.416813303
       lower_bound : 0.1
       upper_bound : 0.9
     hub_tip_ratio_in :
-      value : [0.6, 0.6]
+      value : [0.715998818, 0.715998818]
       lower_bound : [0.6, 0.6]
       upper_bound : [0.9, 0.9]
     hub_tip_ratio_out :
-      value : [0.6, 0.6]
+      value : [0.715998818, 0.674840305]
       lower_bound : [0.6, 0.6]
       upper_bound : [0.9, 0.9]
     aspect_ratio :
-      value : [1.5, 1.5]
+      value : [1.2855504587156, 1.40214888718342]
       lower_bound : [1.0, 1.0]
       upper_bound : [2.0, 2.0]
     pitch_chord_ratio :
-      value : [0.9, 0.9]
-      lower_bound : [0.75, 0.75]
+      value : [0.699311926605505, 0.584804298]
+      lower_bound : [0.5, 0.5]
       upper_bound : [1.10, 1.10]
     trailing_edge_thickness_opening_ratio :
-      value : [0.1, 0.1]
+      value : [0.066889342, 0.068006543]
       lower_bound : [0.05, 0.05]
       upper_bound : [0.4, 0.4]
     leading_edge_angle :
-      value : [0.41, 0.5]
[…]
