### 2024-07-26 57aae82
add note about efficiency definitions
