+++ docs/dev_notes/efficiency_definitions.md
@@ -1,0 +1,32 @@
+## Efficiency definitions
+
+Total-to-total esficiency definition
+$$
+\eta_{\text{tt}}=\frac{h_{01}-h_{02}}{h_{01}-h_{02s}}=\frac{h_{01}-h_{02}}{h_{01}-h_{2s}-{v_2^2}/{2}}
+$$
+
+Total-to-static esficiency definition
+$$
+\eta_{\text{ts}}=\frac{h_{01}-h_{02}}{h_{01}-h_{0s}}
+$$
+
+Taking the inverse of $\eta_{\text{tt}}$ we see that:
+$$
+\frac{1}{\eta_{\text{tt}}}=\frac{h_{01}-h_{2s}-{v_2^2}/{2}}{h_{01}-h_{02}}=\frac{1}{\eta_{\text{ts}}}-\frac{v_2^2 / 2}{h_{01}-h_{02}} = \frac{1}{\eta_{\text{ts}}} - \Delta \eta_{ke}
+$$
+
+Solving for the total-to-total efficiency we get:
+
+$$
+\eta_{\text{tt}}=\frac{\eta_{\text{ts}}}{(1-  \eta_{\text{ts}} \,\Delta \eta_{ke})} 
+$$
+
+
+Alternatively, solving for the total-to-static efficiency we get
+
+$$
+\eta_{\text{ts}}=\frac{\eta_{\text{tt}}}{(1 + \eta_{\text{tt}} \,\Delta \eta_{ke})} 
+$$
+
+From these formulas it is clear that $\eta_{\text{tt}} > \eta_{\text{ts}}$ for $\Delta \eta_{ke}>0$.
+
