### 2024-07-29 feeb86b
Modify nonlinear solver to raise exception when the residual is not actually zero
