### 2024-07-30 01ed491
rename choking criterion models
