+++ turboflow/axial_turbine/__init__.py
@@ -3,5 +3,5 @@ from .design_optimization import *
 from .deviation_model import *
 from .geometry_model import *
 from .flow_model import *
-from .choking_model import *
+from .choking_criterion import *
 from .loss_model import *
