### 2024-07-30 2226254
rename test files and update config files
