+++ tests/config_files/design_optimization.yaml
@@ -11,7 +11,7 @@ operation_points:
 simulation_options:
   deviation_model : zero_deviation
   blockage_model : 0.00  #flat_plate_turbulent
-  choking_model : evaluate_cascade_isentropic_throat # evaluate_cascade_critical, evaluate_cascade_throat, evaluate_cascade_isentropic_throat
+  choking_criterion : critical_isentropic_throat # evaluate_cascade_critical, evaluate_cascade_throat, evaluate_cascade_isentropic_throat
   rel_step_fd: 1e-4
   loss_model:
     model: kacker_okapuu
@@ -131,9 +131,10 @@ design_optimization:
   solver_options:
     library: scipy
     method: slsqp
-    tolerance: 1e-6
-    max_iterations: 100
     derivative_method: 2-point
     print_convergence: True
     plot_convergence: False
     update_on: "gradient"
+    options : 
+      tol: 1e-6
+      max_iter: 100
