+++ turboflow/axial_turbine/choking_criterion.py
@@ -1,0 +1,762 @@
+from scipy.optimize._numdiff import approx_derivative
+from scipy import interpolate
+import numpy as np
+
+from .. import math
+from . import flow_model as fm
+from . import deviation_model as dm
+
+CHOKING_CRITERIONS = [
+    "critical_mass_flow_rate",
+    "critical_mach_number",
+    "critical_isentropic_throat",
+]
+
+REGRESSION_COEFFICIENTS = {'air' : [0.999984354, -0.368037583, 0.202442269, -0.096333086, 0.023783628],
+                           'R125' : [0.999983586, -0.319079827, 0.165487456, -0.078246564, 0.01996359]}
+
+
+def evaluate_choking(
+    choking_input,
+    inlet_plane,
+    exit_plane,
+    fluid,
+    geometry,
+    angular_speed,
+    model_options,
+    reference_values,
+):
+    r"""
+
+    Calculate condition for choking and evaluate wheter or not the cascade is choked, based on selected choking model.
+
+    Parameters
+    ----------
+    choking_input : dict
+        Dictionary containing necessary input parameters required for selected choking model. See each models documentation for specifics.
+    inlet_plane : dict
+        Dictionary containing data on the inlet plane for the actual cascade operating condition.
+    exit_plane : dict
+        Dictionary containing data on the exit plane for the actual cascade operating condition.
+    fluid : object
+        A fluid object with methods for thermodynamic property calculations.
+    geometry : dict
+        Geometric parameters of the cascade.
+    angular_speed : float
+        Angular speed of the cascade.
+    model_options : dict
+        Options for the model used to evaluate choking.
+    reference_values : dict
+        Reference values used in the calculation, including the reference mass flow rate.
+
+    Returns
+    -------
+    dict
+        Dictionary containing the residuals of choking model
+    dict
+        Dictionary containing state information on the critical conditions.
+
+    Raises
+    ------
+    ValueError
+        If an invalid choking model is provided.
+
+    """
+
+    critical_cascade_functions = {
+        CHOKING_CRITERIONS[0]: critical_mass_flow_rate,
+        CHOKING_CRITERIONS[1]: critical_mach_number,
+        CHOKING_CRITERIONS[2]: critical_isentropic_throat,
+    }
+
+    # Evaluate loss model
+    model = model_options["choking_criterion"]
+    if model in critical_cascade_functions:
+        return critical_cascade_functions[model](
+            choking_input,
+            inlet_plane,
+            exit_plane,
+            fluid,
+            geometry,
+            angular_speed,
+            model_options,
+            reference_values,
+        )
+    else:
+        options = ", ".join(f"'{k}'" for k in CHOKING_CRITERIONS)
+        raise ValueError(
+            f"Invalid critical cascade model '{model}'. Available options: {options}"
+        )
+
+def critical_mach_number(
+    choking_input,
+    inlet_plane,
+    exit_plane,
+    fluid,
+    geometry,
+    angular_speed,
+    model_options,
+    reference_values,
+):
+    r"""
+
+    Calculate condition for choking and evaluate wheter or not the cascade is choked, based on the critical_mach_number choking model.
+
+    This choking model evaluates the cascade throat and checks if the throat mach number exceed the critical. The critical mach number is calculated
[…]
