+++ dev_projects/Kofskey1972_one_stage/plot_kofskey1972_1stage.py
@@ -652,123 +652,131 @@ elif Case == "test":
     # filename_1 = "output/performance_analysis_2024-07-20_17-39-40.xlsx" # Lagrange, 5/6
     # filename_2 = "output/performance_analysis_2024-07-20_17-33-26.xlsx" # Critical mach, Madrid correlation, 5/6
     # filename_3 = "output/performance_analysis_2024-07-20_17-47-54.xlsx" # Critical mach, Lasse correlation, 5/6
-    filename_1 = "output/performance_analysis_2024-07-20_20-01-58.xlsx" # Lagrange, 1.0
+    # filename_1 = "output/performance_analysis_2024-07-20_20-01-58.xlsx" # Lagrange, 1.0
     # filename_2 = "output/performance_analysis_2024-07-20_19-57-29.xlsx" # Critical mach, Lasse correlation, 1.0
-    filename_2 = "output/performance_analysis_2024-07-20_20-05-38.xlsx" # Critical mach, Madrid correlation, 1.0
+    # filename_2 = "output/performance_analysis_2024-07-20_20-05-38.xlsx" # Critical mach, Madrid correlation, 1.0
+    filename_1 = "output/performance_map_evaluate_cascade_throat.xlsx"
+    filename_2 = "output/performance_analysis_2024-03-13_23-14-55.xlsx"
 
     data_1 = tf.plot_functions.load_data(filename_1)
     data_2 = tf.plot_functions.load_data(filename_2)
-    # data_3 = tf.plot_functions.load_data(filename_3)
 
-    subsets = ["omega", 1627]
+    subsets = ["omega"] + list(1627*np.array([0.7, 1.0]))
+    colors = plt.get_cmap('magma')(np.linspace(0.2, 0.7, 2))
     fig1, ax1 = tf.plot_functions.plot_lines(
         data_1,
         x_key="PR_ts",
         y_keys=["mass_flow_rate"],
+        subsets=subsets,
         xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
         ylabel="Mass flow rate [kg/s]",
-        colors='k',
+        colors=colors,
+        linestyles=['-']*2,
         filename = 'mass_flow_rate',
         outdir = "figures",
-        save_figs=True,
+        save_figs=False,
     )
 
     fig1, ax1 = tf.plot_functions.plot_lines(
         data_2,
         x_key="PR_ts",
         y_keys=["mass_flow_rate"],
+        subsets=subsets,
         xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
         ylabel="Mass flow rate [kg/s]",
-        colors=['r'],
+        colors=colors,
+        linestyles=[":"]*2,
         fig = fig1,
         ax = ax1, 
-        linestyles = ["--"],
         filename = 'mass_flow_rate',
         outdir = "figures",
-        save_figs=True,
+        save_figs=False,
     )
 
-    # fig1, ax1 = tf.plot_functions.plot_lines(
-    #     data_3,
-    #     x_key="PR_ts",
-    #     y_keys=["mass_flow_rate"],
-    #     xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
-    #     ylabel="Mass flow rate [kg/s]",
-    #     colors=['r'],
-    #     fig = fig1,
-    #     ax = ax1, 
-    #     linestyles = ["--"],
-    #     filename = 'mass_flow_rate',
-    #     outdir = "figures",
-    #     save_figs=True,
-    # )
[…]
