+++ dev_projects/Kofskey1974/plot_kofskey1974.py
@@ -712,66 +712,136 @@ elif Case == "error_plot":
 elif Case == "test":
 
     # Load performance data
-    filename_1 = "output/performance_analysis_2024-03-16_20-10-25.xlsx" # Lagrange, 5/6
-    filename_2 = "output/performance_analysis_2024-07-20_20-09-13.xlsx" # Critical mach, Madrid correlation, 5/6
-    # filename_1 = "output/performance_analysis_2024-07-20_21-07-30.xlsx" # Lagrange, 1.0
-    # filename_2 = "output/performance_analysis_2024-07-20_21-01-37.xlsx" # Critical mach, Madrid correlation, 1.0
+    filename_1 = "output/performance_map_evaluate_cascade_throat.xlsx"
+    filename_2 = "output/performance_analysis_2024-03-16_20-10-25.xlsx"
 
     data_1 = tf.plot_functions.load_data(filename_1)
-    sheets = data_1.keys()
-    filtered_indices = data_1["overall"][data_1["overall"]["angular_speed"] == 1963].index
-    for sheet in sheets:
-        data_1[sheet] = data_1[sheet].loc[filtered_indices, :]
     data_2 = tf.plot_functions.load_data(filename_2)
-    # data_3 = tf.plot_functions.load_data(filename_3)
 
-    subsets = ["omega", 1627]
+    subsets = ["omega"] + list(1963*np.array([0.7, 1.0]))
+    linestyles = ["-", ":"]
+    colors = plt.get_cmap('magma')(np.linspace(0.2, 0.7, 2))
     fig1, ax1 = tf.plot_functions.plot_lines(
         data_1,
         x_key="PR_ts",
         y_keys=["mass_flow_rate"],
+        subsets=subsets,
         xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
         ylabel="Mass flow rate [kg/s]",
-        colors='k',
+        colors=colors,
+        # color_map = 'magma',
+        linestyles=['-']*2,
         filename = 'mass_flow_rate',
         outdir = "figures",
-        save_figs=True,
+        save_figs=False,
     )
 
     fig1, ax1 = tf.plot_functions.plot_lines(
         data_2,
         x_key="PR_ts",
         y_keys=["mass_flow_rate"],
+        subsets=subsets,
         xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
         ylabel="Mass flow rate [kg/s]",
-        colors=['r'],
+        linestyles=[':']*2,
+        colors=colors, 
+        # color_map = 'Blues',       
         fig = fig1,
         ax = ax1, 
-        linestyles = ["--"],
         filename = 'mass_flow_rate',
         outdir = "figures",
-        save_figs=True,
+        save_figs=False,
     )
 
-    ax1.legend(labels = ["Maxmized mass flow rate", "Critical mach"], loc = "lower right")
-    ax1.set_ylim([1.8, 2.4])
-    ax1.set_title("Kofskey 1974")
-    fig1.tight_layout(pad=1, w_pad=None, h_pad=None)
+    fig2, ax2 = tf.plot_functions.plot_lines(
+        data_1,
+        x_key="PR_ts",
+        y_keys=["torque"],
+        subsets=subsets,
+        xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+        ylabel="Torque [Nm]",
+        colors=colors,
+        # color_map = 'magma',
+        linestyles=['-']*2,
+        filename = 'torque',
+        outdir = "figures",
[…]
