### 2024-07-30 64690a7
restructure optimization options
