+++ turboflow/pysolver_view/optimization_wrappers.py
@@ -35,10 +35,19 @@ PYGMO_SOLVERS = [
     "ipopt",
     "snopt",
     # "worhp",
+    "de",
+    "sga",
     "pso",
 ]
 """List of valid Pygmo solvers"""
 
+GENETIC_SOLVERS = [
+    "de",
+    "sga",
+    "pso",
+]
+
+GRADIENT_SOLVERS = list(set(PYGMO_SOLVERS + SCIPY_SOLVERS) - set(GENETIC_SOLVERS))
 
 NLOPT_SOLVERS = [
     "auglag",
@@ -100,9 +109,8 @@ def minimize_scipy(problem, x0, method, options):
         raise ValueError(error_message)
 
     # Define the options dictionary
-    options = options.copy()  # Work with a copy to avoid side effects
-    tol = options.pop("tol")
-    max_iter = options.pop("max_iter")
+    tol = 1e-6 # Default tolerance
+    max_iter = 100 # Default iterations
     default_options = {
         "bfgs": {"gtol": tol, "maxiter": max_iter},
         "l-bfgs-b": {"ftol": tol, "gtol": tol, "maxiter": max_iter},
@@ -112,6 +120,7 @@ def minimize_scipy(problem, x0, method, options):
     default_options = (
         default_options[method] if method in default_options.keys() else {}
     )
+    options = options.copy()  # Work with a copy to avoid side effects
     combined_options = default_options | options  # Merge defaults with input
 
     # Convert bounds from Pygmo to Scipy convention
@@ -188,6 +197,10 @@ def minimize_pygmo(problem, x0, method, options):
         return _minimize_pygmo_ipopt(problem, x0, options)
     # elif method == "worhp":
     #     return _minimize_pygmo_worhp(problem, x0, options)
+    elif method == "de":
+        return _minimize_pygmo_de(problem, x0, options)
+    elif method == "sga":
+        return _minimize_pygmo_sga(problem, x0, options)
     elif method == "pso":
         return _minimize_pygmo_pso(problem, x0, options)
     else:
@@ -198,14 +211,22 @@ def minimize_pygmo(problem, x0, method, options):
         )
         raise ValueError(error_message)
 
-def _minimize_pygmo_pso(problem, x0, options):
-    """Solve optimization problem using Pygmo's wrapper to PSO"""
+def _minimize_pygmo_de(problem, x0, options):
+    """Solve optimization problem using Pygmo's wrapper to de"""
+
+    # Define solver options
+    default_options = {
+        "gen": 10, 
+        "pop_size": 10,
+    }
+    options = options.copy()  # Work with a copy to avoid side effects
+    combined_options = default_options | options
     
     # Solve the problem
-    algorithm = pg.algorithm(pg.nsga2(gen = 500))
+    algorithm = pg.algorithm(pg.de(gen = int(combined_options["gen"])))
     algorithm.set_verbosity(50)
     problem = pg.problem(problem)
-    population = pg.population(problem, 20)
+    population = pg.population(problem, int(combined_options["pop_size"]))
     population = algorithm.evolve(population)
 
     # Optimization output
@@ -213,6 +234,56 @@ def _minimize_pygmo_pso(problem, x0, options):
     message = ""
 
     return success, message
+
+def _minimize_pygmo_sga(problem, x0, options):
+    """Solve optimization problem using Pygmo's wrapper to sga"""
+
+    # Define solver options
+    default_options = {
+        "gen": 10, 
+        "pop_size": 10,
+    }
+    options = options.copy()  # Work with a copy to avoid side effects
+    combined_options = default_options | options
+    
+    # Solve the problem
+    algorithm = pg.algorithm(pg.sga(gen = int(combined_options["gen"])))
+    algorithm.set_verbosity(50)
+    problem = pg.problem(problem)
[…]
