### 2024-07-30 fc59d18
add tuple to convert_numpy_to_python function
