+++ _sources/source/api/turboflow.axial_turbine.choking_criterion.rst.txt
@@ -1,0 +1,7 @@
+turboflow.axial\_turbine.choking\_criterion module
+==================================================
+
+.. automodule:: turboflow.axial_turbine.choking_criterion
+   :members:
+   :undoc-members:
+   :show-inheritance:
