### 2024-08-02 2a71135
add list of previous 'x' to optimization solver convergence history
