### 2024-08-19 069889b
change structure of saving black-box optimization results
