+++ turboflow/axial_turbine/design_optimization.py
@@ -4,6 +4,7 @@ import CoolProp as cp
 import datetime
 import os
 import yaml
+import pickle
 from .. import pysolver_view as psv
 from .. import utilities as utils
 from . import geometry_model as geom
@@ -288,7 +289,7 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
 
         # Initialize initial guess
         initial_guess = self.index_variables(design_variables)
-        self.design_variables_keys = initial_guess.keys()
+        self.design_variables_keys = list(initial_guess.keys())
         self.initial_guess = np.array(list(initial_guess.values()))
 
         # Index fixed variables
@@ -357,7 +358,7 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
 
         # Initialize initial guess
         initial_guess = self.index_variables(design_variables)
-        self.design_variables_keys = initial_guess.keys()
+        self.design_variables_keys = list(initial_guess.keys())
         self.initial_guess = np.array(list(initial_guess.values()))
 
         # Index fixed variables
@@ -380,8 +381,9 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
 
         # Set fitness function
         self.fitness = self.fitness_gradient_free
-        self.iteration = 1
-        self.success = 0
+
+        # Define variables (development variables) 
+        self.optimization_process = BlackBoxOptimization()
 
     def fitness(self, x):
         
@@ -441,49 +443,47 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
                 self.geometry[key] = self.geometry[key] * angle_range + angle_min
 
         self.geometry = geom.prepare_geometry(self.geometry, self.radius_type)
-        try:
         # Evaluate turbine model
-            solver, results = pa.compute_single_operation_point(
-                self.operation_point,
-                self.x0,
-                self.geometry,
-                self.simulation_options,
-                self.solver_options,
-            )
-            if solver.success and np.min(list(results["residuals"].values())) < 1e-6 : 
-                self.results = results
-                self.geometry = results["geometry"]
-                # Update initial guess
-                # self.x0 = solvers[0].problem.vars_real
+        solver, results = pa.compute_single_operation_point(
+            self.operation_point,
+            self.x0,
+            self.geometry,
+            self.simulation_options,
+            self.solver_options,
+        )
+        if solver.success: 
+            self.results = results
+            self.geometry = results["geometry"]
 
-                # Evaluate objective function
-                self.f = self.get_nested_value(self.results, self.obj_func["variable"])[0]/self.obj_func["scale"] # self.obj.func on the form "key.column"
+            # Evaluate objective function
+            self.f = self.get_nested_value(self.results, self.obj_func["variable"])[0]/self.obj_func["scale"] # self.obj.func on the form "key.column"
 
-                # Evaluate additional constraints
-                self.results["additional_constraints"] = pd.DataFrame({"interspace_area_ratio": self.geometry["A_in"][1:]
-                    / self.geometry["A_out"][0:-1]})
-                
-                # Evaluate constraints
-                c_eq = self.evaluate_constraints(self.eq_constraints)
-                c_ineq = self.evaluate_constraints(self.ineq_constraints)
+            # Evaluate additional constraints
+            self.results["additional_constraints"] = pd.DataFrame({"interspace_area_ratio": self.geometry["A_in"][1:].values
+                / self.geometry["A_out"][0:-1].values})
+                            
+            # Evaluate constraints
+            c_eq = self.evaluate_constraints(self.eq_constraints)
+            c_ineq = self.evaluate_constraints(self.ineq_constraints)
 
[…]
