### 2024-08-19 d8489bc
change critical mach to be unity for initial guess
