+++ turboflow/axial_turbine/performance_analysis.py
@@ -575,9 +575,6 @@ def get_initial_guess(initial_guess, problem, boundary_conditions, geometry, flu
             initial_guess = {key: val for key, val in initial_guess.items() if not (key.startswith("v_crit_in") or key.startswith("s_crit_throat"))}
         
         initial_guesses[i] = initial_guess
-        # Always normalize initial guess
-        # initial_guess_scaled = scale_values(initial_guess)
-        # initial_guesses_scaled += initial_guess_scaled
 
     return initial_guesses
 
@@ -818,433 +815,6 @@ class CascadesNonlinearSystemProblem(psv.NonlinearSystemProblem):
 
         return scaled_variables
 
-    # def get_initial_guess(self, initial_guess=None):
-    #     """
-    #     Determine the initial guess for the performance analysis based on the given parameters or default values.
-
-    #     The given `initial_guess` make take three forms:
-
-    #         - None, which generate a default initial guess
-    #         - A dictionary which generates an initial guess through `compute_heuristic_initial_guess`. Required elements are:
-
-    #             - `enthalpy_loss_fractions`, which is a list containing the assumed fractions of enthalpy loss that occurs for each cascade.
-    #             - `eta_ts`, which is the assumed total-to-static efficiency.
-    #             - `eta_tt`, which is the assumed total-to-total efficiency.
-    #             - `Ma_crit`, which is the assumed critical mash number.
-
-    #         - A dictionary containing the full set of variables needed to evaluate turbine performance. This option require that the user has complete knowledge of what are the required variables, and the setup of the inital guess dictionary for the given configuration.
-
-    #     The initial guess is scaled in this function.
-
-    #     Parameters
-    #     ----------
-    #     initial_guess : dict, optional
-    #         A dictionary containing the initial guess parameters.
-
-    #     Returns
-    #     -------
-    #     dict
-    #         Initial guess for the performance analysis.
-
-    #     Raises
-    #     ------
-    #     ValueError
-    #         If the provided initial_guess is invalid or incompatible with the model options.
-
-    #     """
-
-    #     number_of_cascades = self.geometry["number_of_cascades"]
-
-    #     # Define initial guess
-    #     if isinstance(initial_guess, dict):
-    #         # valid_keys_1 = ["enthalpy_loss_fractions", "eta_ts", "eta_tt", "Ma_crit"]
-    #         valid_keys_1 = ["eta_tt", "eta_ts", "mach_1", "mach_2"]
-    #         valid_keys_2 = [
-    #             "w_out",
-    #             "s_out",
-    #             "beta_out",
-    #             "v_crit_in",
-    #             "w_crit_throat",
-    #             "s_crit_throat",
-    #         ]
-    #         valid_keys_2 = ["v_in"] + [
-    #             f"{key}_{i+1}"
-    #             for i in range(number_of_cascades)
-    #             for key in valid_keys_2
-    #         ]
-    #         valid_keys_3 = [
-    #             key for key in valid_keys_2 if not key.startswith("v_crit_in")]
-    #         valid_keys_4 = [
-    #             key
-    #             for key in valid_keys_2
-    #             if not (
-    #                 key.startswith("v_crit_in")
-    #                 or key.startswith("s_crit_throat")
-    #                 or key.startswith("beta_crit_throat")
-    #             )
-    #         ]
-
-    #         check = []
-    #         check.append(set(valid_keys_1) == set(list(initial_guess.keys())))
-    #         check.append(set(valid_keys_2) == set(list(initial_guess.keys())))
[…]
