### 2024-08-19 1ced1c4
add function for saving and loading objects through pickle
