### 2024-09-02 84ddefb
replacing min by smooth min
