+++ turboflow/axial_turbine/design_optimization.py
@@ -74,16 +74,21 @@ def compute_optimal_turbine(
     # Perform initial function call to initialize problem
     # This populates the arrays of equality and inequality constraints
     # TODO: it might be more intuitive to create a new method called initialize_problem() that generates the initial guess and evaluates the fitness() function with it
-    problem.fitness(problem.initial_guess)
+    problem.fitness(problem.initial_guesses[0])
 
     # Load solver configuration
     solver_config = config["design_optimization"]["solver_options"]
 
-    # Initialize solver object using keyword-argument dictionary unpacking
-    solver = psv.OptimizationSolver(problem, **solver_config)
-
-    # Solve optimization problem for initial guess x0
-    solver.solve(problem.initial_guess)
+    solver_container = SolverContainer()
+    success = []
+    for ig in problem.initial_guesses:
+        # Initialize solver object using keyword-argument dictionary unpacking
+        solver = psv.OptimizationSolver(problem, **solver_config)
+        
+         # Solve optimization problem for initial guess x0
+        solver.solve(ig)
+        solver_container.solver_container.append(solver)
+        success.append(solver.success)
 
     dfs = {
         "operation point": pd.DataFrame(
@@ -137,7 +142,10 @@ def compute_optimal_turbine(
             for sheet_name, df in dfs.items():
                 df.to_excel(writer, sheet_name=sheet_name, index=True)
 
-    return solver        
+    if len(solver_container.solver_container) == 1:
+        return solver_container.solver_container[0]
+
+    return solver_container     
 
 class CascadesOptimizationProblem(psv.OptimizationProblem):
     """
@@ -230,6 +238,8 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
         )
 
         algorithm = config["design_optimization"]["solver_options"]["method"]
+        
+        # self.init_black_box_constrained(config)
         if algorithm in GRADIENT_ALGORITHMS:
             self.init_gradient_based(config)
         elif algorithm in GENETIC_ALGORITHMS:
@@ -290,7 +300,7 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
         # Initialize initial guess
         initial_guess = self.index_variables(design_variables)
         self.design_variables_keys = list(initial_guess.keys())
-        self.initial_guess = np.array(list(initial_guess.values()))
+        # self.initial_guess = np.array(list(initial_guess.values()))
 
         # Index fixed variables
         self.fixed_params = self.index_variables(fixed_params)
@@ -298,6 +308,15 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
         # Initialize bounds
         lb, ub = self.get_given_bounds(design_variables)
         self.bounds = (lb, ub)
+
+        # Initialize a range of initial guesses for multistart optimization
+        starts = config["design_optimization"]["multistarts"]
+        if starts > 0:
+            bounds = [(lower, upper) for lower, upper in zip(lb, ub)]
+            initial_guesses = pa.latin_hypercube_sampling(bounds, starts)
+            self.initial_guesses = np.insert(initial_guesses, 0, np.array(list(initial_guess.values())), axis = 0)
+        else:
+            self.initial_guesses = [np.array(list(initial_guess.values()))]            
 
         # Adjust set of independent variables according to choking model
         self.independent_variables = [
@@ -359,7 +378,7 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
         # Initialize initial guess
         initial_guess = self.index_variables(design_variables)
         self.design_variables_keys = list(initial_guess.keys())
-        self.initial_guess = np.array(list(initial_guess.values()))
+        self.initial_guesses = [np.array(list(initial_guess.values()))]
 
         # Index fixed variables
         self.fixed_params = self.index_variables(fixed_params)
[…]
