### 2024-09-03 b59517a
add functions to prepare solver object for pickle
