### 2024-09-03 4a1e586
add scripts for paper 3
