+++ dev_projects/optimization_algos/convergence_analysis.py
@@ -1,0 +1,193 @@
+
+import turboflow as tf
+import numpy as np
+import matplotlib.pyplot as plt
+from mpl_toolkits.axes_grid1.inset_locator import inset_axes, mark_inset
+import matplotlib.cm as cm
+import seaborn as sns
+
+algorthims = False
+pop_size = True
+save_fig = True
+
+# Load object
+filename_slsqp = "output/pickle_slsqp_2024-08-19_14-30-05.pkl"
+filename_ipopt = "output/pickle_ipopt_2024-08-21_13-42-48.pkl"
+filename_snopt = "output/pickle_snopt_2024-08-19_20-45-38.pkl"
+filename_sga_20 = "output/pickle_sga_2024-08-26_05-00-45.pkl"
+filename_pso_20 = "output/pickle_pso_2024-08-23_09-31-51.pkl"
+filename_slsqp_BB = "output/pickle_slsqp_BB_2024-08-23_10-49-46.pkl"
+filename_snopt_BB = "output/pickle_snopt_BB_2024-08-23_15-33-27.pkl"
+filename_ipopt_BB = "output/pickle_ipopt_BB_2024-08-24_01-35-28.pkl"
+
+solver_slsqp = tf.load_from_pickle(filename_slsqp)
+solver_ipopt = tf.load_from_pickle(filename_ipopt)
+solver_snopt = tf.load_from_pickle(filename_snopt)
+solver_sga_20 = tf.load_from_pickle(filename_sga_20)
+solver_pso_20 = tf.load_from_pickle(filename_pso_20)
+solver_slsqp_BB = tf.load_from_pickle(filename_slsqp_BB)
+solver_snopt_BB = tf.load_from_pickle(filename_snopt_BB)
+solver_ipopt_BB = tf.load_from_pickle(filename_ipopt_BB)
+
+# Close open figures
+plt.close()
+
+# Access the tab10 colormap
+# cmap = plt.get_cmap('tab10')
+# colors = [cmap(i) for i in range(cmap.N)]
+# colors = plt.get_cmap('magma')(np.linspace(0.2, 1.0, 7))
+colors = sns.color_palette("colorblind", n_colors=8)  # Adjust n_colors as needed
+
+
+# Linestyles
+linestyles = ['-', '--', ':', '-.', '-', '-', '--', ':']
+
+def adjust_objective_values(objective_values, constraint_values):
+
+    objective_values = objective_values.copy()
+    constraint_values = constraint_values.copy()
+    for i in range(1, len(objective_values)):
+
+        if objective_values[i] > objective_values[i-1] or np.isnan(objective_values[i]) :
+            objective_values[i] = objective_values[i-1]
+            constraint_values[i] = constraint_values[i-1]
+
+    return objective_values, constraint_values
+
+obj_sga_20, cons_sga_20 = adjust_objective_values(solver_sga_20.convergence_history["objective_value"][:-1], solver_sga_20.problem.optimization_process.constraint_violation[1:])
+obj_pso_20, cons_pso_20 = adjust_objective_values(solver_pso_20.convergence_history["objective_value"][:-1], solver_pso_20.problem.optimization_process.constraint_violation[1:])
+
+if algorthims:
+     
+        convergence_data = {"EO, SLSQP" : {"objective" : solver_slsqp.convergence_history["objective_value"], 
+                                        "constraint_violation" : solver_slsqp.convergence_history["constraint_violation"], 
+                                        "iterations" : solver_slsqp.convergence_history["grad_count"]},
+                        "EO, IPOPT" : {"objective" : solver_ipopt.convergence_history["objective_value"],
+                                        "constraint_violation" : solver_ipopt.convergence_history["constraint_violation"],
[…]
