### 2024-09-04 2e4597e
Add automatic differentiation demo using JAX
