### 2024-09-13 #6979 @kyleniemeyer
@lasseband the issue isn't really finding an editor, it's just that our editorial board is already at capacity handling other submissions.
