### 2024-09-25 33bb55e
try to fix one-stage example problems
