+++ examples/one-stage turbine/one_stage_plots.py
@@ -26,7 +26,8 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     data,
     x_key="PR_ts",
     y_keys=["mass_flow_rate"],
-    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    # xlabel= "Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    xlabel= r"Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
     ylabel="Mass flow rate [kg/s]",
     colors='k',
     filename = 'mass_flow_rate',
@@ -40,7 +41,8 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     data,
     x_key="PR_ts",
     y_keys=["p_1", "p_2", "p_3", "p_4"],
-    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    # xlabel= "Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    xlabel= r"Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
     ylabel="Static pressure [Pa]",
     linestyles=["-", ":", "--", "-."],
     color_map='Reds',
@@ -62,7 +64,8 @@ fig1, ax1 = tf.plot_functions.plot_lines(
         "loss_trailing_4",
         "loss_incidence_4",
     ],
-    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    # xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    xlabel= r"Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
     ylabel="Loss coefficient [-]", 
     color_map='Reds',
     labels = labels,
