### 2024-09-25 #6979 @philipcardiff
Thanks @lasseband,

Ah OK. A documentation link as a badge is reasonable, but some users might overlook it (like "slow-witted me :D). For better clarity, it would be helpful to also include an explicit link to the documentation in a clearly labeled section, such as "Getting Started" or "Documentation," in addition to the badge. This ensures that users can easily find the information they need without relying solely on badges.
