### 2024-10-15 #6979 @lasseband
Hi @philipcardiff,

I apologize for the delayed response. I added an explicit link to the documentation in the README file under "Core features", so it should be easier to find it now :D.

Thanks for adding the reviewers. I look forward to addressing any feedback that comes in to help improve our work.
