### 2024-10-16 #6979 @philipcardiff
Hi @talbring, Would you be interested in reviewing the submission “TurboFlow: Meanline Modelling of Turbomachinery,” which we received at JOSS? The submission link is https://github.com/openjournals/joss-reviews/issues/6979, and the repository is at https://turbo-sim.github.io/turboflow/.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. To get an idea of what is expected, you can look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html).
