### 2024-10-17 f43e777
fix degrees to radians in documentation
