### 2024-10-20 8f8785d
add files for centrifugal compressor development
