+++ dev_projects/centrifugal_compressor/compressor_geom/functions/turbomachinery/plot_centrifugal_compressor/plot_compressor_meridional_plane_bis.m
@@ -1,0 +1,154 @@
+function plot_compressor_meridional_plane(geom, nameValueArgs)
+
+    arguments
+        geom (1, 1) struct
+        nameValueArgs.axial_offset (1, 1) double = 0.00;
+        nameValueArgs.edge_width (1, 1) double = 0.50;
+        nameValueArgs.color_blades (1, 3) double = 0.75*[1, 1, 1];
+        nameValueArgs.color_hub (1, 3) double = 0.90*[1, 1, 1];
+        nameValueArgs.symmetric (1, 1) logical = false
+    end
+
+    % Load geometry
+    has_vanes = geom.has_vanes;
+    r_1_h = 0.5*geom.D_1_h;
+    r_1_s = 0.5*geom.D_1_s;
+    r_2 = 0.5*geom.D_2;
+    r_3 = 0.5*geom.D_3;
+    r_4 = 0.5*geom.D_4;
+    r_5 = 0.5*geom.D_5;
+    r_6 = 0.5*geom.D_6;
+    b_1 = geom.b_1;
+    b_2 = geom.b_2;
+    b_2 = geom.b_2;
+    b_3 = geom.b_3;
+    b_5 = geom.b_5;
+    b_6 = geom.b_6;
+    L_z = geom.L_z;
+    eps_a = geom.eps_a;
+    eps_r = geom.eps_r;
+    eps_b = geom.eps_b;
+    LR_split = geom.LR_split;
+   
+    % Define additional geometric parameters for plotting 
+    z0 = 0;
+    r_shaft = r_1_h;
+    t_casing = 0.5*b_2;
+    t_backplate = 0.40*b_2;
+    L_inducer = geom.L_inducer;
+    L_shaft = L_z/2;
+    inducer_type = geom.inducer_type;
+    
+    % Impeller
+    theta = linspace(3/2*pi, 2*pi, 50)';
+    x_h_impeller = L_z*cos(theta);
+    y_h_impeller = r_2 + (r_2 - r_1_h)*sin(theta);
+    x_s_impeller = (L_z - b_2)*cos(theta);
+    y_s_impeller = r_2 + (r_2 - r_1_s)*sin(theta);
+    xy_impeller = [x_h_impeller, y_h_impeller;
+                   flip(x_s_impeller), flip(y_s_impeller)];
+    
+    % Splitter blades
+    theta = linspace(3/2*pi+(1-LR_split)*pi/2, 2*pi, 50)';
+    x_h_splitter = L_z*cos(theta);
+    y_h_splitter = r_2 + (r_2 - r_1_h)*sin(theta);
+    x_s_splitter = (L_z - b_2)*cos(theta);
+    y_s_splitter = r_2 + (r_2 - r_1_s)*sin(theta);
+    xy_splitter = [x_h_splitter, y_h_splitter;
+                   flip(x_s_splitter), flip(y_s_splitter)];
+    
+    % Diffuser
+    x_h_diffuser = [L_z, L_z]';
+    y_h_diffuser = [r_3, r_5]';
+    x_s_diffuser = [L_z - b_3, L_z - b_5]';
+    y_s_diffuser = [r_3, r_5]';
+    xy_diffuser = [x_h_diffuser, y_h_diffuser
+                   flip(x_s_diffuser), flip(y_s_diffuser)];
+    
+    % Hub and shaft
+    if strcmp(inducer_type, "flat")
+        x_nose = z0-[L_inducer, L_inducer]';
+        y_nose = [0, r_1_h]';
+    elseif strcmp(inducer_type, "round")
+        phi = linspace(pi, 1/2*pi, 50)';
+        x_nose = z0 + L_inducer*cos(phi);
[…]
