### 2024-10-23 7281396
Add notes on compressor equations
