### 2024-10-23 c06166f
centrifugal compressor model
