+++ turboflow/centrifugal_compressor/flow_model.py
@@ -1,0 +1,454 @@
+
+import numpy as np
+import pandas as pd
+import CoolProp as cp
+from . import loss_model as lm
+from . import slip_model as sm
+from .. import utilities as utils
+from .. import math
+import scipy.linalg
+import scipy.integrate
+
+
+def evaluate_centrifugal_compressor(
+    variables,
+    boundary_conditions,
+    geometry,
+    fluid,
+    model_options,
+    reference_values,
+):
+
+    # Load reference_values
+    v_max = reference_values["v_max"]
+    s_range = reference_values["s_range"]
+    s_min = reference_values["s_min"]
+    angle_range = reference_values["angle_range"]
+    angle_min = reference_values["angle_min"]
+
+    # Initialize results structure
+    compressor_results = {"residuals" : {},
+                          "planes" : pd.DataFrame(),
+                          "boundary_conditions" : boundary_conditions}
+
+    # Map functions for each component
+    component_function = {
+        "impeller" : evaluate_impeller,
+        "vaneless_diffuser" : evaluate_vaneless_diffuser,
+        "vaned_diffuser" : evaluate_vaned_diffuser,
+        "volute" : evaluate_volute,
+    }
+
+    # Prepare first set of inputs
+    input = {"h0_in" : boundary_conditions["h0_in"],
+             "s_in" : boundary_conditions["s_in"],
+             "v_in" : variables["v_in"]*v_max,
+             "alpha_in" : boundary_conditions["alpha_in"],
+             "w_out" : variables["w_out"]*v_max,
+             "beta_out" : variables["beta_out"]* angle_range + angle_min,
+             "s_out" : variables["s_out"]* s_range + s_min}
+
+    # Evaluate components
+    for key in geometry.keys():
+        results, residuals = component_function[key](input, boundary_conditions, geometry[key], fluid, model_options)
+
+        # Store result in result structure
+        compressor_results[key] = results
+        compressor_results["planes"].loc[len(compressor_results["planes"]), results["inlet_plane"].keys()] = results["inlet_plane"]
+        compressor_results["planes"].loc[len(compressor_results["planes"]), results["exit_plane"].keys()] = results["exit_plane"]
+
+        residuals = utils.add_string_to_keys(residuals, f"_{key}")
+        compressor_results["residuals"].update(residuals)
+
+        # Prepare calculation of the subsequent component
+        input = compressor_results["planes"].loc[len(compressor_results["planes"])-1].to_dict()
+
+    # Compute overall performance
+    compressor_results["overall"] = compute_overall_performance(compressor_results, fluid, geometry)
+
+    return compressor_results
+
+def evaluate_impeller(input, boundary_conditions, geometry, fluid, model_options):
+
+    # Load input variables
+    h0_in = input["h0_in"]
+    s_in = input["s_in"]
+    v_in = input["v_in"] # IV
+    alpha_in = input["alpha_in"]
+    w_out = input["w_out"] # IV
+    beta_out = input["beta_out"] # IV
+    s_out = input["s_out"] # IV
+
+    # Load boundary conditions
+    mass_flow_rate = boundary_conditions["mass_flow_rate"]
+    omega = boundary_conditions["omega"]
+
+    # Load impeller geometry
+    A_in = geometry["area_in"]
+    A_out = geometry["area_out"]
[…]
