+++ turboflow/centrifugal_compressor/performance_analysis.py
@@ -1,0 +1,616 @@
+import os
+import yaml
+import copy
+import time
+import datetime
+import itertools
+import numpy as np
+import pandas as pd
+import CoolProp as cp
+import matplotlib.pyplot as plt
+
+from .. import math
+from .. import pysolver_view as psv
+from .. import utilities as utils
+from .. import properties as props
+from . import geometry_model as geom
+from . import flow_model as flow
+from scipy.stats import qmc
+from scipy import optimize
+
+
+SOLVER_MAP = {"lm": "Lavenberg-Marquardt", "hybr": "Powell's hybrid"}
+"""
+Available solvers for performance analysis.
+"""
+
+def compute_single_operation_point(
+    operating_point,
+    initial_guess,
+    geometry,
+    simulation_options,
+    solver_options,
+    out_dir = "output",
+    out_filename = None,
+    export_results = False,
+):
+
+    # Initialize problem object
+    problem = CentrifugalCompressorProblem(geometry, simulation_options)
+    problem.update_boundary_conditions(operating_point)
+    solver_options = copy.deepcopy(solver_options)
+
+    # Get solver method array
+    solver_methods = [solver_options["method"]] + [method for method in SOLVER_MAP.keys() if method != solver_options["method"]]
+
+    x0 = np.array(list(initial_guess.values()))
+    problem.keys = list(initial_guess.keys())
+    solver = psv.NonlinearSystemSolver(problem, **solver_options)
+    solver.solve(x0)
+    
+    if not solver.success:
+        print("WARNING: All attempts failed to converge")
+        # TODO: Add messages to Log file
+
+    # Export results
+    overall = solver.problem.results["overall"]
+    overall = {key : [val] for key, val in overall.items()}
+    overall = pd.DataFrame(overall)
+    planes = solver.problem.results["planes"]
+
+    if export_results:
+        if not os.path.exists(out_dir):
+                os.makedirs(out_dir)
+
+        # Define filename with unique date-time identifier
+        if out_filename == None:
+            current_time = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+            out_filename = f"cc_{current_time}"
+
+        filepath = os.path.join(out_dir, f"{out_filename}.xlsx")
+        with pd.ExcelWriter(filepath, engine='openpyxl') as writer:
+            # Save the dictionary DataFrame to the first sheet
+            overall.to_excel(writer, sheet_name='Overall', index=True)
+            
+            # Save the original DataFrame to the second sheet
+            planes.to_excel(writer, sheet_name='Planes', index=True)
+
+    return solver, problem.results
+
+
+def find_closest_operation_point(current_op_point, operation_points, solution_data):
+    """
+    Find the solution vector and index of the closest operation point in the historical data.
+
+    Parameters
+    ----------
+    current_op_point : dict
+        The current operation point we want to compare.
+    operation_points : list of dict
+        A list of historical operation points to search through.
+    solution_data : list
+        A list of solution vectors corresponding to each operation point.
+
+    Returns
+    -------
+    tuple
+        A tuple containing the closest solution vector and the one-based index of the closest operation point.
+
+    """
+    min_distance = float("inf")
+    closest_point_x = None
+    closest_index = None
+
+    for i, op_point in enumerate(operation_points):
+        distance = get_operation_point_distance(current_op_point, op_point)
+        if distance < min_distance:
+            min_distance = distance
+            closest_point_x = solution_data[i]
+            closest_index = i
+
+    return closest_point_x, closest_index
+
+
+def get_operation_point_distance(point_1, point_2, delta=1e-8):
+    """
+    Calculate the normalized distance between two operation points, with special consideration
[…]
