+++ turboflow/centrifugal_compressor/loss_model_1.py
@@ -1,0 +1,4 @@
+
+def compute_losses():
+
+    pass

[…]
