### 2024-10-23 436804d
add notes on diffuser modeling
