### 2024-10-23 b86a249
add volute modeling notes
