+++ _sources/source/api/turboflow.centrifugal_compressor.rst.txt
@@ -1,0 +1,21 @@
+turboflow.centrifugal\_compressor package
+=========================================
+
+.. automodule:: turboflow.centrifugal_compressor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   turboflow.centrifugal_compressor.choking_model
+   turboflow.centrifugal_compressor.flow_model
+   turboflow.centrifugal_compressor.geometry_model
+   turboflow.centrifugal_compressor.loss_model
+   turboflow.centrifugal_compressor.loss_model_1
+   turboflow.centrifugal_compressor.performance_analysis
+   turboflow.centrifugal_compressor.slip_model
