+++ dev_projects/centrifugal_compressor/compressor_model/Notes/Vaneless diffuser.md
@@ -1,5 +1,4 @@
-In Oh's paper: The vaneless diffuser losses included as internal losses?
-In Meroni's paper: The vaneless diffuser modelled by using a pressure coefficient
+
 
 # 1D model
 The flow in the vaneless diffuser can be described by the following set of differential equations:
@@ -42,3 +41,347 @@ and the source vector is given by:
 $$S = \begin{bmatrix} -\frac{\rho v_m}{b r} \frac{d}{dm}(br) \\ \frac{\rho v_\theta^2}{r} \sin(\phi) - \frac{2 \tau_w}{b} \cos(\alpha) \\ -\frac{\rho v_\theta v_m}{r}  \sin(\phi) - \frac{2 \tau_w}{b} \sin(\alpha) \\ \frac{2 (\tau_w v + q_w)}{b  (\frac{\partial e}{\partial p})_\rho} \end{bmatrix}$$
 The system of of ODEs can be solved using a linear equation solver to get $\frac{dU}{dm}$ .
 Once $\frac{dU}{dm}$  is known, the solution vector $U$ at any position $x$ can be obtained by integrating $\frac{dU}{dm}$​ over the interval $[0, x]$. This is typically done using numerical methods for solving initial value problems, such as the Runge-Kutta method.
+
+# Algebraic model
+The inlet of the vaneless diffuser is the same as the exit of the impeller
+- Know: Inlet state, geometry $\vec{G}$
+- Independents: $v_{out}$, $s_{out}$, $\alpha_{out}$
+- Assumption: Assume skin friction factor $C_f$
+
+$v_{m,out} = v_{out} \cos \alpha_{out}$
+$v_{\theta,out} = v_{out} \cos \alpha_{out}$
+$h_{0,out} = h_{0,in}$
+$h_{out} = h_{0,out} - 0.5v_{out}^2$
+$\rho_{out}, p_{out} = f(h_{out}, s_{out})$
+$p_{0, out} = f(h_{0, out}, s_{out})$
+$\dot{m}_{out} = \rho_{out}v_{m, out}A_{out}$
+$Y^{LM} = 2C_f(r_{out}-r_{in})/b_{in}$
+$Y^{def} = 2(h_{out} - h_{out,is})/v_{in}^2$
+
+### Residuals
+$\Delta \dot{m}_{out}  = (\dot{m}_{out} - \dot{m})/\dot{m}$
+$\Delta Y = Y^{LM} - Y^{def}$
+$\Delta AM = \frac{r_{out} v_{\theta, out}}{r_{in} v_{\theta, in}} -  \exp\left({-\frac{c_f (r_{out} - r_{in})}{b_{in} \cos{\alpha_{in}}}}\right)$
+
+# Derivations
+## Flow equations in simple ducts
+
+The governing equations for steady-state flow in a one-dimensional duct with area change, heat transfer, and friction are given by:
+$$
+
+\begin{gather}
+
+v \frac{\text{d} \rho}{\text{d} x}+\rho \frac{\text{d} v}{\text{d}x} = -\frac{\rho v}{A} \frac{\text{d} A}{\text{d} x} \\
+
+\rho v \frac{\text{d} v}{\text{d} x}+\frac{\text{d} p}{d x} = -\frac{4}{D_h} \tau_\text{w} \\
+
+\frac{\text{d} p}{\text{d} x}-a^2 \frac{\text{d} \rho}{\text{d} x} =  \left(\frac{4G}{v D_h} \right) \left(\tau_\text{w} \, v+\dot{q}_\text{w}\right)
+
+\end{gather}
+
+$$
+where:
+
+- $A$ is the cross sectional area of the duct.
+
+- $D_h = 4 A/P$ is the hydraulic diameter.
+
[…]
