+++ dev_projects/centrifugal_compressor/compressor_model/Notes/Volute.md
@@ -1,18 +1,63 @@
-Meroni: 
-- Assume purely no tangential component at the volute exit
-- Assume incompressible flow
-- Enthalpy loss calculated with an empirical correlation
-$\rho_{out} = \rho_{in}$
-$v_{out} = \dot{m}/\rho_{out}A_{out}$
-$h_{out} = h_{0, in} - 0.5v_{out}^2$
-$s_{out} = f(h_{out}, \rho_{out})$
-$p_{0,out} = f(h_{0, in}, s_{out})$
-$h_{0, out} - h(p_{0,out}, s_{in}) = k_vv_{in}^2$ 
+# Model Equations
+- Known: Inlet state, geometry $\vec{G}$
+- Independent variables: $v_{out}$, $s_{out}$
+- Assumptions: Purely meridional velocity component at the exit
 
-- This systems seems overdefined? 
-- $h(p_{0,out}, s_{in})$ can now be calculated both from the empirical correlation and the exit total pressure
+$h_{0,out} = h_{0,in}$
+$h_{out} = h_{0,out} - 0.5v_{out}^2$
+$\rho_{out}, p_{out} = f(h_{out}, s_{out})$
+$p_{0, out} = f(h_{0, out}, s_{out})$
+$\dot{m}_{out} = \rho_{out}v_{out}A_{out}$
+$v_{scroll} = v_{out}A_{out}/A_{scroll}$
+$Y^{LM} = f(all)$
+$Y^{def} = h_{0,out} - h_{0,out,is}$
 
-Aungier:
-- Make one of the assumption as above (either compressible or velocity)
-- Need one independent if only one of the assumptions are made, and the loss correlation is more complex than the one above
-- Just a more complex loss correlation
+### Residuals
+$\Delta \dot{m}_{out}  = (\dot{m}_{out} - \dot{m})/\dot{m}$
+$\Delta Y = Y^{LM} - Y^{def}$
+
+# Loss equations
+According to Ceyrowsky et al, the loss coefficient in the volute is the sum of 3 components:
+
+$$ \zeta_\text{volute} = \zeta_\text{radial} + \zeta_\text{tangential} + \zeta_\text{cone}$$
+
+where the original expressions provided by Ceyrowsky can be reformulated as:
+$$
+
+\begin{gather}
+
+\zeta_\text{radial} = \frac{h-h_s}{v_4^2/2} = \frac{\Delta p_0}{\rho v_4^2/2} = v_{r,4}^2/v_4^2 = \cos{(\alpha_4)}^2 \\
+
+\zeta_\text{tangential} = \frac{h-h_s}{v_4^2/2} = \frac{\Delta p_0}{\rho v_4^2/2} = \left( v_{\theta,4}/v_4\right)^2 \left( 1 - \frac{A_4\sin{\alpha_4}}{A_5}\right)^2 = \sin({\alpha_4})^2 \left( 1 - \frac{A_4\sin{\alpha_4}}{A_5}\right)^2 \\
+
+  
+
+\zeta_\text{cone} = \frac{h-h_s}{v_4^2/2} = \frac{\Delta p_0}{\rho v_4^2/2} = \left(v_{5}/v_4 \right)^2 \left( 1 - \frac{A_5}{A_6}\right)^2 = \cos({\alpha_4})^2 \left( \frac{A_4}{A_5} \right)^2\left( 1 - \frac{A_5}{A_6}\right)^2 \\
+
+  
+
+\end{gather}
+$$
[…]
