+++ turboflow/centrifugal_compressor/flow_model.py
@@ -19,13 +19,6 @@ def evaluate_centrifugal_compressor(
     reference_values,
 ):
 
-    # Load reference_values
-    v_max = reference_values["v_max"]
-    s_range = reference_values["s_range"]
-    s_min = reference_values["s_min"]
-    angle_range = reference_values["angle_range"]
-    angle_min = reference_values["angle_min"]
-
     # Initialize results structure
     compressor_results = {"residuals" : {},
                           "planes" : pd.DataFrame(),
@@ -42,15 +35,13 @@ def evaluate_centrifugal_compressor(
     # Prepare first set of inputs
     input = {"h0_in" : boundary_conditions["h0_in"],
              "s_in" : boundary_conditions["s_in"],
-             "v_in" : variables["v_in"]*v_max,
              "alpha_in" : boundary_conditions["alpha_in"],
-             "w_out" : variables["w_out"]*v_max,
-             "beta_out" : variables["beta_out"]* angle_range + angle_min,
-             "s_out" : variables["s_out"]* s_range + s_min}
+             }
 
     # Evaluate components
     for key in geometry.keys():
-        results, residuals = component_function[key](input, boundary_conditions, geometry[key], fluid, model_options)
+        independents = {var : val for var, val in variables.items() if var.endswith(key)}
+        results, residuals = component_function[key](independents, input, boundary_conditions, geometry[key], fluid, model_options, reference_values)
 
         # Store result in result structure
         compressor_results[key] = results
@@ -68,16 +59,30 @@ def evaluate_centrifugal_compressor(
 
     return compressor_results
 
-def evaluate_impeller(input, boundary_conditions, geometry, fluid, model_options):
+def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid, model_options, reference_values):
+    """
+    Evaluate impelkler performance
+    - Deviation calculated according to slip factor
+    - Losses calculated according to correlation for loss coefficient
+    """
+
+    # Load reference_values
+    v_max = reference_values["v_max"]
+    s_range = reference_values["s_range"]
+    s_min = reference_values["s_min"]
+    angle_range = reference_values["angle_range"]
+    angle_min = reference_values["angle_min"]
+
+    # Load indepndent variables
+    v_in = independents["v_in_impeller"]*v_max
+    w_out = independents["w_out_impeller"]*v_max 
+    beta_out = independents["beta_out_impeller"]*angle_range + angle_min 
+    s_out = independents["s_out_impeller"]*s_range + s_min
 
     # Load input variables
     h0_in = input["h0_in"]
     s_in = input["s_in"]
-    v_in = input["v_in"] # IV
     alpha_in = input["alpha_in"]
-    w_out = input["w_out"] # IV
-    beta_out = input["beta_out"] # IV
-    s_out = input["s_out"] # IV
 
     # Load boundary conditions
     mass_flow_rate = boundary_conditions["mass_flow_rate"]
@@ -194,14 +199,27 @@ def evaluate_impeller(input, boundary_conditions, geometry, fluid, model_options
 
     return results, resiudals
 
-def evaluate_vaned_diffuser(input, boundary_conditions, geometry, fluid, model_options):
+def evaluate_vaned_diffuser(independents, input, boundary_conditions, geometry, fluid, model_options, reference_values):
 
+    """
+    Evaluate vaned diffuser
+    - Deviation calculated according to aungier 
[…]
