### 2024-10-24 615321d
update volute modeling note
