### 2024-10-24 d0b9a8a
add derivation of relation between stagnation pressure and entropy loss
