### 2024-10-25 c89ac7b
update compressor model
