+++ turboflow/centrifugal_compressor/loss_model_1.py
@@ -1,4 +1,0 @@
-
-def compute_losses():
-
-    pass
