+++ turboflow/centrifugal_compressor/flow_model.py
@@ -40,7 +40,10 @@ def evaluate_centrifugal_compressor(
 
     # Evaluate components
     for key in geometry.keys():
+        # Prepare set of independent variables
         independents = {var : val for var, val in variables.items() if var.endswith(key)}
+
+        # Evaluate component
         results, residuals = component_function[key](independents, input, boundary_conditions, geometry[key], fluid, model_options, reference_values)
 
         # Store result in result structure
@@ -97,7 +100,7 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
 
     # Load model options
     slip_model = model_options ["slip_model"]
-    loss_model = model_options ["loss_model"]
+    loss_model = model_options ["loss_model"]["impeller"]
 
     # evaluate inlet velocity triangle
     u_in = omega*r_mean_in
@@ -158,18 +161,33 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
     stagnation_properties_out = fluid.get_props(cp.HmassSmass_INPUTS, h0_out, s_out)
     relative_properties_out = fluid.get_props(cp.HmassSmass_INPUTS, h0_rel_out, s_out)  
     isentropic_relative_properties_out = fluid.get_props(cp.PSmass_INPUTS, relative_properties_out["p"], s_in)
-    isentropic_properties_out = fluid.get_props(cp.PSmass_INPUTS, static_properties_out["p"], s_in)
+    isentropic_static_properties_out = fluid.get_props(cp.PSmass_INPUTS, static_properties_out["p"], s_in)
+    isentropic_properties_out = {**isentropic_static_properties_out,**utils.add_string_to_keys(isentropic_relative_properties_out, "0")}
 
     # evaluate exit mass flow rate
     m_out = static_properties_out["d"]*w_m_out*A_out
 
+    # Collect inlet plane and exit plane results 
+    inlet_plane = {**velocity_triangle_in,
+                   **static_properties_in,
+                   **utils.add_string_to_keys(stagnation_properties_in, "0"),
+                   **utils.add_string_to_keys(relative_properties_in, "0_rel")}
+    
+    exit_plane = {**velocity_triangle_out,
+                   **static_properties_out,
+                   **utils.add_string_to_keys(stagnation_properties_out, "0"),
+                   **utils.add_string_to_keys(relative_properties_out, "0_rel"),
+                   }
+
     # Evaluate loss coefficient
     loss_input = {
-        "h0_rel_out" : h0_rel_out,
-        "h0_rel_out_is" : isentropic_relative_properties_out["h"],
-        "h_out" : h_out,
-        "h_out_is" : isentropic_properties_out["h"]}
-    loss_dict = lm.evaluate_loss_model(loss_model, loss_input)
+        "inlet_plane" : inlet_plane,
+        "exit_plane" : exit_plane,
+        "geometry" : geometry,
+        "isentropic" : isentropic_properties_out,
+    }
+    loss_dict = lm.evaluate_loss_model(loss_model, loss_input, "impeller")
+    exit_plane = {**exit_plane, **loss_dict}
 
     # Evaluate slip
     slip_velocity = u_out + v_m_out*math.tand(theta_out) - v_t_out
@@ -183,19 +201,9 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
                  "slip" : slip_velocity-slip_model}
 
     # Store impeller results
-    inlet_plane = {**velocity_triangle_in,
-                   **static_properties_in,
[…]
