+++ turboflow/centrifugal_compressor/loss_model_custom.py
@@ -1,0 +1,107 @@
+import numpy as np
+from .. import math
+
+def compute_losses(input_parameters, component):
+
+    # Map loss model functions
+    component_function = {
+        "impeller" : compute_impeller_losses,
+        "vaneless_diffuser" : compute_vaneless_diffuser_losses,
+        "vaned_diffuser" : compute_vaned_diffuser_losses,
+        "volute" : compute_volute_losses,
+    }   
+
+    loss_dict = component_function[component](input_parameters)
+
+    return loss_dict
+
+def compute_vaned_diffuser_losses(input_parameters):
+    pass
+
+def compute_volute_losses(input_parameters):
+    
+    # Load parameters
+    alpha_in = input_parameters["inlet_plane"]["alpha"]
+    A_in = input_parameters["geometry"]["area_in"]
+    A_scroll = input_parameters["geometry"]["area_scroll"]
+    A_out = input_parameters["geometry"]["area_out"]
+
+    # Calculate dissipation of radial kinetic energy
+    zeta_rad = math.cosd(alpha_in)**2
+
+    # Calculate dissipation of tangential kinetic energy due to expansion loss in scroll
+    zeta_scroll = math.sind(alpha_in)**2*(1-A_in*math.sind(alpha_in)/A_scroll)**2
+
+    # Calculate dissipation of tangential kinetic energy due to expansion loss in cone
+    zeta_cone = math.cosd(alpha_in)**2*(A_in/A_scroll)**2*(1-A_scroll/A_out)**2
+
+    # Store losses in dict
+    loss_dict = {"loss_radial" : zeta_rad,
+                 "loss_scroll" : zeta_scroll,
+                 "loss_cone" : zeta_cone,
+                 "loss_total" : zeta_rad + zeta_scroll + zeta_cone}
+
+    return loss_dict
+
+
+def compute_vaneless_diffuser_losses(input_parameters):
+    
+    # Load parameters
+    Cf = input_parameters["Cf"]
+    r_out = input_parameters["geometry"]["radius_out"]
+    r_in = input_parameters["geometry"]["radius_in"]
+    b_in = input_parameters["geometry"]["width_in"]
+
+    # Store losses in dict
+    loss_dict = {
+        "loss_total" : 2*Cf*(r_out - r_in)/b_in
+    }
+
+    return loss_dict
+
+def compute_impeller_losses(input_parameters):
+    
+    # Load inlet plane
+    v_in = input_parameters["inlet_plane"]["v"]
+    v_m_in = input_parameters["inlet_plane"]["v_m"]
+    w_in = input_parameters["inlet_plane"]["w"]
+    w_t_in = input_parameters["inlet_plane"]["w_t"]
+    d_in = input_parameters["inlet_plane"]["d"]
+    h0_out = input_parameters["inlet_plane"]["h0"]
+
+    # Load exit plane
+    v_out = input_parameters["exit_plane"]["v"]
+    v_t_out = input_parameters["exit_plane"]["v_t"]
+    w_out = input_parameters["exit_plane"]["w"]
+    u_out = input_parameters["exit_plane"]["blade_speed"]
+    d_out = input_parameters["exit_plane"]["d"]
+    h0_out = input_parameters["exit_plane"]["h0"]
+
+    # Load geometry
+    t_cl = input_parameters["geometry"]["tip_clearance"]
+    b_out = input_parameters["geometry"]["width_out"]
[…]

[…]
