+++ dev_projects/centrifugal_compressor/volute_modeling.md
@@ -12,8 +12,6 @@ $$
 \end{gather}
 $$
 
-
-
 ## Derivation of sudden expansion loss
 
 Mass balance:
