+++ _sources/source/api/turboflow.centrifugal_compressor.rst.txt
@@ -16,6 +16,6 @@ Submodules
    turboflow.centrifugal_compressor.flow_model
    turboflow.centrifugal_compressor.geometry_model
    turboflow.centrifugal_compressor.loss_model
-   turboflow.centrifugal_compressor.loss_model_1
+   turboflow.centrifugal_compressor.loss_model_custom
    turboflow.centrifugal_compressor.performance_analysis
    turboflow.centrifugal_compressor.slip_model
