### 2024-10-29 1c4c801
update one-stage example file
