### 2024-11-01 c649331
dev projects regarding choking
