### 2024-11-01 37cf7be
example files for centrifugal compressor case
