### 2024-11-01 8114314
improve loss modeling, structure and hoking evaluations
