### 2024-11-01 09a048d
add centrifugal compressor configuration
