+++ examples/one-stage turbine/one-stage_config.yaml
@@ -1,7 +1,7 @@
 turbomachinery: axial_turbine
 operation_points: 
-  fluid_name: air          # Fluid
-  T0_in: 295.6             # Inlet total pressure (Inlet total temperature)
+  fluid_name: water          # Fluid
+  T0_in: 600            # Inlet total pressure (Inlet total temperature)
   p0_in: 13.8e4            # Inlet total temperature (Inlet total pressure)
   p_out : 13.8e4/2.298     # Exit static pressure
   omega: 1627              # Flow angle at first stator (Rotational speed)
@@ -9,8 +9,8 @@ operation_points:
 
 
 simulation_options: 
-  deviation_model : aungier 
-  choking_criterion : evaluate_cascade_throat
+  deviation_model : aungier
+  choking_criterion : critical_mass_flow_rate
   rel_step_fd: 1e-4 
   loss_model:  
     model: benner  
