### 2024-11-02 b254632
re-structure compressor development notes and files
