+++ dev_projects/centrifugal_compressor/vaneless_diffuser_model/functions/fluid_properties/PropsSI_scalar.m
@@ -1,0 +1,358 @@
+function out = PropsSI_scalar(varargin)
+    
+    % Wrapper around CoolProp high-level interface
+    % This function accepts scalars as input variables
+
+    % Calculate trivial thermodynamic properties
+    if nargin == 2
+        
+        % Call Coolprop wrapper
+        out = PropsSI_wrapper(varargin{:});
+    
+    % Calculate non-trivial thermodynamic properties
+    elseif nargin == 6
+        
+        % Rename arguments
+        name_out = varargin{1};
+
+        % Speed of sound with Homogeneous Frozen Model
+        if strcmp(name_out, 'A_HFM')
+            out = compute_speed_of_sound_HFM(varargin{:});
+
+        % Speed of sound with Homogeneous Equilibrium Model
+        elseif strcmp(name_out, 'A_HEM')
+            out = compute_speed_of_sound_HEM(varargin{:});
+        
+        % Speed of sound with smoothed Homogeneous Equilibrium Model
+        elseif strcmp(name_out, 'A_HEMs')
+            out = compute_speed_of_sound_HEMs(varargin{:});
+                
+        % Speed of sound with Finite Difference Method
+        elseif strcmp(name_out, 'A_HEMcore')
+            out = compute_speed_of_sound_HEM_core(varargin{:});
+        
+        % Speed of sound with Finite Difference Method
+        elseif strcmp(name_out, 'A_FDM')
+            out = compute_speed_of_sound_FDM(varargin{:});
+
+        % Compressibility factor according to its definition
+        elseif strcmp(name_out, 'Z_def')  
+            out = compute_compressibility_factor(varargin{:});
+            
+        % Vapor quality based on mass
+        elseif strcmp(name_out, 'Qmass')
+            out = compute_vapor_quality_mass(varargin{:});
+
+        % Vapor quality based on volume
+        elseif strcmp(name_out, 'Qvol')
+            out = compute_vapor_quality_volume(varargin{:});
+
+        % Isobaric expansion coefficient with HFM
+        elseif strcmp(name_out, 'isobaric_expansion_2phase')
+            out = compute_isobaric_expansion(varargin{:});
+
+        % Isobaric expansion coefficient with HFM
+        elseif strcmp(name_out, 'isothermal_compressibility_2phase')
+            out = compute_isothermal_compressibility(varargin{:});
+
+        % All other thermodynamic properties
+        else              
+            out = PropsSI_wrapper(varargin{:});
+        
+        end
+
+    
+    % Invalid number of arguments
+    else
+        
+        errID = 'thermodynamicEvaluation:inputError';
+        msgtext = 'The number of arguments must be 2 or 6';
+        throw(MException(errID,msgtext));
+
+    end
+
+end
+
+
+%% Speed of sound models
+function out = compute_speed_of_sound_HFM(varargin)
+
+    % Use HFM model within two-phase region (piecewise function)
+    if PropsSI_wrapper('Phase', varargin{2:end}) == 6  % inside two-phase region
+
+        % Temperature and density of the two-phase mixture
+        fluid_name = varargin{6};
+        T_mix = PropsSI_wrapper('T', varargin{2:end});
[…]
