+++ dev_projects/centrifugal_compressor/validation_cases/data_conversion/files_yaml/NASA_vaned.yaml
@@ -1,0 +1,97 @@
+casename: NASA_vaned
+fluid_name: air
+p0_in: 101325
+T0_in: 288.15
+mass_flow_PR:
+  - [4.18171454312, 3.52934945688, 2.70146241032, 2.09878832768]
+  - [4.30119521184, 3.60547580224, 2.79450773728, 2.2066343616]
+  - [4.38366730928, 3.73658203392, 2.90764265392, 2.29227706712]
+  - [4.4428746730399995, 3.77781808264, 2.98165526056, 2.41175773584]
+  - [4.49891142872, 3.85183068928, 3.06412282208, 2.52171750848]
+  - [4.53592, 3.86134704944, 3.12122098304, 2.5830431468799997]
+  - [4.6606850155199995, 3.8708634095999996, 3.1762008693599997, 2.61264909672]
+  - [4.6606850155199995, 3.87826603104, 3.18677409888, 2.63696616384]
+  - [4.6606850155199995, 3.8846072472, 3.19417672032, 2.64859626272]
+  - [4.65222552472, .nan, .nan, 2.64859626272]
+  - [4.64587977264, .nan, .nan, .nan]
+  - [4.6374248177599995, .nan, .nan, .nan]
+  - [4.6057051292, .nan, .nan, .nan]
+  - [4.56763968856, .nan, .nan, .nan]
+RPM_PR:
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+  - [21789.0, 19610.100000000002, 17431.199999999997, 15252.3]
+PR_tt:
+  - [4.16291, 3.29809, 2.64991, 2.13692]
+  - [4.15945, 3.25477, 2.62912, 2.10745]
+  - [4.11438, 3.14905, 2.58059, 2.07279]
+  - [4.07626, 3.11092, 2.5338, 2.026]
+  - [4.02946, 2.96014, 2.46101, 1.94627]
+  - [3.96534, 2.87348, 2.38475, 1.87522]
+  - [2.84749, 2.7799, 2.19237, 1.80416]
+  - [3.12825, 2.4922, 2.0208, 1.72964]
+  - [3.36742, 2.34662, 1.90295, 1.66724]
+  - [3.56499, .nan, .nan, 1.59099]
+  - [3.68977, .nan, .nan, .nan]
+  - [3.77643, .nan, .nan, .nan]
+  - [3.85269, .nan, .nan, .nan]
+  - [3.91161, .nan, .nan, .nan]
+mass_flow_eta:
+  - [4.18171454312, 3.52934945688, 2.70146241032, 2.09878832768]
[…]
