+++ dev_projects/centrifugal_compressor/validation_cases/Eckardt/eckardt_impeller_A/eckardt_impeller_A.yaml
@@ -1,0 +1,43 @@
+casename: Eckardt_impeller_A
+fluid_name: air
+p0_in: 101325
+T0_in: 288.15
+mass_flow_PR:
+  - [2.5175, 3.02453, 3.53027, 4.23505]
+  - [3.10586, 3.87816, 4.56075, 5.34315]
+  - [3.8277, 4.49707, 5.34378, 6.12606]
+  - [4.59817, 5.30175, 6.10493, 6.91558]
+  - [5.04023, 5.99258, 6.79877, 7.19985]
+RPM_PR:
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+PR_tt:
+  - [1.4234, 1.64808, 1.942, 2.30472]
+  - [1.41157, 1.6248, 1.91348, 2.27623]
+  - [1.38632, 1.59471, 1.86373, 2.20918]
+  - [1.34619, 1.54305, 1.78225, 2.05847]
+  - [1.32181, 1.47259, 1.62958, 1.92971]
+mass_flow_eta:
+  - [2.5142, 5.98125, 6.78409, 4.21534]
+  - [3.10227, 5.2892, 6.09205, 5.325]
+  - [3.81818, 4.48807, 5.32841, 6.1125]
+  - [4.59034, 3.87273, 4.54602, 6.89659]
+  - [5.02841, 3.01364, 3.51477, 7.18295]
+RPM_eta:
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+  - [10000, 12000, 14000, 16000]
+eta_tt:
+  - [0.891877, 0.804042, 0.75409, 0.857863]
+  - [0.899788, 0.861963, 0.848046, 0.875938]
+  - [0.892859, 0.885756, 0.875938, 0.867969]
+  - [0.862887, 0.896785, 0.885871, 0.820847]
+  - [0.828932, 0.888816, 0.867911, 0.763099]
+partial_geometry: [0.12, 0.28, 0.4, 0.00211, 0.00108, 20, 0, 0.6748000000000001, 0.026,
+  0.01730258, -51.5, -40, -63, -30, 2.0e-06, 1, 0.02, 0, 0.2, 0, 0.000235, 0, none,
+  0.13]

[…]
