+++ examples/one-stage turbine/one-stage_main.py
@@ -7,11 +7,12 @@ import matplotlib.pyplot as plt
 import turboflow as tf
 
 # Define mode 
-MODE = "design_optimzation"
+# MODE = "design_optimzation"
+MODE = "performance_analysis"
 
 # Load configuration file
 CONFIG_FILE = os.path.abspath("one-stage_config.yaml")
-config = tf.load_config(CONFIG_FILE, print_summary=False)
+config = tf.load_config(CONFIG_FILE, print_summary=True)
 
 if MODE == "performance_analysis":
     
