+++ dev_projects/Kofskey1972_one_stage/kofskey1972_1stage.yaml
@@ -1,12 +1,11 @@
 turbomachinery: axial_turbine
 operation_points: 
-  - fluid_name: air
-    T0_in: 295.6
-    p0_in: 13.8e4
-    p_out : 13.8e4/2.298 #, 13.8e4/2.5]
-    # p_out : 52252.42718
-    omega: 1627
-    alpha_in: 0
+  fluid_name: air
+  T0_in: 295.6
+  p0_in: 13.8e4
+  p_out : 13.8e4/2.298 
+  omega: 1627
+  alpha_in: 0
 
 simulation_options: 
   deviation_model : aungier 
@@ -146,85 +145,6 @@ design_optimization:
       value : [0.00, 0.030e-2]
     cascade_type : 
       value : ["stator", "rotor"]
-  # variables :
-  #   specific_speed :
-  #     value : 1.2
-  #     lower_bound : 0.01
-  #     upper_bound : 10
-  #   blade_jet_ratio :
-  #     value : 0.5
-  #     lower_bound : 0.1
-  #     upper_bound : 0.9
-  #   hub_tip_ratio_in :
-  #     value : [0.6, 0.6]
-  #     lower_bound : [0.6, 0.6]
-  #     upper_bound : [0.9, 0.9]
-  #   hub_tip_ratio_out :
-  #     value : [0.6, 0.6]
-  #     lower_bound : [0.6, 0.6]
-  #     upper_bound : [0.9, 0.9]
-  #   aspect_ratio :
-  #     value : [1.5, 1.5]
-  #     lower_bound : [1.0, 1.0]
-  #     upper_bound : [2.0, 2.0]
-  #   pitch_chord_ratio :
-  #     value : [0.9, 0.9]
-  #     lower_bound : [0.75, 0.75]
-  #     upper_bound : [1.10, 1.10]
-  #   trailing_edge_thickness_opening_ratio :
-  #     value : [0.1, 0.1]
-  #     lower_bound : [0.05, 0.05]
-  #     upper_bound : [0.4, 0.4]
-  #   leading_edge_angle :
-  #     value : [0.41, 0.5]
-  #     lower_bound : [0.41, 0.08]
-  #     upper_bound : [0.92, 0.58]
-  #   gauging_angle :
-  #     value : [0.17, 0.94]
-  #     lower_bound : [0.06, 0.72]
-  #     upper_bound : [0.28, 0.94]
-  #   v_in :
-  #     value : 0.1
-  #     lower_bound : 0.001
-  #     upper_bound : 0.5
-  #   w_out : 
-  #     value : [0.65, 0.65]
-  #     lower_bound : [0.1, 0.1]
-  #     upper_bound : [1.0, 1.0]
-  #   s_out : 
-  #     value : [0.15, 0.15]
-  #     lower_bound : [0.0, 0.0]
-  #     upper_bound : [0.32, 0.32]
-  #   beta_out : 
-  #     value : [0.83, 0.17]
-  #     lower_bound : [0.72, 0.06]
-  #     upper_bound : [0.94, 0.28]
-  #   v_crit_in : 
-  #     value : [0.4, 0.4]
-  #     lower_bound : [0.1, 0.1]
-  #     upper_bound : [1.0, 1.0]
-  #   beta_crit_throat : 
-  #     value : [0.83, 0.17]
-  #     lower_bound : [0.72, 0.06]
-  #     upper_bound : [0.94, 0.28]
-  #   w_crit_throat : 
-  #     value : [0.65, 0.65]
-  #     lower_bound : [0.1, 0.1]
-  #     upper_bound : [1.0, 1.0]
-  #   s_crit_throat : 
-  #     value : [0.15, 0.15]
-  #     lower_bound : [0.0, 0.0]
[…]
