+++ turboflow/centrifugal_compressor/flow_model.py
@@ -4,13 +4,12 @@ import pandas as pd
 import CoolProp as cp
 from . import loss_model as lm
 from . import slip_model as sm
+from . import choking_criterion as cm
 from .. import utilities as utils
 from .. import math
 import scipy.linalg
 import scipy.integrate
 from scipy.optimize._numdiff import approx_derivative
-
-
 
 def evaluate_centrifugal_compressor(
     variables,
@@ -78,12 +77,11 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
     angle_range = reference_values["angle_range"]
     angle_min = reference_values["angle_min"]
 
-    # Load indepndent variables
+    # Load independent variables
     v_in = independents["v_in_impeller"]*v_max
     w_out = independents["w_out_impeller"]*v_max 
     beta_out = independents["beta_out_impeller"]*angle_range + angle_min 
     s_out = independents["s_out_impeller"]*s_range + s_min
-    w_throat = independents["w_throat_impeller"]*v_max
 
     # Load input variables
     h0_in = input["h0_in"]
@@ -105,7 +103,7 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
     slip_model = model_options ["slip_model"]
     loss_model = model_options ["loss_model"]["impeller"]
 
-    # evaluate inlet velocity triangle
+    # Evaluate inlet velocity triangle
     u_in = omega*r_mean_in
     v_m_in = v_in*math.cosd(alpha_in)
     v_t_in = v_in*math.sind(alpha_in)
@@ -125,20 +123,20 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
                    "blade_speed" : u_in,
                    }
 
-    # evaluate inlet thermodynamic state
+    # Evaluate inlet thermodynamic state
     h_in = h0_in - 0.5*v_in**2
     h0_rel_in = h_in + 0.5*w_in**2
     static_properties_in = fluid.get_props(cp.HmassSmass_INPUTS, h_in, s_in)
     stagnation_properties_in = fluid.get_props(cp.HmassSmass_INPUTS, h0_in, s_in)
     relative_properties_in = fluid.get_props(cp.HmassSmass_INPUTS, h0_rel_in, s_in)
 
-    # evaluate inlet mass flow rate, rothalpy and mach
+    # Evaluate inlet mass flow rate, rothalpy and mach
     m_in = w_m_in*static_properties_in["d"]*A_in
     rothalpy = h0_rel_in - 0.5*u_in**2
     Ma_in = v_in/static_properties_in["a"]
     Ma_rel_in = w_in/static_properties_in["a"]
 
-    # evaluate exit velocity triangle
+    # Evaluate exit velocity triangle
     u_out = omega*r_out
     w_m_out = w_out*math.cosd(beta_out)
     w_t_out = w_out*math.sind(beta_out)
@@ -158,7 +156,7 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
                    "blade_speed" : u_out,
                    }
 
-    # evaluate inlet thermodynamic state
+    # Evaluate inlet thermodynamic state
     h0_rel_out = rothalpy + 0.5*u_out**2
     h_out = h0_rel_out - 0.5*w_out**2
     h0_out = h_out + 0.5*v_out**2
@@ -169,18 +167,17 @@ def evaluate_impeller(independents, input, boundary_conditions, geometry, fluid,
     isentropic_static_properties_out = fluid.get_props(cp.PSmass_INPUTS, static_properties_out["p"], s_in)
[…]
