+++ _sources/source/api/turboflow.centrifugal_compressor.loss_model_oh.rst.txt
@@ -1,0 +1,7 @@
+turboflow.centrifugal\_compressor.loss\_model\_oh module
+========================================================
+
+.. automodule:: turboflow.centrifugal_compressor.loss_model_oh
+   :members:
+   :undoc-members:
+   :show-inheritance:
