+++ _sources/source/api/turboflow.centrifugal_compressor.rst.txt
@@ -12,9 +12,12 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
+   turboflow.centrifugal_compressor.choking_criterion
    turboflow.centrifugal_compressor.flow_model
    turboflow.centrifugal_compressor.geometry_model
    turboflow.centrifugal_compressor.loss_model
    turboflow.centrifugal_compressor.loss_model_custom
+   turboflow.centrifugal_compressor.loss_model_oh
+   turboflow.centrifugal_compressor.loss_model_zhang_set1
    turboflow.centrifugal_compressor.performance_analysis
    turboflow.centrifugal_compressor.slip_model
