+++ _sources/source/api/turboflow.centrifugal_compressor.choking_criterion.rst.txt
@@ -1,0 +1,7 @@
+turboflow.centrifugal\_compressor.choking\_criterion module
+===========================================================
+
+.. automodule:: turboflow.centrifugal_compressor.choking_criterion
+   :members:
+   :undoc-members:
+   :show-inheritance:
