### 2024-11-07 4d26b46
add smoothed abs for throat residual to dev script
