### 2024-11-08 3168d97
add method to determine throat
