+++ turboflow/centrifugal_compressor/choking_criterion.py
@@ -2,10 +2,12 @@
 import numpy as np
 import CoolProp as cp
 from scipy.optimize._numdiff import approx_derivative
+from .. import math
 
 CHOKING_CRITERIONS = [
     "no_throat",
-    "critical_isentropic_throat",
+    "minimize_throat_residual",
+    "evaluate_throat"
 ]
 
 
@@ -19,12 +21,52 @@ def evaluate_choking(independents, inlet_state, boundary_conditions, geometry, f
 
     throat_functions = {
         "no_throat" : lambda *args: ({}, {}),
-        "critical_isentropic_throat" : get_throat_gradient,
+        "minimize_throat_residual" : get_throat_gradient,
+        "evaluate_throat" : evaluate_throat,
     }
 
     residuals, results = throat_functions[choking_criterion](independents, inlet_state, boundary_conditions, geometry, fluid, model_options, reference_values, component, u_throat)
 
     return residuals, results
+
+def get_res_throat(w_throat, fluid, mass_flow_rate, rothalpy, s_in, A_throat, u_throat):
+
+        """ 
+        Get mass flow rate residual at the throat for normalized velocity x
+        """
+
+        w_throat = w_throat[0]
+
+        # Throat
+        h0_rel_throat = rothalpy + 0.5*u_throat**2 # Assume same radius at throat and inlet
+        h_throat = h0_rel_throat - 0.5*w_throat**2
+        s_throat = s_in # Assume isentropic flow
+        statsic_properties_throat = fluid.get_props(cp.HmassSmass_INPUTS, h_throat, s_throat)
+        d_throat = statsic_properties_throat["d"]
+        m_throat = d_throat*w_throat*A_throat
+
+        res = 1-m_throat/mass_flow_rate
+
+        return np.array([res, math.smooth_abs(res, method="logarithmic", epsilon=1e-1)])
+
+def get_res_gradient(w_throat, fluid, mass_flow_rate, rothalpy, s_in, A_throat, u_throat, eps, f0):
+    jac = approx_derivative(
+        get_res_throat,
+        w_throat,
+        abs_step = eps,
+        f0 = f0,
+        method="3-point",
+        args = (
+            fluid,
+            mass_flow_rate,
+            rothalpy,
+            s_in,
+            A_throat, 
+            u_throat,
+        )
+    )
+
+    return jac[:, 0]
 
 def get_throat_gradient(independents, inlet_state, boundary_conditions, geometry, fluid, model_options, reference_values, component, u_throat = 0):
 
@@ -53,50 +95,131 @@ def get_throat_gradient(independents, inlet_state, boundary_conditions, geometry
 
     # Load model options
     rel_step_fd = model_options["rel_step_fd"]
-
-    def get_res_throat(w, rothalpy, u_throat, s_in, mass_flow_rate):
-
-        """ 
-        Get mass flow rate residual at the throat for normalized velocity x
-        """
-
-        # Throat
-        h0_rel = rothalpy + 0.5*u_throat**2 # Assume same radius at throat and inlet
-        h = h0_rel - 0.5*w**2
-        s = s_in # Assume isentropic flow
-        statsic_properties = fluid.get_props(cp.HmassSmass_INPUTS, h, s)
-        d = statsic_properties["d"]
-        m = d*w*A_throat
-
-        res = 1-m/mass_flow_rate
-
-        return res
     
     # Evaluate gradient
-    x = np.array([w_throat])
-    res = get_res_throat(w_throat, rothalpy, u_throat, s_in, mass_flow_rate)
-    eps = rel_step_fd * x
-    jac = approx_derivative(
-        get_res_throat,
-        w_throat,
-        abs_step = eps,
[…]
