### 2024-11-20 a1c6660
improve heurisitc guess function
