+++ turboflow/centrifugal_compressor/performance_analysis.py
@@ -54,27 +54,28 @@ def compute_performance(
                 operation_point,
                 operation_points[:i],  # Use up to the previous point
                 solution_data[:i],  # Use solutions up to the previous point
+                solver_container[:i],
             )
             print(f" Using solution from point {closest_index+1} as initial guess")
             initial_guess = closest_x
+        # try:
+        # Compute performance
+        solver = compute_single_operation_point(
+                            operation_point,
+                            initial_guess,
+                            config["geometry"],
+                            config["simulation_options"],
+                            config["performance_analysis"]["solver_options"],
+                            )
 
-        try:
-            # Compute performance
-            solver = compute_single_operation_point(
-                                operation_point,
-                                initial_guess,
-                                config["geometry"],
-                                config["simulation_options"],
-                                config["performance_analysis"]["solver_options"],
-                                )
+        # Collect solution and solver
+        solution_data.append(solver.problem.vars_real)
+        solver_container.append(solver)
 
-            # Collect solution and solver
-            solution_data.append(solver.problem.vars_real)
-            solver_container.append(solver)
-
-        except Exception as e:
-            handle_failure(i, e, stop_on_failure)
-            solver_container.append(None)
+        # except Exception as e:
+        #     handle_failure(i, e, stop_on_failure)
+        #     solver_container.append(None)
+        #     solution_data.append([])
 
     # Print final report
     print_simulation_summary(solver_container)
@@ -212,7 +213,7 @@ def compute_single_operation_point(
     return solver
 
 
-def find_closest_operation_point(current_op_point, operation_points, solution_data):
+def find_closest_operation_point(current_op_point, operation_points, solution_data, solver_container):
     """
     Find the solution vector and index of the closest operation point in the historical data.
 
@@ -224,6 +225,8 @@ def find_closest_operation_point(current_op_point, operation_points, solution_da
         A list of historical operation points to search through.
     solution_data : list
         A list of solution vectors corresponding to each operation point.
+    solver_container : list
+        A list of solver objects corresponding to each operation point.
 
     Returns
     -------
@@ -237,7 +240,7 @@ def find_closest_operation_point(current_op_point, operation_points, solution_da
 
     for i, op_point in enumerate(operation_points):
         distance = get_operation_point_distance(current_op_point, op_point)
-        if distance < min_distance:
+        if distance < min_distance and solver_container[i].success:
             min_distance = distance
             closest_point_x = solution_data[i]
             closest_index = i
@@ -534,33 +537,34 @@ class CentrifugalCompressorProblem(psv.NonlinearSystemProblem):
 
         """
         3 options:
-            - Give guess of efficiency_tt and phi_out. Initial guess calculated heuristically
-            - Give bounds of efficiency_tt and phi_out, and a sample size. Initial guess calculated by the best 
-            - Give complete set (real values)
+            1. Give guess of efficiency_tt and phi_out. Initial guess calculated heuristically
+            2. Give bounds of efficiency_tt and phi_out, and a sample size. Initial guess calculated by the best 
[…]
