### 2024-11-20 a8d2bab
improve choking calculation
