+++ turboflow/centrifugal_compressor/choking_criterion.py
@@ -2,11 +2,12 @@
 import numpy as np
 import CoolProp as cp
 from scipy.optimize._numdiff import approx_derivative
+from .. import utilities as utils
 from .. import math
 
 CHOKING_CRITERIONS = [
     "no_throat",
-    "minimize_throat_residual",
+    "evaluate_critical_throat",
     "evaluate_throat"
 ]
 
@@ -21,37 +22,98 @@ def evaluate_choking(independents, inlet_state, boundary_conditions, geometry, f
 
     throat_functions = {
         "no_throat" : lambda *args: ({}, {}),
-        "minimize_throat_residual" : get_throat_gradient,
-        "evaluate_throat" : evaluate_throat,
+        "evaluate_critical_throat" : get_critical_throat,
+        "evaluate_throat" : get_throat,
     }
 
     residuals, results = throat_functions[choking_criterion](independents, inlet_state, boundary_conditions, geometry, fluid, model_options, reference_values, component, u_throat)
 
     return residuals, results
 
-def get_res_throat(w_throat, fluid, mass_flow_rate, rothalpy, s_in, A_throat, u_throat):
+def evaluate_throat(w_throat, fluid, mass_flow_rate, rothalpy, s_in, A_throat, theta_throat, u_throat):
 
-        """ 
-        Get mass flow rate residual at the throat for normalized velocity x
-        """
+    """
+    Evaluate throat based on w_throat. (See get_res_throat, but more detailed)
+    """
 
-        w_throat = w_throat[0]
+    w_throat = w_throat[0]
 
-        # Throat
-        h0_rel_throat = rothalpy + 0.5*u_throat**2 # Assume same radius at throat and inlet
-        h_throat = h0_rel_throat - 0.5*w_throat**2
-        s_throat = s_in # Assume isentropic flow
-        statsic_properties_throat = fluid.get_props(cp.HmassSmass_INPUTS, h_throat, s_throat)
-        d_throat = statsic_properties_throat["d"]
-        m_throat = d_throat*w_throat*A_throat
+    # Evaluate velocity triangle
+    beta_throat = theta_throat
+    w_m_throat = w_throat*math.cosd(beta_throat)
+    w_t_throat = w_throat*math.sind(beta_throat)
+    v_m_throat = w_m_throat
+    v_t_throat = w_t_throat + u_throat
+    v_throat = np.sqrt(v_t_throat**2 + v_m_throat**2)
+    alpha_throat = math.arctand(v_t_throat/v_m_throat)
 
-        res = 1-m_throat/mass_flow_rate
+    velocity_triangle_throat = {"v_t" : v_t_throat,
+                   "v_m" : v_m_throat,
+                   "v" : v_throat,
+                   "alpha" : alpha_throat,
+                   "w_t" : w_t_throat,
+                   "w_m" : w_m_throat,
+                   "w" : w_throat,
+                   "beta" : beta_throat,
+                   "blade_speed" : u_throat,
+                   }
 
-        return np.array([res, math.smooth_abs(res, method="logarithmic", epsilon=1e-1)])
+    # Evaluate thermodynamic state
+    h0_rel_throat = rothalpy + 0.5*u_throat**2 # Assume same radius at throat and inlet
+    h_throat = h0_rel_throat - 0.5*w_throat**2
+    h0_throat = h_throat + 0.5*v_throat**2 
+    s_throat = s_in # Assume isentropic flow
+    static_properties_throat = fluid.get_props(cp.HmassSmass_INPUTS, h_throat, s_throat)
+    stagnation_properties_throat = fluid.get_props(cp.HmassSmass_INPUTS, h0_throat, s_throat)
[…]
