### 2024-11-20 5017c9e
add and fix loss models
