+++ turboflow/centrifugal_compressor/loss_model_zhang_set3.py
@@ -18,13 +18,12 @@ def compute_losses(input_parameters, component):
 def compute_vaneless_diffuser_losses(input_parameters):
 
     """
-    Oh's vanless diffuser losses
+    Zhang set 1 vanless diffuser losses
     """
     
     # Load parameters
     T0_in = input_parameters["inlet_plane"]["T0"]
     p0_in = input_parameters["inlet_plane"]["p0"]
-    v_in = input_parameters["inlet_plane"]["v"]
     cp = input_parameters["inlet_plane"]["cp"]
     cv = input_parameters["inlet_plane"]["cv"]
     p_out = input_parameters["exit_plane"]["p"]
@@ -36,9 +35,8 @@ def compute_vaneless_diffuser_losses(input_parameters):
     Y_tot = cp*T0_in*((p_out/p0_out)**alpha - (p_out/p0_in)**alpha)
 
     # Store losses in dict
-    scale = 0.5*v_in**2
     loss_dict = {
-        "loss_total" : Y_tot/scale
+        "loss_total" : Y_tot
     }
 
     return loss_dict
@@ -46,7 +44,8 @@ def compute_vaneless_diffuser_losses(input_parameters):
 def compute_impeller_losses(input_parameters):
 
     """
-    Oh. Impeller losses
+    Zhang set 1 Impeller losses
+    Throat not evaluated for now
     """
     
     # Load inlet plane
@@ -58,18 +57,31 @@ def compute_impeller_losses(input_parameters):
     d_in = input_parameters["inlet_plane"]["d"]
     h0_in = input_parameters["inlet_plane"]["h0"]
     a_in = input_parameters["inlet_plane"]["a"]
+    p_in = input_parameters["inlet_plane"]["p"]
+    Ma_rel_in = input_parameters["inlet_plane"]["Ma_rel"]
     mu_in = input_parameters["inlet_plane"]["mu"]
+    p0_in = input_parameters["inlet_plane"]["p0"]
+    T0_in = input_parameters["inlet_plane"]["T0"]
+    gamma_in = input_parameters["inlet_plane"]["gamma"]
+    cp_in = input_parameters["inlet_plane"]["cpmass"]
+    cv_in = input_parameters["inlet_plane"]["cvmass"]
 
     # Load exit plane
     v_out = input_parameters["exit_plane"]["v"]
+    v_m_out = input_parameters["exit_plane"]["v_m"]
     v_t_out = input_parameters["exit_plane"]["v_t"]
     alpha_out = input_parameters["exit_plane"]["alpha"]
     beta_out = input_parameters["exit_plane"]["beta"]
     w_out = input_parameters["exit_plane"]["w"]
+    w_t_out = input_parameters["exit_plane"]["w_t"]
     u_out = input_parameters["exit_plane"]["blade_speed"]
     d_out = input_parameters["exit_plane"]["d"]
     h0_out = input_parameters["exit_plane"]["h0"]
     a_out = input_parameters["exit_plane"]["a"]
+
+    # Load throat plane
+    w_th = input_parameters["throat_plane"]["w"]
+    p_th = input_parameters["throat_plane"]["p"]
 
     # Load geometry
     t_cl = input_parameters["geometry"]["tip_clearance"]
[…]
