+++ turboflow/centrifugal_compressor/loss_model_zhang_set1.py
@@ -24,6 +24,7 @@ def compute_vaneless_diffuser_losses(input_parameters):
     # Load parameters
     T0_in = input_parameters["inlet_plane"]["T0"]
     p0_in = input_parameters["inlet_plane"]["p0"]
+    v_in = input_parameters["inlet_plane"]["v"]
     cp = input_parameters["inlet_plane"]["cp"]
     cv = input_parameters["inlet_plane"]["cv"]
     p_out = input_parameters["exit_plane"]["p"]
@@ -35,8 +36,9 @@ def compute_vaneless_diffuser_losses(input_parameters):
     Y_tot = cp*T0_in*((p_out/p0_out)**alpha - (p_out/p0_in)**alpha)
 
     # Store losses in dict
+    scale = 0.5*v_in**2
     loss_dict = {
-        "loss_total" : Y_tot
+        "loss_total" : Y_tot/scale
     }
 
     return loss_dict
@@ -58,6 +60,11 @@ def compute_impeller_losses(input_parameters):
     h0_in = input_parameters["inlet_plane"]["h0"]
     a_in = input_parameters["inlet_plane"]["a"]
     mu_in = input_parameters["inlet_plane"]["mu"]
+    p0_in = input_parameters["inlet_plane"]["p0"]
+    T0_in = input_parameters["inlet_plane"]["T0"]
+    gamma_in = input_parameters["inlet_plane"]["gamma"]
+    cp_in = input_parameters["inlet_plane"]["cpmass"]
+    cv_in = input_parameters["inlet_plane"]["cvmass"]
 
     # Load exit plane
     v_out = input_parameters["exit_plane"]["v"]
@@ -73,57 +80,62 @@ def compute_impeller_losses(input_parameters):
     a_out = input_parameters["exit_plane"]["a"]
 
     # Load throat plane
-    # w_th = input_parameters["throat_plane"]["w"]
-    # A_crit = input_parameters["throat_plane"]["area_crit"]  # Critical area for given mass flow rate
+    w_th = input_parameters["throat_plane"]["w"]
 
     # Load geometry
     t_cl = input_parameters["geometry"]["tip_clearance"]
     b_out = input_parameters["geometry"]["width_out"]
-    L_ax = input_parameters["geometry"]["impeller_length"]
+    L_ax = input_parameters["geometry"]["length_axial"]
     z = input_parameters["geometry"]["number_of_blades"]
     r_in_hub = input_parameters["geometry"]["radius_hub_in"]
     r_in_tip = input_parameters["geometry"]["radius_tip_in"]
     r_out = input_parameters["geometry"]["radius_out"]
     A_out = input_parameters["geometry"]["area_out"]
     A_in = input_parameters["geometry"]["area_in"]
-    # A_th = input_parameters["geometry"]["area_throat"]
+    A_th = input_parameters["geometry"]["area_throat"]
+    theta_in = input_parameters["geometry"]["leading_edge_angle"]
+
+    # Load boundary conditions
+    mass_flow_rate = input_parameters["boundary_conditions"]["mass_flow_rate"]
 
     # Load model factors
-    Cf = input_parameters["factors"]["skin_friction"] 
[…]
