### 2024-11-20 c1e14b9
add validation and test cases
