+++ dev_projects/centrifugal_compressor/validation_plots_parasitic_losses.py
@@ -4,8 +4,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 impeller = "O" # A, O
+model = "zhang" # oh, zhang
 save_figs = True
 
+# Map files
 files_exp = {
     "A" : "validation_cases\Eckardt\eckardt_impeller_A\eckardt_impeller_A.xlsx",
     "O" : "validation_cases\Eckardt\eckardt_impeller_O\eckardt_impeller_O.xlsx",
@@ -33,13 +35,14 @@ files_sim_incl =  {
     }
 }
 
+
 filename_exp = files_exp[impeller]
-filename_oh = files_sim_incl[impeller]["oh"]
-filename_zhang = files_sim_incl[impeller]["zhang"]
+filename_incl = files_sim_incl[impeller][model]
+filename_exl = files_sim_exl[impeller][model]
 
 data_exp = tf.load_data(filename_exp, ["pr_tt", "eta_tt"])
-data_oh = tf.load_data(filename_oh)
-data_zhang = tf.load_data(filename_zhang)
+data_sim_incl = tf.load_data(filename_incl)
+data_sim_exl = tf.load_data(filename_exl)
 
 # Plot simulation data
 omegas = [10e3, 12e3, 14e3, 16e3]
@@ -49,7 +52,7 @@ subsets = ["omega"] + list(2*np.pi/60*np.array(omegas))
 
 # Mass flow rate
 fig1, ax1 = tf.plot_functions.plot_lines(
-    data_oh,
+    data_sim_exl,
     x_key="mass_flow_rate",
     y_keys=["PR_tt"],
     subsets = subsets,
@@ -58,22 +61,23 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     colors=colors,
 )
 fig1, ax1 = tf.plot_functions.plot_lines(
-    data_zhang,
+    data_sim_incl,
     x_key="mass_flow_rate",
     y_keys=["PR_tt"],
     subsets = subsets,
     fig = fig1,
-    ax = ax1,
+    ax = ax1, 
     xlabel="Mass flow rate [kg/s]",
     ylabel="Total-to-total pressure ratio",
     linestyles=["--"]*4,
+    title = model,
     colors=colors,
 )
-ax1.legend(omegas, loc = 'best')
+ax1.legend(omegas, loc = 'upper left')
 
 # Efficiency
 fig2, ax2 = tf.plot_functions.plot_lines(
-    data_oh,
+    data_sim_exl,
     x_key="mass_flow_rate",
     y_keys=["efficiency_tt"],
     subsets = subsets,
@@ -82,18 +86,19 @@ fig2, ax2 = tf.plot_functions.plot_lines(
     colors=colors,
 )
 fig2, ax2 = tf.plot_functions.plot_lines(
-    data_zhang,
+    data_sim_incl,
     x_key="mass_flow_rate",
     y_keys=["efficiency_tt"],
     subsets = subsets,
     fig = fig2,
-    ax = ax2,
+    ax = ax2, 
     xlabel="Mass flow rate [kg/s]",
     ylabel="Total-to-total efficiency",
     linestyles=["--"]*4,
+    title = model,
     colors=colors,
 )
-ax2.legend(omegas, loc = 'best')
+ax2.legend(omegas, loc = 'lower left')
 ax2.set_xlim([2.0, 7.5])
 ax2.set_ylim([50, 100])
 
@@ -101,33 +106,30 @@ ax2.set_ylim([50, 100])
 subset = ["omega"] + [2*np.pi/60*14e3]
 colors_losses = [plt.get_cmap(color_map)(i / (10 - 1)) for i in range(10)]
 fig3, ax3 = tf.plot_functions.plot_lines(
-    data_oh,
[…]

[…]
