+++ _sources/source/api/turboflow.centrifugal_compressor.loss_model_zhang_set3.rst.txt
@@ -1,0 +1,7 @@
+turboflow.centrifugal\_compressor.loss\_model\_zhang\_set3 module
+=================================================================
+
+.. automodule:: turboflow.centrifugal_compressor.loss_model_zhang_set3
+   :members:
+   :undoc-members:
+   :show-inheritance:
