### 2024-11-20 84779e8
Add Automatic Differentiation changes
