+++ turboflow/axial_turbine/loss_model_benner.py
@@ -1,5 +1,10 @@
-import numpy as np
-from .. import math
+# import numpy as np
+from turboflow import math
+
+import jax
+import jax.numpy as jnp
+
+from turboflow.axial_turbine import loss_model_kacker_okapuu as lm_ko
 
 # TODO: Add smoothing to the min/max/abs/piecewise functions
 
@@ -7,9 +12,6 @@ from .. import math
 # TODO: Raise warninings to a log file to be aware when the loss model is outside its range of validity and because of which variables
 # TODO: the log object can be passed as an additional variable to the functions with a default parameter Logger=None
 # TODO: Whenever something is out of range we can do "if Logger: " to check if the logger was actually passed as an input and log an descriptive message
-
-# TODO: Write detailed docstring in .rst format explaning the equations of the loss model?
-# TODO: This could be prepared in latex and then transformed into .rst with chatGPT
 
 # TODO: For all functions:
 # TODO: add docstring explaning the equations and the original paper
@@ -75,18 +77,20 @@ def compute_losses(input_parameters):
     delta_height = delta_height * (flow_parameters["Re_in"] / 3e5) ** (-1 / 7)
 
     # Profile loss coefficient
-    Y_p = get_profile_loss(flow_parameters, geometry)
+    Y_p = lm_ko.get_profile_loss(flow_parameters, geometry)
 
     # Trailing edge coefficient
-    Y_te = get_trailing_edge_loss(flow_parameters, geometry)
+    Y_te = lm_ko.get_trailing_edge_loss(flow_parameters, geometry)
 
     # Secondary loss coefficient
+    # Y_s = 0.0
     Y_s = get_secondary_loss(flow_parameters, geometry, delta_height)
 
     # Tip clearance loss coefficient
-    Y_cl = get_tip_clearance_loss(flow_parameters, geometry)
+    Y_cl = lm_ko.get_tip_clearance_loss(flow_parameters, geometry)
 
     # Incidence loss coefficient
+    # Y_inc = 0.0
     Y_inc = get_incidence_loss(flow_parameters, geometry, beta_des)
 
     # Penetration depth to blade height ratio
@@ -108,185 +112,6 @@ def compute_losses(input_parameters):
     }
 
     return losses
-
-
-def get_profile_loss(flow_parameters, geometry):
-    r"""
-    Calculate the profile loss coefficient for the current cascade using the Kacker and Okapuu loss model.
-    The equation for :math:`\mathrm{Y_p}` is given by:
-
-    .. math::
-
-        \mathrm{Y_p} = \mathrm{Y_{reaction}} - \left|\frac{\theta_\mathrm{in}}{\beta_\mathrm{out}}\right| \cdot
-        \left(\frac{\theta_\mathrm{in}}{\beta_\mathrm{out}}\right) \cdot (\mathrm{Y_{impulse}} - \mathrm{Y_{reaction}})
-
-    where:
-
-        - :math:`\mathrm{Y_{reaction}}` is the reaction loss coefficient computed using Aungier correlation.
-        - :math:`\mathrm{Y_{impulse}}` is the impulse loss coefficient computed using Aungier correlation.
-        - :math:`\theta_\mathrm{in}` is the inlet metal angle.
-        - :math:`\beta_\mathrm{out}` is the exit flow angle.
-
-    The function also applies various corrections based on flow parameters and geometry factors.
-
-    Parameters
-    ----------
[…]
