+++ dev_projects/automatic_differentiation/design_optimization_jax_copy.py
@@ -1,0 +1,1205 @@
+import numpy as np
+import pandas as pd
+import CoolProp as cp
+import datetime
+import os
+import yaml
+import pickle
+from turboflow import pysolver_view as psv
+from turboflow import utilities as utils
+from turboflow.axial_turbine import geometry_model as geom
+from turboflow.axial_turbine import flow_model as flow
+from turboflow import properties as props
+from turboflow.axial_turbine import performance_analysis as pa
+
+from turboflow.properties import perfect_gas_props
+import jax
+import jax.numpy as jnp
+
+
+RADIUS_TYPE = ["constant_mean", "constant_hub", "constant_tip"]
+ANGLE_KEYS = ["leading_edge_angle", "gauging_angle"]
+INDEPENDENT_VARIABLES = [
+    "v_in",
+    "w_out",
+    "s_out",
+    "beta_out",
+    "v_crit_in",
+    "beta_crit_throat",
+    "w_crit_throat",
+    "s_crit_throat",
+]
+INDEXED_VARIABLES = [
+    "hub_tip_ratio_in",
+    "hub_tip_ratio_out",
+    "aspect_ratio",
+    "pitch_chord_ratio",
+    "trailing_edge_thickness_opening_ratio",
+    "leading_edge_angle",
+    "gauging_angle",
+    "throat_location_fraction",
+    "leading_edge_diameter",
+    "leading_edge_wedge_angle",
+    "tip_clearance",
+    "cascade_type",
+]
+
+GENETIC_ALGORITHMS = psv.GENETIC_SOLVERS
+GRADIENT_ALGORITHMS = psv.GRADIENT_SOLVERS
+
+def compute_optimal_turbine(
+    config,
+    out_filename=None,
+    out_dir="output",
+    export_results=True,
+):
+    r"""
+    Calculate the optimal turbine configuration based on the specified optimization problem.
+
+    The function checks the configuration file before performing design optimization.
+
+    Parameters
+    ----------
+    config : dict
+        A dictionary containing necessary configuration options for computing optimal turbine.
+    initial_guess : None or dict
+        The initial guess of the design optimization. If None, a defualt initial guess is provided. If given, he initial guess must correspond
+        with the given set of design variables given in the configuration dictionary.
+
+    Returns
+    -------
+    solution : object
+        The solution object containing the results of the design optimization.
+    """
+
+    # Initialize problem object
+    problem = CascadesOptimizationProblem(config)
+
+    # Perform initial function call to initialize problem
+    # This populates the arrays of equality and inequality constraints
+    # TODO: it might be more intuitive to create a new method called initialize_problem() that generates the initial guess and evaluates the fitness() function with it
+    problem.fitness(problem.initial_guess)
+
+    # Load solver configuration
+    solver_config = config["design_optimization"]["solver_options"]
+
+    # Initialize solver object using keyword-argument dictionary unpacking
+    solver = psv.OptimizationSolver(problem, **solver_config)
+
+    # Solve optimization problem for initial guess x0
+    solver.solve(problem.initial_guess)
+
+    dfs = {
+        "operation point": pd.DataFrame(
+            {key: pd.Series(val) for key, val in problem.boundary_conditions.items()}
+        ),
+        "overall": pd.DataFrame(problem.results["overall"], index=[0]),
+        "planes": problem.results["planes"],
+        "cascades": problem.results["cascades"],
+        "stage": problem.results["stage"],
+        "geometry": pd.DataFrame(
+            {key: pd.Series(val) for key, val in problem.geometry.items()}
+        ),
+        "solver": pd.DataFrame(
+            {
+                "completed": pd.Series(True, index=[0]),
+                "success": pd.Series(solver.success, index=[0]),
+                "message": pd.Series(solver.message, index=[0]),
[…]
