+++ dev_projects/automatic_differentiation/merge_dict_test.py
@@ -1,0 +1,8 @@
+dict1 = {"a": 1, "b": 2}
+dict2 = {"a": 3, "b": 5, "c": 8}
+
+merge_dict = {}
+
+merge_dict = merge_dict | dict1 
+
+print(merge_dict)

[…]
