### 2024-11-20 c222fc2
update the SNOPT wrapper function and option parsing in pysolverview
