+++ dev_projects/automatic_differentiation/two_stage_dummy_config.yaml
@@ -3,7 +3,7 @@ operation_points:
   fluid_name: air          # Fluid
   T0_in: 295.6             # Inlet total temperature
   p0_in: 13.8e4            # Inlet total pressure
-  p_out : 13.8e4/2.298     # Exit static pressure
+  p_out : 13.8e4/5.280     # Exit static pressure
   omega: 1627              # Rotational speed
   alpha_in: 0              # Flow angle at first stator
 
@@ -40,113 +40,22 @@ performance_analysis :
       ma_2 : [0.8, 0.8]
 
 geometry:
-  cascade_type: ["stator", "rotor"]
-  radius_hub_in: [0.084785, 0.084785]
-  radius_hub_out: [0.084785, 0.081875]
-  radius_tip_in: [0.118415, 0.118415]
-  radius_tip_out: [0.118415, 0.121325]
-  pitch: [1.8294e-2, 1.524e-2]
-  chord: [2.616e-2, 2.606e-2]
-  stagger_angle: [+43.03, -31.05]
-  opening: [0.747503242e-2, 0.735223377e-2]
-  leading_edge_angle : [0.00, 29.60]
-  leading_edge_wedge_angle : [50.00, 50.00]
-  leading_edge_diameter : [2*0.127e-2, 2*0.081e-2]
-  trailing_edge_thickness : [0.050e-2, 0.050e-2]
-  maximum_thickness : [0.505e-2, 0.447e-2]
-  tip_clearance: [0.00, 0.030e-2]
-  throat_location_fraction: [1, 1] # Should be changed
-  
-
-# design_optimization:
-#   objective_function: 
-#     variable : overall.efficiency_ts
-#     scale : 10
-#     type : maximize
-#   variables :
-#     specific_speed :
-#       value : 1.2
-#       lower_bound : 0.01
-#       upper_bound : 10
-#     blade_jet_ratio :
-#       value : 0.5
-#       lower_bound : 0.1
-#       upper_bound : 0.9
-#     hub_tip_ratio_in :
-#       value : [0.6, 0.6]
-#       lower_bound : [0.6, 0.6]
-#       upper_bound : [0.9, 0.9]
-#     hub_tip_ratio_out :
-#       value : [0.6, 0.6]
-#       lower_bound : [0.6, 0.6]
-#       upper_bound : [0.9, 0.9]
-#     aspect_ratio :
-#       value : [1.5, 1.5]
-#       lower_bound : [1.0, 1.0]
-#       upper_bound : [2.0, 2.0]
-#     pitch_chord_ratio :
-#       value : [0.9, 0.9]
-#       lower_bound : [0.75, 0.75]
-#       upper_bound : [1.10, 1.10]
-#     trailing_edge_thickness_opening_ratio :
-#       value : [0.1, 0.1]
-#       lower_bound : [0.05, 0.05]
-#       upper_bound : [0.4, 0.4]
-#     leading_edge_angle :
-#       value : [0.41, 0.5]
-#       lower_bound : [0.41, 0.08]
-#       upper_bound : [0.92, 0.58]
-#     gauging_angle :
-#       value : [0.17, 0.94]
-#       lower_bound : [0.06, 0.72]
-#       upper_bound : [0.28, 0.94]
-#     v_in :
-#       value : 0.1
-#       lower_bound : 0.001
-#       upper_bound : 0.5
-#     w_out : 
-#       value : [0.65, 0.65]
-#       lower_bound : [0.1, 0.1]
-#       upper_bound : [1.0, 1.0]
-#     s_out : 
-#       value : [0.15, 0.15]
-#       lower_bound : [0.0, 0.0]
[…]
