### 2024-11-22 45e57ef
update pysolverview option handling and pydantic validation
