+++ turboflow/pysolver_view/pysolver_utilities.py
@@ -286,3 +286,76 @@ def create_logger(name, path=None, use_datetime=True):
     logger = logging.getLogger()
 
     return logger
+
+
+def print_dict(data, indent=0, return_output=False):
+    """
+    Recursively prints nested dictionaries with indentation or returns the formatted string.
+
+    Parameters
+    ----------
+    data : dict
+        The dictionary to print.
+    indent : int, optional
+        The initial level of indentation for the keys of the dictionary, by default 0.
+    return_output : bool, optional
+        If True, returns the formatted string instead of printing it.
+
+    Returns
+    -------
+    str or None
+        The formatted string representation of the dictionary if return_output is True, otherwise None.
+    """
+    lines = []
+    for key, value in data.items():
+        line = "    " * indent + str(key) + ": "
+        if isinstance(value, dict):
+            if value:
+                lines.append(line)
+                lines.append(print_dict(value, indent + 1, True))
+            else:
+                lines.append(line + "{}")
+        else:
+            lines.append(line + str(value))
+
+    output = "\n".join(lines)
+    if return_output:
+        return output
+    else:
+        print(output)
+
+
+def print_object(obj):
+    """
+    Prints all attributes and methods of an object, sorted alphabetically.
+
+    - Methods are identified as callable and printed with the prefix 'Method: '.
+    - Attributes are identified as non-callable and printed with the prefix 'Attribute: '.
+
+    Parameters
+    ----------
+    obj : object
+        The object whose attributes and methods are to be printed.
+
+    Returns
+    -------
+    None
+        This function does not return any value. It prints the attributes and methods of the given object.
+
+    """
+    # Retrieve all attributes and methods
+    attributes = dir(obj)
+
+    # Sort them alphabetically, case-insensitive
+    sorted_attributes = sorted(attributes, key=lambda x: x.lower())
+
+    # Iterate over sorted attributes
+    for attr in sorted_attributes:
+        # Retrieve the attribute or method from the object
+        attribute_or_method = getattr(obj, attr)
+
+        # Check if it is callable (method) or not (attribute)
+        if callable(attribute_or_method):
+            print(f"Method: {attr}")
+        else:
+            print(f"Attribute: {attr}")
