+++ dev_projects/automatic_differentiation/one_stage_config.yaml
@@ -244,18 +244,20 @@ design_optimization:
 
   radius_type : "constant_mean"
   solver_options:
-    library: pygmo
-    method: snopt
-    # library: scipy
-    # method: slsqp
+    # library: pygmo
+    # method: snopt
+    library: scipy
+    method: slsqp
     derivative_method: 2-point
     derivative_abs_step: 1e-9
     print_convergence: True
     plot_convergence: False
     update_on: "gradient"
-    options:
-      tol: 1e-6
-      max_iter: 500 
+    tolerance: 1e-6
+    max_iterations: 1000
+    # extra_options:
+    #   tol: 1e-6
+    #   max_iter: 500 
 
 
 
