+++ dev_projects/automatic_differentiation/two_stage_dummy_config.yaml
@@ -76,63 +76,63 @@ design_optimization:
       lower_bound : [0.6, 0.6, 0.6, 0.6]
       upper_bound : [0.9, 0.9, 0.9, 0.9]
     hub_tip_ratio_out :
-      value : [0.715999, 0.674840]
-      lower_bound : [0.6, 0.6]
-      upper_bound : [0.9, 0.9]
+      value : [0.715999, 0.674840, 0.715999, 0.674840]
+      lower_bound : [0.6, 0.6, 0.6, 0.6]
+      upper_bound : [0.9, 0.9, 0.9, 0.9]
     aspect_ratio :
-      value : [1.285550, 1.402149]
-      lower_bound : [1.0, 1.0]
-      upper_bound : [2.0, 2.0]
+      value : [1.285550, 1.402149, 1.285550, 1.402149]
+      lower_bound : [1.0, 1.0, 1.0, 1.0]
+      upper_bound : [2.0, 2.0, 2.0, 2.0]
     pitch_chord_ratio :
-      value : [0.699312, 0.584804]
-      lower_bound : [0.5, 0.5]
-      upper_bound : [1.10, 1.10]
+      value : [0.699312, 0.584804, 0.699312, 0.584804]
+      lower_bound : [0.5, 0.5, 0.5, 0.5]
+      upper_bound : [1.10, 1.10, 1.10, 1.10]
     trailing_edge_thickness_opening_ratio :
-      value : [0.066889, 0.068007]
-      lower_bound : [0.05, 0.05]
-      upper_bound : [0.2, 0.2]
+      value : [0.066889, 0.068007, 0.066889, 0.068007]
+      lower_bound : [0.05, 0.05, 0.05, 0.05]
+      upper_bound : [0.2, 0.2, 0.2, 0.2]
     leading_edge_angle : # (-70-15)
-      value : [0.5, 0.66444444]
-      lower_bound : [0.11111111, 0.41666667]
-      upper_bound : [0.58333333, 0.88888889]
+      value : [0.5, 0.66444444, 0.5, 0.66444444]
+      lower_bound : [0.11111111, 0.41666667, 0.11111111, 0.41666667]
+      upper_bound : [0.58333333, 0.88888889, 0.58333333, 0.88888889]
     gauging_angle : # (55-75)
-      value : [0.86601513, 0.16024573]
-      lower_bound : [0.80555556, 0.08333333]
-      upper_bound : [0.91666667, 0.19444444]
+      value : [0.86601513, 0.16024573, 0.86601513, 0.16024573]
+      lower_bound : [0.80555556, 0.08333333, 0.80555556, 0.08333333]
+      upper_bound : [0.91666667, 0.19444444, 0.91666667, 0.19444444]
     v_in :
       value : 0.22715060950254562
       lower_bound : 0.001
       upper_bound : 0.5
     w_out :
-      value : [0.7560168311370046, 0.714178597145251]
-      lower_bound : [0.1, 0.1]
-      upper_bound : [1.0, 1.0]
+      value : [0.7560168311370046, 0.714178597145251, 0.7560168311370046, 0.714178597145251]
+      lower_bound : [0.1, 0.1, 0.1, 0.1]
+      upper_bound : [1.0, 1.0, 1.0, 1.0]
     s_out :
-      value : [0.02393594702543656, 0.07236629937075202]
-      lower_bound : [0.0, 0.0]
-      upper_bound : [0.32, 0.32]
+      value : [0.02393594702543656, 0.07236629937075202, 0.02393594702543656, 0.07236629937075202]
+      lower_bound : [0.0, 0.0, 0.0, 0.0]
+      upper_bound : [0.32, 0.32, 0.32, 0.32]
     beta_out :
[…]
