### 2024-11-23 019228b
Update IPOPT settings for nice termination
