### 2024-11-29 3e3afd2
Add calculation of nondimensional mass flow and speed
