+++ dev_projects/centrifugal_compressor/validation_plots_parasitic_losses.py
@@ -5,7 +5,7 @@ import numpy as np
 
 impeller = "O" # A, O
 model = "zhang" # oh, zhang
-save_figs = True
+save_figs = False
 
 # Map files
 files_exp = {
