+++ turboflow/axial_turbine/choking_criterion.py
@@ -214,6 +214,8 @@ def critical_mach_number(
     }
     throat_plane = {f"{key}_throat": val for key, val in throat_plane.items()}
 
+    
+
     return residuals_critical, {**critical_state, **throat_plane}
 
 
@@ -260,6 +262,10 @@ def get_mach_crit(gamma, eta):
         + jnp.sqrt((1 + eta) ** 2 + 4 * alpha * (1 + alpha - 3 * eta))
     ) / (4 * alpha - 2)
     Ma_crit = jnp.sqrt(2 / (gamma - 1) * (1 / T_hat_crit - 1))
+
+    print("This is Mach Critical")
+    print(Ma_crit, T_hat_crit)
+
     return Ma_crit
 
 
