### 2024-11-29 9ea7d57
add modifications to prevent NaNs in loss model and critical Mach number functions
