### 2024-11-29 d9a024a
update dev scripts to test functionality to prevent NaNs
