+++ turboflow/axial_turbine/choking_criterion.py
@@ -258,9 +258,9 @@ def get_mach_crit(gamma, eta):
     ) / (4 * alpha - 2)
     Ma_crit = jnp.sqrt(2 / (gamma - 1) * (1 / T_hat_crit - 1))
 
-    print("This is Mach Critical")
-    print(Ma_crit, T_hat_crit)
-
+    # print("This is Ma_crit")
+    # print(Ma_crit, T_hat_crit)
+    
     return Ma_crit
 
 
