+++ dev_projects/automatic_differentiation/two_stage_config.yaml
@@ -85,7 +85,7 @@ design_optimization:
       lower_bound : [0.6, 0.6, 0.6, 0.6]
       upper_bound : [0.9, 0.9, 0.9, 0.9]
     hub_tip_ratio_out :
-      value : [0.715999, 0.674840, 0.63851147, 0.597547073]
+      value : [0.715999, 0.674840, 0.63851147, 0.6100000 ] #0.597547073  ]
       # value : [0.715999, 0.674840, 0.715999, 0.674840]
       lower_bound : [0.6, 0.6, 0.6, 0.6]
       upper_bound : [0.9, 0.9, 0.9, 0.9]
@@ -110,7 +110,7 @@ design_optimization:
       lower_bound : [0.11111111, 0.41666667, 0.11111111, 0.41666667]
       upper_bound : [0.58333333, 0.88888889, 0.58333333, 0.88888889]
     gauging_angle : # (55-75)
-      value : [0.86601513, 0.16024573, 0.82190876, 0.22530754]
+      value : [0.86601513, 0.16024573, 0.82190876, 0.16024573]
       # value : [0.86601513, 0.16024573, 0.86601513, 0.16024573]
       lower_bound : [0.80555556, 0.08333333, 0.80555556, 0.08333333]
       upper_bound : [0.91666667, 0.19444444, 0.91666667, 0.19444444]
