### 2024-11-30 33c0c92
add preliminary changes to check bounds
