+++ turboflow/axial_turbine/design_optimization.py
@@ -119,6 +119,13 @@ def compute_optimal_turbine(
     # Initialize problem object
     problem = CascadesOptimizationProblem(config)
 
+    # Check that the initial guess is within bounds and clip if necessary
+    problem.initial_guess = check_and_clip_initial_guess(
+        problem.initial_guess,
+        problem.bounds,
+        problem.design_variables_keys
+    )
+
     # Perform initial function call to initialize problem
     # This populates the arrays of equality and inequality constraints
     # TODO: it might be more intuitive to create a new method called initialize_problem() that generates the initial guess and evaluates the fitness() function with it
@@ -133,60 +140,62 @@ def compute_optimal_turbine(
     # Solve optimization problem for initial guess x0
     solver.solve(problem.initial_guess)
 
-    dfs = {
-        "operation point": pd.DataFrame(
-            {key: pd.Series(val) 
-             for key, val in problem.boundary_conditions.items()},
-        ),
-        "overall": pd.DataFrame(problem.results["overall"], index=[0]),
-        "planes": pd.DataFrame(problem.results["planes"]),
-        "cascades": pd.DataFrame(problem.results["cascades"]),
-        "stage": pd.DataFrame(problem.results["stage"]),
-        "geometry": pd.DataFrame(
-            {key: pd.Series(val) for key, val in problem.geometry.items()}
-        ),
-        "solver": pd.DataFrame(
-            {
-                "completed": pd.Series(True, index=[0]),
-                "success": pd.Series(solver.success, index=[0]),
-                "message": pd.Series(solver.message, index=[0]),
-                "grad_count": solver.convergence_history["grad_count"],
-                "func_count": solver.convergence_history["func_count"],
-                "func_count_total": solver.convergence_history["func_count_total"],
-                "objective_value": solver.convergence_history["objective_value"],
-                "constraint_violation": solver.convergence_history[
-                    "constraint_violation"
-                ],
-                "norm_step": solver.convergence_history["norm_step"],
-            },
-            index=range(len(solver.convergence_history["grad_count"])),
-        ),
-    }
+    # dfs = {
+    #     "operation point": pd.DataFrame(
+    #         {key: pd.Series(val) 
+    #          for key, val in problem.boundary_conditions.items()},
+    #     ),
+    #     "overall": pd.DataFrame(problem.results["overall"], index=[0]),
+    #     "planes": pd.DataFrame(problem.results["planes"]),
+    #     "cascades": pd.DataFrame(problem.results["cascades"]),
+    #     "stage": pd.DataFrame(problem.results["stage"]),
+    #     "geometry": pd.DataFrame(
+    #         {key: pd.Series(val) for key, val in problem.geometry.items()}
+    #     ),
+    #     "solver": pd.DataFrame(
+    #         {
+    #             "completed": pd.Series(True, index=[0]),
+    #             "success": pd.Series(solver.success, index=[0]),
+    #             "message": pd.Series(solver.message, index=[0]),
+    #             "grad_count": solver.convergence_history["grad_count"],
+    #             "func_count": solver.convergence_history["func_count"],
+    #             "func_count_total": solver.convergence_history["func_count_total"],
+    #             "objective_value": solver.convergence_history["objective_value"],
+    #             "constraint_violation": solver.convergence_history[
+    #                 "constraint_violation"
+    #             ],
+    #             "norm_step": solver.convergence_history["norm_step"],
+    #         },
+    #         index=range(len(solver.convergence_history["grad_count"])),
+    #     ),
+    # }
 
-    if export_results:
-        # Create a directory to save simulation results
[…]
