+++ dev_projects/automatic_differentiation/one_stage_config.yaml
@@ -251,7 +251,7 @@ design_optimization:
 
     derivative_method: 2-point
     derivative_abs_step: 1e-9
-    print_convergence: False
+    print_convergence: True
     plot_convergence: False
     update_on: "gradient"
     tolerance: 1e-8
