### 2024-11-30 c0eeb8b
refactor source couse and prepare automation for ASME paper
