+++ dev_projects/automatic_differentiation/plot_convergence_history.py
@@ -1,0 +1,190 @@
+
+import turboflow as tf
+import numpy as np
+import matplotlib.pyplot as plt
+from mpl_toolkits.axes_grid1.inset_locator import inset_axes, mark_inset
+import matplotlib.cm as cm
+import seaborn as sns
+import os
+
+algorthims = False
+pop_size = True
+
+save_fig = True
+
+figures_dir = "figures"
+os.makedirs(figures_dir, exist_ok=True)
+
+# Main directory
+output_dir = "output"
+
+# Load object
+filename_jax_slsqp = os.path.join(output_dir, '1stage_jax_slsqp.pkl')
+filename_FD_3_slsqp = os.path.join(output_dir, '1stage_FD_1e-3_slsqp.pkl')
+filename_FD_6_slsqp = os.path.join(output_dir, '1stage_FD_1e-6_slsqp.pkl')
+filename_FD_9_slsqp = os.path.join(output_dir, '1stage_FD_1e-9_slsqp.pkl')
+
+filename_FD_3_ipopt = os.path.join(output_dir, '1stage_FD_1e-3_ipopt.pkl')
+filename_FD_6_ipopt = os.path.join(output_dir, '1stage_FD_1e-6_ipopt.pkl')
+filename_FD_9_ipopt = os.path.join(output_dir, '1stage_FD_1e-9_ipopt.pkl')
+
+filename_FD_3_snopt = os.path.join(output_dir, '1stage_FD_1e-3_snopt.pkl')
+filename_FD_6_snopt = os.path.join(output_dir, '1stage_FD_1e-6_snopt.pkl')
+filename_FD_9_snopt = os.path.join(output_dir, '1stage_FD_1e-9_snopt.pkl')
+
+
+solver_jax_slsqp = tf.load_from_dill(filename_jax_slsqp)
+solver_FD_3_slsqp = tf.load_from_dill(filename_FD_3_slsqp)
+solver_FD_6_slsqp = tf.load_from_dill(filename_FD_6_slsqp) 
+solver_FD_9_slsqp = tf.load_from_dill(filename_FD_9_slsqp)
+
+solver_FD_3_ipopt = tf.load_from_dill(filename_FD_3_ipopt)
+solver_FD_6_ipopt = tf.load_from_dill(filename_FD_6_ipopt) 
+solver_FD_9_ipopt = tf.load_from_dill(filename_FD_9_ipopt)
+
+solver_FD_3_snopt = tf.load_from_dill(filename_FD_3_snopt)
+solver_FD_6_snopt = tf.load_from_dill(filename_FD_6_snopt) 
+solver_FD_9_snopt = tf.load_from_dill(filename_FD_9_snopt)
+
+# Close open figures
+plt.close()
+
+# Access the tab10 colormap
+# cmap = plt.get_cmap('tab10')
+# colors = [cmap(i) for i in range(cmap.N)]
+# colors = plt.get_cmap('magma')(np.linspace(0.2, 1.0, 7))
+colors = sns.color_palette("colorblind", n_colors=3)  # Adjust n_colors as needed
+
+
+# Linestyles
+# linestyles = ['-', '--', ':', '-.', '-', '-', '--', ':']
+linestyles = ['-', '--', ':']
+
+def adjust_objective_values(objective_values, constraint_values):
+
+    objective_values = objective_values.copy()
+    constraint_values = constraint_values.copy()
+    for i in range(1, len(objective_values)):
+
+        if objective_values[i] > objective_values[i-1] or np.isnan(objective_values[i]) :
+            objective_values[i] = objective_values[i-1]
+            constraint_values[i] = constraint_values[i-1]
+
+    return objective_values, constraint_values
+
[…]
