### 2024-12-01 8f7a3e2
add logging capabilities and functionality to postprocess automated results run
