### 2024-12-01 0f8ae82
updated case summary files
