### 2024-12-01 203dd13
modify loss models to improve convergence in cases with tight tolerance (tip clearance, benner incidence)
