+++ turboflow/axial_turbine/loss_model_benner.py
@@ -5,7 +5,9 @@ import jax
 import jax.numpy as jnp
 
 from . import loss_model_kacker_okapuu as lm_ko
-from .loss_coefficient_conversion import convert_kinetic_energy_to_stagnation_pressure_loss
+from .loss_coefficient_conversion import (
+    convert_kinetic_energy_to_stagnation_pressure_loss,
+)
 
 # TODO: Add smoothing to the min/max/abs/piecewise functions
 
@@ -88,6 +90,7 @@ def compute_losses(input_parameters):
     Y_s = get_secondary_loss(flow_parameters, geometry, delta_height)
 
     # Tip clearance loss coefficient
+    # Y_cl = 0.0
     Y_cl = lm_ko.get_tip_clearance_loss(flow_parameters, geometry)
 
     # Incidence loss coefficient
@@ -95,6 +98,7 @@ def compute_losses(input_parameters):
     Y_inc = get_incidence_loss(flow_parameters, geometry, beta_des)
 
     # Penetration depth to blade height ratio
+    # ZTE = 0.0
     ZTE = get_penetration_depth(flow_parameters, geometry, delta_height)
 
     # Correct profile losses according to penetration depth
@@ -207,93 +211,28 @@ def get_incidence_loss(flow_parameters, geometry, beta_des):
         Incidence loss coefficient.
     """
 
+    # Rename parameters
     beta_in = flow_parameters["beta_in"]
     gamma = flow_parameters["gamma_out"]
     Ma_rel_out = flow_parameters["Ma_rel_out"]
-
     le = geometry["leading_edge_diameter"]
     s = geometry["pitch"]
     We = geometry["leading_edge_wedge_angle"]
     theta_in = geometry["leading_edge_angle"]
     theta_out = math.arccosd(geometry["A_throat"] / geometry["A_out"])
+    type = geometry["cascade_type"]
 
-    chi = get_incidence_parameter(le, s, We, theta_in, theta_out, beta_in, beta_des)
-
-    dPhip = get_incidence_profile_loss_increment(chi)
-
-    Y_inc = convert_kinetic_energy_to_stagnation_pressure_loss(Ma_rel_out, gamma, dPhip, limit_output=True)
+    # Evaluate incidence model
+    chi = get_incidence_parameter(
+        le, s, We, theta_in, theta_out, beta_in, beta_des, type
+    )
+    delta_phi2 = get_incidence_profile_loss_increment(chi)
+    # Y_inc = delta_phi2
+    Y_inc = convert_kinetic_energy_to_stagnation_pressure_loss(
+        Ma_rel_out, gamma, delta_phi2, limit_output=True
+    )
 
     return Y_inc
-
-
-def get_penetration_depth(flow_parameters, geometry, delta_height):
-    r"""
-    Calculated the penetration depth of the passage vortex separation line relative to the blade span.
-
-    The endwall inlet boundary layer generate a vortex which propagtes through the cascade section,
-    and the spanwise penetration of this vortex affect the magnutude of the secondary loss coefficient.
-    This function approximate the vortex penetration depth to the blade span by the correlation developed by :cite:`benner_empirical_2006-1`.
-
-    The quantity is calculated as:
-
-    .. math::
-
-        \frac{\mathrm{Z_{te}}}{H} = \frac{0.10F_t^{0.79}}{\sqrt{CR}\left(\frac{H}{c}\right)^{0.55}} + 32.70\frac{\delta^*}{H}^2
-
-    where:
-
-        - :math:`CR = \frac{\cos(\beta_\mathrm{in})}{\cos(\beta_\mathrm{out})}` is the convergence ratio.
[…]
