### 2024-12-04 d649c62
add dense Hessians computation test
