+++ turboflow/axial_turbine/design_optimization.py
@@ -537,6 +537,21 @@ class CascadesOptimizationProblem(psv.OptimizationProblem):
         return grad
 
 
+    def hessians(self, x):
+        return compute_hessians_jax(self.fitness, x, lower_triangular=True)
+
+
+    def hessians_jax(self, x, lower_triangular=True):
+        return compute_hessians_jax(self.fitness, x, lower_triangular=lower_triangular)
+
+
+    def hessians_approx(self, x):
+        H = psv.approx_jacobian_hessians(
+            self.fitness, x, abs_step=1e-4, lower_triangular=True
+        )
+        return H
+
+
     def get_objective_function(self, objective):
         """
         Change scale for the objective function depending on its type.
@@ -1165,3 +1180,50 @@ def check_and_clip_initial_guess(initial_guess, bounds, variable_names):
             )
 
     return initial_guess
+
+
+
+
+def compute_hessians_jax(fitness_function, x, lower_triangular=True):
+    """
+    Compute the Hessians of a vector-valued fitness function.
+
+    Parameters
+    ----------
+    fitness_function : callable
+        The vector-valued function for which to compute the Hessians.
+    x : array-like
+        The point at which to compute the Hessians.
+    lower_triangular : bool, optional
+        If True, return the Hessians in lower triangular format. Default is True.
+
+    Returns
+    -------
+    Hessians : ndarray
+        A tensor where each slice along the first dimension corresponds to the Hessian
+        of a component of the fitness function. If lower_triangular is True, the Hessians
+        are returned in a flattened lower triangular format.
+    """
+    # Compute full Hessians with JAX
+    H_full = jax.jacfwd(jax.jacrev(fitness_function))(x)
+    # H_full = jax.jacrev(jax.jacfwd(fitness_function))(x)
+
+    # Convert to NumPy for post-processing
+    H_full = np.array(H_full)
+
+    if lower_triangular:
+        # Extract the lower triangular part
+        n = H_full.shape[1]
+        k = (n * (n + 1)) // 2  # Number of elements in the lower triangular part
+        H_lower_triangular = np.zeros((H_full.shape[0], k))
+
+        # Loop over all components of the fitness
+        for i in range(H_full.shape[0]):
+            H_lower_triangular[i] = H_full[i][np.tril_indices(n)]
+
+        return H_lower_triangular
+
+    return H_full
+
+
+
