### 2024-12-08 #6979 @lasseband
Hi @philipcardiff,

We have contacted a potential reviewer who agreed to participate. His experience within turbomachinery and meanline modeling makes him suitable for this review. Could you contact him and add him to this issue?
Name: Alessandro Romei
Affiliation: Politecnico di Milano

Do we need any additional reviewers?

Best,
Lasse
