### 2024-12-09 #6979 @philipcardiff
Thanks, Lasse. Yes, I will contact Alessandro.
