### 2024-12-19 6234f4f
update vaneless diffuser model script
