+++ dev_projects/centrifugal_compressor/compressor_model/vaneless_diffuser/vaneless_diffuser_model.py
@@ -1,23 +1,38 @@
 
-import numpy as np
-import CoolProp as cp
 import scipy.linalg
 import scipy.integrate
+import scipy.optimize
+import numpy as np
+import CoolProp as cp
+import matplotlib.pyplot as plt
 
 
-def radial_vaneless_diffuser_1d(
-        fluid,
+def evaluate_vaneless_diffuser_1d(
+        p0_in,
+        T0_in,
+        Ma_in,
+        alpha_in,
         Cf,
         q_w,
-        inlet_state,
+        fluid_name,
         diffuser_geometry,
+        number_of_points=None,
+        tol=1e-6,
         ):
-    
-    # Rename inlet state properties
-    v_m_in = inlet_state["v_m"]
-    v_t_in = inlet_state["v_t"]
-    d_in = inlet_state["d"]
-    p_in = inlet_state["p"]
+    """Evaluate one-dimensional flow in an generic annular duct"""
+
+    # Create fluid object
+    fluid = cp.AbstractState("HEOS", fluid_name)
+
+    # Compute inlet static state
+    p_in, s_in = compute_inlet_static_state(p0_in, T0_in, Ma_in, fluid)
+    fluid.update(cp.PSmass_INPUTS, p_in, s_in)
+    d_in = fluid.rhomass()
+    p_in = fluid.p()
+    a_in = fluid.speed_sound()
+    v_in = Ma_in*a_in
+    v_m_in = v_in*np.cos(alpha_in)
+    v_t_in = v_in*np.sin(alpha_in)
 
     # Rename geometry
     phi = diffuser_geometry["phi"]
@@ -33,7 +48,7 @@ def radial_vaneless_diffuser_1d(
 
         # Rename from ODE terminology to physical variables
         m = t     
-        v_m, v_t, d, p = y
+        v_m, v_t, d, p, s_gen, theta = y
 
         # Calculate velocity
         v = np.sqrt(v_t**2 + v_m**2)
@@ -60,9 +75,11 @@ def radial_vaneless_diffuser_1d(
         a = fluid.speed_sound()
         h = fluid.hmass()
         s = fluid.smass()
+        T = fluid.T()
         h0 = h + 0.5*v**2
 
         # Stress at the wall
+        # TODO implement function for friction factor
         tau_w = Cf*d*v**2/2
 
         # Compute coefficient matrix
@@ -87,89 +104,265 @@ def radial_vaneless_diffuser_1d(
 
         dy = scipy.linalg.solve(M, S)
 
+
+        # Compute entropy generation rate
+        sigma = 2/b*(tau_w*v)
+        dy = np.append(dy, sigma/(d*v_m)/T)
+        
+        # Compute streamline wrapping) angle (only for phi=pi/2)
+        dy = np.append(dy, (v_t/v_m)/r)
+
+        # Store data in dictionary
         out = {"v_t" : v_t,
                "v_m" : v_m,
                "v" : v,
+               "Ma" : v / a,
+               "Ma_m" : v_m / a,
+               "Ma_t" : v_t / a,
                "alpha" : alpha,
                "d" : d,
                "p" : p,
                "s" : s,
+               "T" : T,
+               "s_gen" : s_gen,
                "h" : h,
-               "h0" : h0}
+               "h0" : h0,
+               "theta" : theta,
+               "r": r,
+               "b": b,
+               "m": m,
+               "radius_ratio": r/r_in,
+               "area_ratio": (b*r)/(b_in*r_in),
+               "Cp": (p-p_in)/(p0_in-p_in),
+               "Re": "",
+               }
 
         return dy, out
-    
-    solution = scipy.integrate.solve_ivp(
-        lambda t,y: odefun(t,y)[0],
-        [0, m_out],
-        [v_m_in, v_t_in, d_in, p_in],
+
[…]
