### 2025-01-04 4ccf59d
correction of gradient free optimization
