### 2025-01-04 32cab9d
update convergence plots
