### 2025-01-04 c3e9f85
update optimization algos running scipts
