+++ dev_projects/optimization_algos/kofskey1972_1stage.yaml
@@ -25,11 +25,11 @@ performance_analysis :
     alpha_in: 0 
   solver_options: 
     method: hybr  
-    tolerance: 1e-8  
+    tolerance: 1e-6  
     max_iterations: 100  
     derivative_method: 2-point  
     derivative_abs_step: 1e-6
-    print_convergence : False  
+    print_convergence : True  
     plot_convergence: False
   initial_guess :
     n_samples : 100
@@ -73,64 +73,127 @@ design_optimization:
     scale : 1
     type : maximize
   variables :
+    # specific_speed : # Limited 
+    #   value : 0.713756
+    #   lower_bound : 0.5
+    #   upper_bound : 1.5
+    # blade_jet_ratio : # Limited
+    #   value : 0.466501
+    #   lower_bound : 0.3
+    #   upper_bound : 0.8
+    # hub_tip_ratio_in :
+    #   value : [0.715999, 0.715999]
+    #   lower_bound : [0.6, 0.6]
+    #   upper_bound : [0.9, 0.9]
+    # hub_tip_ratio_out :
+    #   value : [0.715999, 0.674840]
+    #   lower_bound : [0.6, 0.6]
+    #   upper_bound : [0.9, 0.9]
+    # aspect_ratio :
+    #   # value : [1.285550, 1.402149]
+    #   value : [1.285550, 2.0]
+    #   lower_bound : [1.0, 1.0]
+    #   upper_bound : [2.0, 2.0]
+    # pitch_chord_ratio :
+    #   value : [0.699312, 0.584804]
+    #   lower_bound : [0.5, 0.5]
+    #   upper_bound : [1.10, 1.10]
+    # trailing_edge_thickness_opening_ratio : 
+    #   value : [0.066889, 0.068007]
+    #   lower_bound : [0.05, 0.05]
+    #   upper_bound : [0.2, 0.2]
+    # leading_edge_angle : # (-70-15)
+    #   # value : [0.5, 0.66444444]
+    #   value : [0.58333333, 0.66444444]
+    #   lower_bound : [0.11111111, 0.41666667]
+    #   upper_bound : [0.58333333, 0.88888889]
+    # gauging_angle : # (55-75)
+    #   value : [0.86601513, 0.16024573]
+    #   lower_bound : [0.80555556, 0.08333333]
+    #   upper_bound : [0.91666667, 0.19444444]
+    # v_in :
+    #   value : 0.22715060950254562
+    #   lower_bound : 0.001
+    #   upper_bound : 0.5
+    # w_out : 
+    #   value : [0.7560168311370046, 0.714178597145251]
+    #   lower_bound : [0.1, 0.1]
+    #   upper_bound : [1.0, 1.0]
+    # s_out : 
+    #   value : [0.02393594702543656, 0.07236629937075202]
+    #   lower_bound : [0.0, 0.0]
+    #   upper_bound : [0.32, 0.32]
+    # beta_out : 
+    #   value : [0.8650417242559838, 0.16150395974996423]
+    #   lower_bound : [0.72, 0.06]
+    #   upper_bound : [0.94, 0.28]
+    # w_crit_throat : 
+    #   value : [0.954641310153918, 0.8901549785503726]
+    #   lower_bound : [0.1, 0.1]
+    #   upper_bound : [1.0, 1.0]
+    # s_crit_throat : 
+    #   value : [0.04006361313218885, 0.09614205881131721]
+    #   lower_bound : [0.0, 0.0]
+    #   upper_bound : [0.32, 0.32]
     specific_speed : # Limited 
-      value : 0.713756
+      value : 0.5719122380735249
       lower_bound : 0.5
       upper_bound : 1.5
     blade_jet_ratio : # Limited
[…]
