+++ dev_projects/Kofskey1972_one_stage/validation_critical_mach.py
@@ -1,0 +1,169 @@
+
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+import turboflow as tf
+
+show_figures = True
+save_figs = True
+error_bars = False
+
+# Load performance data
+filename_1 = "output/performance_analysis_2024-09-05_11-28-10.xlsx" # Critical mach
+filename_2 = "output/performance_analysis_2024-03-13_23-14-55.xlsx" # Critical mass flow
+
+data_1 = tf.plot_functions.load_data(filename_1)
+data_2 = tf.plot_functions.load_data(filename_2)
+
+subsets = ["omega"] + list(1627*np.array([0.7, 1.0]))
+colors = plt.get_cmap('magma')(np.linspace(0.2, 0.7, 2))
+fig1, ax1 = tf.plot_functions.plot_lines(
+    data_1,
+    x_key="PR_ts",
+    y_keys=["mass_flow_rate"],
+    subsets=subsets,
+    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    ylabel="Mass flow rate [kg/s]",
+    colors=colors,
+    linestyles=['-']*2,
+    filename = 'mass_flow_rate',
+    outdir = "figures",
+    save_figs=False,
+)
+
+fig1, ax1 = tf.plot_functions.plot_lines(
+    data_2,
+    x_key="PR_ts",
+    y_keys=["mass_flow_rate"],
+    subsets=subsets,
+    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    ylabel="Mass flow rate [kg/s]",
+    colors=colors,
+    linestyles=[":"]*2,
+    fig = fig1,
+    ax = ax1, 
+    filename = 'mass_flow_rate',
+    outdir = "figures",
+    save_figs=False,
+)
+
+fig2, ax2 = tf.plot_functions.plot_lines(
+    data_1,
+    x_key="PR_ts",
+    y_keys=["torque"],
+    subsets=subsets,
+    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    ylabel="Torque [Nm]",
+    colors=colors,
+    linestyles=['-']*2,
+    filename = 'torque',
+    outdir = "figures",
+    save_figs=False,
+)
+
+fig2, ax2 = tf.plot_functions.plot_lines(
+    data_2,
+    x_key="PR_ts",
+    y_keys=["torque"],
+    subsets=subsets,
+    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    ylabel="Torque [Nm]",
+    colors=colors,
+    linestyles=[":"]*2,
+    fig = fig2,
+    ax = ax2, 
+    filename = 'torque',
+    outdir = "figures",
+    save_figs=False,
+)
+
+# Add experimental points
+filename_exp = "experimental_data_Kofskey1972_1stage_interpolated.xlsx"
+data_exp = pd.read_excel(filename_exp, sheet_name = ["Sheet1"])
+
+subsets = ["speed_percent"] + list(np.array([70, 100]))
+fig1, ax1 = tf.plot_functions.plot_lines(
+    data_exp,
+    x_key="pressure_ratio_ts",
+    y_keys=["mass_flow_rate"],
+    subsets=subsets,
+    xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
+    ylabel="Mass flow rate [kg/s]",
+    colors=colors,
+    linestyles = ['none']*2,
+    markers = ["x"]*2,
[…]
