+++ dev_projects/dev_scripts/loss_model_development/incidence_loss_correction.py
@@ -26,27 +26,40 @@ loss_3 = tf.get_incidence_profile_loss_increment(
 # Create figure
 fig = plt.figure(figsize=(6.4, 4.8))
 ax = fig.gca()
-ax.set_title("Benner incidence loss model")
-ax.set_xlabel("$\chi$ - Distance parameter")
-ax.set_ylabel("$\Delta \phi_{p}^2$ - Loss coefficient increment")
+# ax.set_title("Benner incidence loss model")
+# ax.set_xlabel("$\chi$ - Distance parameter")
+# ax.set_ylabel("$\Delta \phi_{p}^2$ - Loss coefficient increment")
 ax.set_xscale("linear")
 ax.set_yscale("linear")
 ax.set_ylim([0, 1])
 
 # Plot simulation data
-ax.plot(chi, loss_1, linewidth=1.25, label="Original correlation")
-ax.plot(chi, loss_2, linewidth=1.25, label="Linear extrapolation")
-ax.plot(chi, loss_3, linewidth=1.25, label="Extrapolation and limiting")
+ax.plot(chi, loss_1, 'k', linewidth=1.25, label="Original correlation")
+# ax.plot(chi, loss_2, linewidth=1.25, label="Linear extrapolation")
+# ax.plot(chi, loss_3, linewidth=1.25, label="Extrapolation and limiting")
 
 # Create legend
-leg = ax.legend(loc="upper left")
+# leg = ax.legend(loc="upper left")
 
 # Adjust PAD
 fig.tight_layout(pad=1, w_pad=None, h_pad=None)
 
+# ax.set_xlabel("Incidence angle")
+ax.set_ylabel("Incidence losses", fontsize = 16)
+ax.set_ylim([0,0.1])
+ax.set_xlim([-5,5])
+tick_positions = [-2,0,2]
+tick_labels = ["Negative incidence", "0", "Positive incidence"]
+ax.set_xticks(tick_positions)
+ax.set_xticklabels(tick_labels, fontsize = 16)
+ax.set_yticklabels([])
+ax.tick_params(axis = "x", which = "both", bottom = False, top = False)
+ax.tick_params(axis = "y", which = "both", bottom = False, top = False)
+ax.grid(False)
+
 # Save plots
-filename = os.path.join(fig_dir, "incidence_loss_modification")
-fig.savefig(filename + ".png", bbox_inches="tight")
+filename = os.path.join(fig_dir, "incidence_losses")
+# fig.savefig(filename + ".png", bbox_inches="tight")
 fig.savefig(filename + ".svg", bbox_inches="tight")
 
 # Show figure
