### 2025-01-15 afbbeb3
update examples
