+++ examples/one-stage turbine/one_stage_plots.py
@@ -5,16 +5,13 @@ Created on Wed Oct  4 13:27:39 2023
 @author: laboan
 """
 
-import numpy as np
-import pandas as pd
 import matplotlib.pyplot as plt
-from scipy.optimize import curve_fit
 import turboflow as tf
 
 RESULTS_PATH = "output"
 
 # Get the name of the latest results file
-filename = tf.utils.find_latest_results_file(RESULTS_PATH)
+filename = tf.utilities.find_latest_results_file(RESULTS_PATH)
 
 save_figs = False
 show_figures = True
@@ -31,7 +28,7 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     colors='k',
     filename = 'mass_flow_rate',
     outdir = "figures",
-    save_figs=True,
+    save_figs=save_figs,
 )
 
 # Print design speed line
@@ -47,7 +44,7 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     labels = labels,
     filename='static_pressure',
     outdir = "figures",
-    save_figs=True,
+    save_figs=save_figs,
 )
 
 # Print stacked losses
@@ -69,7 +66,7 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     stack=True,
     filename="loss_coefficients",
     outdir="figures",
-    save_figs = True,
+    save_figs = save_figs,
 )
 
 
