+++ examples/one-stage turbine/one-stage_config.yaml
@@ -147,6 +147,14 @@ design_optimization:
       type : "="
       value : 1
       normalize : True
+    - variable : geometry.flaring_angle 
+      type : "<"
+      value : 10
+      normalize : True
+    - variable : geometry.flaring_angle 
+      type : ">"
+      value : -10
+      normalize : True
 
   radius_type : "constant_mean"
   solver_options:
