+++ turboflow/__init__.py
@@ -30,7 +30,7 @@ set_plot_options()
 
 
 # Package info
-__version__ = "0.1.12"
+__version__ = "0.1.13"
 PACKAGE_NAME = "turboflow"
 URL_GITHUB = "https://github.com/turbo-sim/turboflow"
 URL_DOCS = "https://turbo-sim.github.io/turboflow/"
