+++ turboflow/__init__.py
@@ -30,7 +30,7 @@ set_plot_options()
 
 
 # Package info
-__version__ = "0.1.14"
+__version__ = "0.1.15"
 PACKAGE_NAME = "turboflow"
 URL_GITHUB = "https://github.com/turbo-sim/turboflow"
 URL_DOCS = "https://turbo-sim.github.io/turboflow/"
