+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "turboflow"
-version = "0.1.14"
+version = "0.1.15"
 description = "A Python tool for meanline analysis and optimization of turbomachinery."
 authors = ["Lasse Borg & Roberto Agromayor"]
 license = "MIT"
