### 2025-01-16 #7588 @lasseband
Hi @philipcardiff,

The missing DOIs should have been added now.
