### 2025-01-22 #7588 @aromei
Hi @lasseband,
I am trying to install the software, but I encounter an error when I run 'pip install turboflow'

ERROR: Ignored the following versions that require a different python version: 0.1.0 Requires-Python <4.0,>=3.11; 0.1.1 Requires-Python <4.0,>=3.11; 0.1.10 Requires-Python <4.0,>=3.11; 0.1.11 Requires-Python <4.0,>=3.11; 0.1.12 Requires-Python <4.0,>=3.11; 0.1.14 Requires-Python <4.0,>=3.11; 0.1.15 Requires-Python <4.0,>=3.11; 0.1.2 Requires-Python <4.0,>=3.11; 0.1.3 Requires-Python <4.0,>=3.11; 0.1.4 Requires-Python <4.0,>=3.11; 0.1.9 Requires-Python <4.0,>=3.11
ERROR: Could not find a version that satisfies the requirement turboflow (from versions: none)
ERROR: No matching distribution found for turboflow
