### 2025-01-23 #7588 @aromei
I’d like to give credit to the authors, as the code appears to be highly efficient in the optimization process, effectively integrating the gradient-based solver in the meanline code. Specifically focusing on the axial turbine, the code is comprehensive, covering design, performance analysis, and optimization. Overall, the usage was quite intuitive, and I did not encounter any significant issues when running the code or making minor changes to test the model's capabilities.

# General checks
- **License** I am not entirely sure, but shouldn’t the copyright belong to all authors rather than being assigned only to the main author?

# Documentation
- **Community guidelines** @philipcardiff I am uncertain about this point in the reviewer’s checklist, as there are no clear guidelines provided. However, since the software is hosted on GitHub, I believe the community can use this platform to contribute to the code directly.
- **Installation instructions** about dependencies, I needed to install pytest to run the test.
- **Functionality documentation** It is unclear which types of turbomachinery the software can handle. Based on the documentation, it appears to be limited to axial turbines. But, the code also contains folders related to centrifugal compressors, which are not documented.
- **Contribution and authorship**  Two of the four listed authors have not made any commits to the project, while there are two contributors who do not appear on the author list. I just wanted to double-check that the author list accurately reflects the contributions made to the software.

# Software paper
- **Summary** The authors can spend a few sentences to clearly state what is a meanline code to help a non-specialist audience.
- **Summary** The authors state that the software includes meanline modeling for turbomachinery. Since "turbomachinery" is a broad term, it raises the question of whether axial compressors and radial turbines are also supported. Overall, the summary lacks specificity and does not accurately reflect the software's capabilities.
- **Summary** Link to ‘documentation pages’ is not working
- Double-check the text with a grammar check, because I found spurious typos (e.g., line 14 strcuture).

# General comment on the code
- I would recommend excluding the dev_projects folder from the main repository, as it seems to contain untested or under-development features, which may create confusion.
- The code architecture is somewhat unclear, as it lacks common coding conventions such as a src folder and a docs folder. Additionally, some folders and files, such as .vscode, could be excluded and added to the .gitignore file to reduce clutter.
- Overall, it is unclear whether the software includes additional functionalities beyond the meanline code for axial turbines. The documentation focuses solely on axial turbines, yet the example folder contains a case for a centrifugal compressor. Additionally, the src folder (turboflow) includes directories related to cycle calculations, which differ from the scope of the software (meanline code for turbomachinery).
- It seems that metastable properties are implemented in fluid_properties.py. How can the code be instructed to consider either metastable or equilibrium properties? It does not appear that there is a line in the configuration file to specify this. Moreover, if I attempt to run fluid_properties.py to execute the calculation example under __name__ == "__main__", it returns an error, likely related to issues with relative or absolute paths.
- The code does not generate a proper log file. While printing everything to the console is fine, it would be helpful to also have an option to save the output to a plain log file. This could be controlled via a flag in the input file to specify whether logging is desired. The same suggestion applies to the results.
- I encountered issues running the Kofskey1974_stator validation test, which gave me two validation errors related to the config file.
- Please organize the folders more systematically, as there are numerous scattered files. For example, some figures are not placed in the figure folder within the validation tests, and there are old or redundant folders within other directories.
- While running the example for the two-stage turbine and plotting the results, I observed an unusual trend in the incidence loss, even though the plot matches the one provided in the repository. According to the log summary, the simulation was successful for all 40 points. Is this behavior (incidence loss going to zero for two distinct operating points) expected?
- The code is intended to be general regarding thermodynamic modeling. However, in the calculation of the critical Mach number, it relies on an equation specific to perfect-gas modeling, as demonstrated by the use of the specific heat ratio. A similar dependency can be seen in the get_flow_capacity function.
- Can you confirm whether the plotting utility plot_axial_radial_plane works only for optimization? I attempted to use it with the two-stage turbine example in performance analysis mode to plot the meridional channel, but it returned an error. It would be helpful if the meridional channel could be plotted regardless of the specific solver used.
- I suggest expanding the examples available. Currently, the repository includes a single-stage and a two-stage turbine, both operating with air under ideal gas conditions. It would be valuable to include a case featuring a turbine operating with a fluid where non-ideal effects play a significant role.
- Does the code handle supersonic cascades? Based on the documentation and the code itself, it seems that it does not, as there are no mentions of shock losses, supersonic deviation, or unique incidence. If this is the case, the software paper and documentation should explicitly report this limitation.
