+++ examples/one-stage turbine/one-stage_main.py
@@ -7,8 +7,8 @@ import matplotlib.pyplot as plt
 import turboflow as tf
 
 # Define mode 
-# MODE = "design_optimzation"
-MODE = "performance_analysis"
+MODE = "design_optimzation"
+# MODE = "performance_analysis"
 
 # Load configuration file
 CONFIG_FILE = os.path.abspath("one-stage_config.yaml")
@@ -37,6 +37,4 @@ elif MODE == "design_optimzation":
 
     # Compute optimal turbine
     operation_points = config["operation_points"]
-    solver = tf.compute_optimal_turbine(config, export_results=True)
-    fig, ax = tf.plot_functions.plot_axial_radial_plane(solver.problem.geometry)
-    fig, ax = tf.plot_functions.plot_velocity_triangles(solver.problem.results["plane"])
+    solver = tf.compute_optimal_turbine(config, export_results=False)
