### 2025-01-28 e1832a4
add jax/coolprop examples from Simone
