+++ dev_projects/automatic_differentiation/jax_implementation/jax_trials.py
@@ -1,0 +1,163 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Mar 28 19:28:25 2024
+
+@author: sipar
+"""
+
+import jax
+import jax.numpy as jnp
+import CoolProp.CoolProp as cp
+
+DT=cp.DmassT_INPUTS
+
+#%% PART 1: Analytical and callback equations of state
+
+#By default jax uses 32 bit float. Force 64 bit.
+jax.config.update("jax_enable_x64", True)
+
+#Define fluid in CoolProp
+rhoT0=jnp.array([100.0,310.0])
+fluid=cp.AbstractState('HEOS','CO2')
+fluid.update(DT,rhoT0[0],rhoT0[1])
+
+#Parameters for Peng Robinson
+fld={'acentric':fluid.acentric_factor(),
+     'R':fluid.gas_constant(),
+     'Tc':fluid.T_critical(),
+     'Pc':fluid.p_critical(),
+     'Pr':fluid.Prandtl(),
+     'k' :fluid.conductivity(),
+     'mu':fluid.viscosity(),
+     'cp0':fluid.cp0molar(),
+     'MM':fluid.molar_mass()}
+
+#define Peng-Robinson EOS (note: copied from Wikipedia, do not use for anything serious)
+def PR(rhoT):
+    rho=rhoT[0]
+    T=rhoT[1]
+    R=fld['R']
+    Tc=fld['Tc']
+    Vm=fld['MM']/rho
+    ηc=(1+(4-8**(1/2))**(1/3)+(4+8**(1/2))**(1/3))**-1
+    Ωa=(8+40*ηc)/(49-37*ηc)
+    a=Ωa*(R*Tc)**2/fld['Pc']
+    Ωb=ηc/(3+ηc)
+    b=Ωb*(R*Tc)/fld['Pc']
+    Tr=T/fld['Tc']
+    κ=0.37464+1.54226*fld['acentric']-0.26992*fld['acentric']**2
+    α=(1+κ*(1-Tr**(1/2)))**2
+    P=(fld['R']*T)/(Vm-b)-a*α/(Vm**2+2*b*Vm-b**2)
+    Pr=P/fld['Pc']
+    hid=T*fld['cp0']/fld['MM']
+    Z=P*Vm/(R*T)
+    B=0.07780*Pr/Tr
+    h=hid+R/fld['MM']*Tc*(Tr*(Z-1)-2.078*(1+κ)*α**0.5*jnp.log((Z+2.414*B)/(Z-0.414*B)))
+    sid=fld['cp0']/fld['MM']*jnp.log(T)-fld['R']/fld['MM']*jnp.log(P)
+    s=sid+R/fld['MM']*(jnp.log(Z-B)-2.078*κ*((1+κ)/Tr**0.5-κ)*jnp.log((Z+2.414*B)/(Z-0.414*B)))
+    return jnp.array([P,h,s,fld['k'],fld['mu']])
+
+#define HEOS function calling CoolProp
+def HEOS_callback(rhoT):
+    fluid.update(cp.DmassT_INPUTS,rhoT[0].astype(float),rhoT[1].astype(float))
+    mu=fld['mu']+0*rhoT[1] #This could be e.g. Sutherland like
+    return jnp.array([fluid.p(),fluid.hmass(),fluid.smass(),fld['k'],mu])
+
+#wrap the callback in jax format. The decorator jax.custom_jvp is for custom gradients
+@jax.custom_jvp
+def HEOS(rhoT):
+    result_shape=jax.ShapeDtypeStruct((5,),jnp.dtype('float'))
[…]
