+++ dev_projects/automatic_differentiation/jax_implementation/jax_bicubic_HEOS_interpolation.py
@@ -1,0 +1,401 @@
+#%% INITIALIZATION
+
+import os,psutil
+#Look at the number of physical cores.
+#NOTE: On modern machines might want to use the number of high performance
+#cores, because different core architectures can cause problems in parallel
+#processing
+NCORES=psutil.cpu_count(logical=False)
+#Before loading jax, force it to see the CPU count we want
+os.environ["XLA_FLAGS"]="--xla_force_host_platform_device_count=%d"%NCORES
+import jax
+#By default jax uses 32 bit, for scientific computing we need 64 bit precisin
+jax.config.update("jax_enable_x64", True)
+import jax.numpy as jnp
+#To read more regarding the automatic parallelization in jax
+#https://jax.readthedocs.io/en/latest/sharded-computation.html
+from jax.experimental import mesh_utils
+from jax.sharding import Mesh, PartitionSpec , NamedSharding
+
+import CoolProp.CoolProp as cp
+import matplotlib.pyplot as plt
+import numpy as np
+import time
+
+float64=jnp.dtype('float64')
+complex128=jnp.dtype('complex128')
+
+#%% ################# USER SETUP #########################
+#Select the property to interpolate. For example iDmass, iSmass and iT
+#The variable names use the symbol D, but can represent any required 
+#keyed output by changing this line.
+iD=cp.iSmass
+#Grid size in h direction
+N=50
+#Grid size in log(P) direction
+M=50
+#Fluid selection
+name="CO2"
+#Grid boundaries
+hmin=1e5
+hmax=10e5
+Pmin=1e6
+Pmax=1e8
+#Number of random points to test the error
+Npoints=50000
+#Number of times to repeat the computation. used to stabilize the timing of  
+#very fast computations when using time.time()
+Nrepeats=int(1e6/Npoints)
+##########################################################
+#Full evaluation of Helmholtz equation of state
+f=cp.AbstractState("HEOS",name)
+#Use bicubic table
+ftab=cp.AbstractState("BICUBIC&HEOS",name)
+#%% FUNCTION TO COMPUTE INTERPOLATION COEFFICIENTS
+#For more details, see:https://en.wikipedia.org/wiki/Bicubic_interpolation
+#In bicubic interpolation, approximate a generic function with a piecewise
+#cubic polynomial in the form:
+#z= aij * x**i * y**j with i,y =[0,1,2,3]
+#The 16 aij coefficients can be computed from the 4 corners of the cell using:
+#   -The value of the function f
+#   -The derivative in the x direction fx
+#   -The derivative in the y direction fy
+#   -The cross derivative fxy
+#We can compute on the fly the coefficients, or compute them in advance for all
+#the cells in the grid. If we use the compuation in advance we can save time 
+#during execution, but the storage and memory requirement increases by a factor
+#of 4x. For an uncompressed 64bit 100*100 grid, storing the function and
+#derivatives takes 320kB per variable, while storing the cubic coefficients
+#takes 1.28MB.
+
+@jax.jit
[…]
