+++ turboflow/properties/fluid_properties.py
@@ -1293,112 +1293,4 @@ def calculate_subcooling(state, fluid):
     return state
 
 
-# def compute_properties_metastable_rhoT(rho, T, fluid):
-#     """
-#     Compute the thermodynamic properties of a fluid using the Helmholtz
-#     energy equation of state. All properties thermodynamic properties can
-#     be derived as combinations of the Helmholtz energy and its
-#     derivatives with respect to density and pressure.
 
-#     This function can be used to estimate metastable properties using the
-#     equation of state beyond the saturation lines.
-#     """
-
-#     # Update thermodynamic state
-#     fluid.update(CP.DmassT_INPUTS, rho, T)
-
-#     # Get fluid constant properties
-#     R = fluid.gas_constant()
-#     M = fluid.molar_mass()
-#     T_crit = fluid.T_critical()
-#     rho_crit = fluid.rhomass_critical()
-
-#     # Compute reduced variables
-#     tau = T_crit / T
-#     delta = rho / rho_crit
-
-#     # Compute from the Helmholtz energy derivatives
-#     alpha = fluid.alpha0() + fluid.alphar()
-#     dalpha_dTau = fluid.dalpha0_dTau() + fluid.dalphar_dTau()
-#     dalpha_dDelta = fluid.dalpha0_dDelta() + fluid.dalphar_dDelta()
-#     d2alpha_dTau2 = fluid.d2alpha0_dTau2() + fluid.d2alphar_dTau2()
-#     d2alpha_dDelta2 = fluid.d2alpha0_dDelta2() + fluid.d2alphar_dDelta2()
-#     d2alpha_dDelta_dTau = fluid.d2alpha0_dDelta_dTau() + fluid.d2alphar_dDelta_dTau()
-
-#     # Compute thermodynamic properties from Helmholtz energy EOS
-#     properties = {}
-#     properties['T'] = T
-#     properties['p'] = (R/M)*T*rho*delta*dalpha_dDelta
-#     properties['rhomass'] = rho
-#     properties['umass'] = (R/M)*T*(tau*dalpha_dTau)
-#     properties['hmass'] = (R/M)*T*(tau*dalpha_dTau+delta*dalpha_dDelta)
-#     properties['smass'] = (R/M)*(tau*dalpha_dTau - alpha)
-#     properties['gibbsmass'] = (R/M)*T*(alpha + delta*dalpha_dDelta)
-#     properties['cvmass'] = (R/M)*(-tau**2*d2alpha_dTau2)
-#     properties['cpmass'] = (R/M)*(-tau**2*d2alpha_dTau2 + (delta*dalpha_dDelta - delta*tau*d2alpha_dDelta_dTau)**2/(2*delta*dalpha_dDelta + delta**2*d2alpha_dDelta2))
-#     properties['gamma'] = properties['cpmass']/properties['cvmass']
-#     properties['compressibility_factor'] = delta*dalpha_dDelta
-#     properties['speed_sound'] = ((R/M)*T*(2*delta*dalpha_dDelta + delta**2*d2alpha_dDelta2 - (delta*dalpha_dDelta - delta*tau*d2alpha_dDelta_dTau)**2/(tau**2*d2alpha_dTau2)))**0.5
-#     properties['isentropic_bulk_modulus'] = rho*(R/M)*T*(2*delta*dalpha_dDelta + delta**2*d2alpha_dDelta2 - (delta*dalpha_dDelta - delta*tau*d2alpha_dDelta_dTau)**2/(tau**2*d2alpha_dTau2))
-#     properties['isentropic_compressibility'] = 1 / properties["isentropic_bulk_modulus"]
-#     properties['isothermal_bulk_modulus'] = (R/M)*T*rho*(2*delta*dalpha_dDelta + delta**2*d2alpha_dDelta2)
[…]
