### 2025-02-01 dcb4b43
remove stator case
