+++ dev_projects/thermodynamic_cycles/sCO2_split_compression/plot_cycle.py
@@ -1,258 +1,0 @@
-import os
-import sys
-import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib.colors as mcolors
-
-import meanline_axial as ml
-
-import barotropy as bpy
-
-from datetime import datetime
-
-
-# TODO
-# Add capability to export configuration file
-# Improve capability to export initial file (now it does not export solver options)
-# Plotting is messy, it would be nice to have a function to plot the processes and points only to be used independently of the heavy features for autoplotting and updatting
-# Do not plot the phase diagram when plotting points and processes
-# Latest file function to code stack, imporve the meanline analysis output structure to folders instead of single files
-
-# Add colorbar
-# Add greyscale
-# Add white patch to the diagram
-
-
-# Define configuration filename
-# CONFIG_FILE = "case_sCO2_recompression.yaml"
-
-# Find results file
-
-
-ml.set_plot_options(grid=False)
-
-
-# Initialize Brayton cycle problem
-file_name = "optimal_cycle.yaml"
-config = ml.utilities.read_configuration_file(file_name)
-thermoCycle = ml.cycles.ThermodynamicCycleProblem(config)
-thermoCycle.fitness(thermoCycle.x0)
-
-fig_x = 7
-fig_y = 4.0
-fig_x = 6
-fig_y = 4.0
-# fig_y = 12
-fig, ax = plt.subplots(figsize=(fig_x, fig_y))
-ax.set_xlabel("Entropy (kJ/kg/K)")
-ax.set_ylabel("Temperature ($^\circ$ C)")
-
-# thermoCycle.fluid.plot_phase_diagram(axes=ax, plot_saturation_line=True, plot_critical_point=True)
-
-
-# # ---------------------------------------------------------------------------- #
-# # Plot inlet states on T-s diagram
-# # ---------------------------------------------------------------------------- #
-# # Create figure
-# fig, ax = plt.subplots(figsize=(6.8, 4.8))
-# ax.set_xlabel("Entropy [kJ/kg/K]")
-# ax.set_ylabel("Temperature [$^\circ$C]")
-# range_x = np.linspace(1000, 2000, 60)
-# range_y = np.linspace(5, 55, 40) + 273.15
-# range_z = np.asarray([50, 55, 60, 65, 70, 75, 80, 85, 90, 95, 100]) * 1e5
-# ax.set_xlim(np.asarray([range_x[0], range_x[-1]]) / 1e3)
-# ax.set_ylim(np.asarray([range_y[0], range_y[-1]]) - 273.15)
-
-# Plot diagram
-
-
-# Define nice colormap
-# colormap = mcolors.LinearSegmentedColormap.from_list(
-#     "truncated_blues", plt.cm.Blues(np.linspace(0.4, 1.0, 256))
-# )
-# colormap = mcolors.LinearSegmentedColormap.from_list(
-#     "truncated_blues", plt.cm.Blues(np.linspace(0.0, 1, 256)[::-1])
-# )
-
-# Create a grayscale colormap
-colormap = mcolors.LinearSegmentedColormap.from_list(
-    "custom_gray", [(i, i, i) for i in np.linspace(0.5, 1, 256)]
-)
-
-
-# Create fluid object
-fluid_name = "CO2"
-fluid = bpy.Fluid(name=fluid_name, backend="HEOS", exceptions=True)
-
-
-Nx = 50
-Ny = 50
-rho_1 = fluid.triple_point_vapor.d / 20
-rho_2 = fluid.critical_point.d / 2
-rho_3 = fluid.set_state(ml.PT_INPUTS, fluid.p_max, fluid.T_min).d
[…]
