+++ dev_projects/thermodynamic_cycles/work_progress.md
@@ -1,141 +1,0 @@
-# Work Progress
-
-
-## Todo
-- Clean the script to plot the Ts diagram with the Z-factor contours
-- Clean the cycle optimization functionality
-  - Move plotting outside of main class
-  - Wrapper should just redefine some methods of the object? or create a new object with just the methods for optimization to prevent deepcoping problems?
-  - Clean script about IPOPT superiority with Hessian
-  - Make folder for the comparison of solvers for cycle optimization
-  - Make some notes describing each case of cycle optimization
-  - Merging of cycle optimization functionality with turbine optimization
-  - Merging of fluid properties from barotropic model and meanline axial
-  - Merging of barotropy with the meanline code?
-  - clean pysolverview  demo scripts and functinality. make sure it is consistent with the meanline code? combine int he same package to share utilities?
-  - Improve fluid property calculations:
-    - Pseudocritical Line
-    - Spinodal line
-    - Superheating
-    - Subcooling
-    - Generalized quality
-
-
-The cycle problem class only takes information about the problem formulation, not the overall config file
-specifying also the settings for the optimization problem
-As a result, the output YAML is not as meaninful as it could be
-
-Should I have the plotting within the CycleProblem class? or move it somewhere else
-Keep in mind that I wanbt to update plot iteratively when calling function plot_cycle_realtime,
-but also update the plots with the latest data during optimization when a callback function is called
-
-Now I export properly the YAML file. Check the naming and the folder structure to comly with the exporting of figures
-Export initial configuration and final configuration separately
-
-I have to create a wraper class for doing the optimization?
-
-Would it be better to have a class for optimization where I also handle the solution of the optimization problem
-In this class I could have the extra functionaliyt about plotting and exporting yaml file
-Would this apporach be more suited to save confguration file at start and ensd of the optimiazation
-Would this approach help me make parameter sweeps more, easily, or it does not really matter?
-
-**TODO options to add as configuration**
-Use plot callback or not, if yes, then make sure a function call to the plot_cycle function is done
-Move the configuration function handling to the configuration management class
-Move the plotting outside of the cycleProblem class
-Update code to handle parameter sweeps gracefully (create ranges of variabtion by permutations of the parameters to be checked, and pass a single list of dctionaries with the updated variable names. Update the keys of the configu dictionary with the keys provided)
-
-optimizer.problem.plot_cycle()  # TODO make sure that the optimization can run even if plot not initialized
-
-
-## Done
-- Ts diagram of the split compression cycle
-- Do inkscape illustration of the split compression cycle
-- Draft description of the split compression cycle (missing main modelling equations)
-- 
-
-
-## Completed Tasks
-
-### Modeling
-- Main translation from the MATLAB code to Python.
-- Implemented functions and classes to evaluate/optimize recuperated Brayton cycle.
-- Implemented polytropic process for expansion compression and isentropic calculations for faster computation speed.
-- Discretization and temperature difference calculation for heat exchangers.
-- Reformulated the problem using components instead of states. This makes the code more flexible and intuitive. There is no need to keep track of many state points numbered from 1 to a big number.
-- Implemented lazy initialization to avoid recomputing the saturation line.
-- Generalized the vapor quality calculation for values outside of the two-phase region.
-- Added calculation of superheating and subcooling that works inside and outside the two-phase region
-- Cleaned up the recompression and recuperated cycle functions and moved them to their own files.
-
-
-### Optimization
-- Formulated the objective function and constraints automatically via configuration.
-- Introduced very flexible constraint definition by accessing variables in nested dictionaries.
-- Added functionality to ensured the objective function are numeric and scalar.
[…]

[…]
