### 2025-02-01 2b5f74a
remove dev_projects from main
