+++ dev_projects/centrifugal_compressor/compressor_geom/nomenclature.m
@@ -1,48 +1,0 @@
-%% Nomenclature
-
-%% Inducer
-% has_inducer = Whether or not the compressor has and inducer
-% inducer_type = The type of inducer geometry
-% L_inducer = Axial length of the inducer
-% D_0_s = Inducer inlet radius at the shroud
-% D_0_h = Inducer inlet radius at the hub
-
-
-%% Impeller
-% D_1_s = Impeller inlet radius at the shroud
-% D_1_h = Impeller inlet radius at the hub
-% D_2 = Impeller outlet radius
-% b_2 = Impeller outlet width
-% L_z = Impeller axial length
-% Z_r_full = Number of full blades
-% Z_r_split = Number of splitter blades
-% LR_split = Ratio of splitter blades length over full blades length
-% beta_1b_s = Impeller inlet blade angle at the shroud
-% beta_1b_h = Impeller inlet blade angle at the hub
-% beta_1b = Impeller inlet blade angle at RMS radius
-% beta_2b = Blade angle at impeller outlet, radial 
-% t_b_1 = Impeller blade thickness (leading edge)
-% t_b_2 = Impeller blade thickness (trailing edge)
-% eps_a = Impeller axial clearance (inlet) 
-% eps_r = Impeller radial clearance (exit) 
-% eps_b = Impeller backplate clearance (exit) 
-% Ra = Impeller blade surface roughness
-
-
-%% Diffuser
-% has_vanes = Whether or not the diffuser has vanes
-% vane_type = Type of diffuser vanes (airfoil or wedge)
-% Z_vd = Number of diffuser vanes
-% b_3 = Channel width at diffuser leading edge
-% b_4 = Channel width at diffuser throat
-% b_5 = Channel width at diffuser outlet
-% D_3 = Diameter at diffuser leading edge
-% D_4 = Diameter at diffuser throat
-% D_5 = Diameter at diffuser outlet
-% alpha_3b = Metal angle at diffuser leading edge
-% alpha_5b = Metal angle at diffuser trailing edge
-% wedge_angle = Wedge angle of the diffuser
-% t_b_3 = Diffuser leading edge thickness
-% t_b_5 = Diffuser trailing edge thickness
-
-
