+++ dev_projects/optimization_algos/two_stage_analysis.py
@@ -1,168 +1,0 @@
-
-import turboflow as tf
-import numpy as np
-import matplotlib.pyplot as plt
-from mpl_toolkits.axes_grid1.inset_locator import inset_axes, mark_inset
-import matplotlib.cm as cm
-import seaborn as sns
-
-print_evals = False
-save_fig = True
-axins = True
-
-def adjust_objective_values(objective_values, constraint_values, efficiency_values, cut_off = 1.0):
-
-    # Adjust array such that the objective value is strictly deacreasing
-    objective_values = objective_values.copy()
-    constraint_values = constraint_values.copy()
-    efficiency_values = efficiency_values.copy()
-    for i in range(1, len(objective_values)):
-        if objective_values[i] > objective_values[i-1] or np.isnan(objective_values[i]) :
-            objective_values[i] = objective_values[i-1]
-            constraint_values[i] = constraint_values[i-1]
-            efficiency_values[i] = efficiency_values[i-1]
-
-    # Cut off objective value
-    if isinstance(cut_off, float):
-        last_val = objective_values[-1]
-        for i in range(1,len(objective_values)):
-            if objective_values[i] <= cut_off*last_val:
-                objective_values = objective_values[:i]
-                constraint_values = constraint_values[:i]
-                efficiency_values = efficiency_values[:i]
-                break
-
-    return objective_values, constraint_values, efficiency_values
-
-# Load object
-filename_slsqp = "output/pickle_2stage_slsqp_2024-09-03_22-57-21.pkl"
-# filename_slsqp_BB = "output/pickle_2stage_BB_slsqp_2024-09-04_12-17-17.pkl"
-# filename_sga = "output/pickle__2stage_sga_2024-08-31_01-38-24.pkl"
-filename_slsqp_BB = "output/pickle_2stage_BB_slsqp_new_2024-11-07_12-34-56.pkl"
-filename_sga = "output/pickle_2stage_BB_sga_new_2024-11-09_09-50-03.pkl"
-filename_ig = "output/initial_guess_2stage_2024-09-20_16-04-06.pkl"
-
-solver_slsqp = tf.load_from_pickle(filename_slsqp)
-solver_slsqp_BB = tf.load_from_pickle(filename_slsqp_BB)
-solver_sga = tf.load_from_pickle(filename_sga)
-initial_guess = tf.load_from_pickle(filename_ig)
-
-# Close open figures
-plt.close()
-
-# Plot options
-# colors = sns.color_palette("colorblind", n_colors=3)  # Adjust n_colors as needed
-colors = cm.magma(np.linspace(0.85, 0.25, 3)) 
-linestyles = ['-', '--', '-.']
-linewidth = 2.5
-fontsize = 22
-
-obj_sga, cons_sga, eff_sga = adjust_objective_values(solver_sga.problem.optimization_process.iterations_objective_function[1:], solver_sga.problem.optimization_process.constraint_violation[1:], solver_sga.problem.optimization_process.iterations_efficiency[1:], cut_off=None)
-
-convergence_data = {"EO, SLSQP" : {"objective" : solver_slsqp.convergence_history["objective_value"], 
-                        "constraint_violation" : solver_slsqp.convergence_history["constraint_violation"], 
-                        "iterations" : solver_slsqp.convergence_history["grad_count"]},
-        "BB, SLSQP" : {"objective" : solver_slsqp_BB.convergence_history["objective_value"],
-                "constraint_violation" : solver_slsqp_BB.convergence_history["constraint_violation"],
[…]
