### 2025-02-01 d5a5d0e
remove cycle calculations
