+++ turboflow/cycles/components.py
@@ -1,313 +1,0 @@
-import numpy as np
-
-from scipy.integrate import solve_ivp
-
-# from .. import utilities
-from .. import fluid_properties as props
-
-
-def heat_exchanger(
-    fluid_hot,
-    h_in_hot,
-    h_out_hot,
-    p_in_hot,
-    p_out_hot,
-    fluid_cold,
-    h_in_cold,
-    h_out_cold,
-    p_in_cold,
-    p_out_cold,
-    num_steps=50,
-    counter_current=True,
-):
-    # Evaluate properties on the hot side
-    hot_side = heat_transfer_process(
-        fluid=fluid_hot,
-        h_1=h_in_hot,
-        h_2=h_out_hot,
-        p_1=p_in_hot,
-        p_2=p_out_hot,
-        num_steps=num_steps,
-    )
-
-    # Evaluate properties in the cold side
-    cold_side = heat_transfer_process(
-        fluid=fluid_cold,
-        h_1=h_in_cold,
-        h_2=h_out_cold,
-        p_1=p_in_cold,
-        p_2=p_out_cold,
-        num_steps=num_steps,
-    )
-
-    # Sort values for temperature difference calculation
-    if counter_current:
-        for key, value in hot_side["states"].items():
-            hot_side["states"][key] = np.flip(value)
-
-    # Compute temperature difference
-    dT = hot_side["states"]["T"] - cold_side["states"]["T"]
-
-    # Compute mass flow ratio
-    # Use 1.0 for cases with no heat exchange (i.e., no recuperator)
-    dh_hot = hot_side["state_in"].h - hot_side["state_out"].h
-    dh_cold = cold_side["state_out"].h - cold_side["state_in"].h
-    mass_ratio = 1.00 if (dh_cold == 0 or dh_hot == 0) else dh_cold / dh_hot
-
-    # Create result dictionary
-    result = {
-        "type": "heat_exchanger",
-        "hot_side": hot_side,
-        "cold_side": cold_side,
-        "q_hot_side": dh_hot,
-        "q_cold_side": dh_cold,
-        "temperature_difference": dT,
-        "mass_flow_ratio": mass_ratio,
-    }
-
-    return result
-
-
-def heat_transfer_process(fluid, h_1, p_1, h_2, p_2, num_steps=25):
-    # Generate linearly spaced arrays for pressure and enthalpy
-    p_array = np.linspace(p_1, p_2, num_steps)
-    h_array = np.linspace(h_1, h_2, num_steps)
-
-    # Initialize lists to store states for hot and cold sides
-    states = []
-
-    # Calculate states for hot side
-    for p, h in zip(p_array, h_array):
-        states.append(fluid.set_state(props.HmassP_INPUTS, h, p))
-
-    # Store inlet and outlet states
-    state_in = states[0]
-    state_out = states[-1]
-
-    # Create result dictionary
-    result = {
-        "states": props.states_to_dict(states),
-        "fluid_name": fluid.name,
-        "state_in": state_in,
-        "state_out": state_out,
-        "mass_flow": np.nan,
-        "color": "black",
-    }
-
-    return result
-
-
-def compression_process(
-    fluid, h_in, p_in, p_out, efficiency, efficiency_type="isentropic", num_steps=10
-):
-    """
-    Calculate properties along a compression process defined by a isentropic or polytropic efficiency
-
-    Parameters
-    ----------
-    fluid : Fluid
-        The fluid object used to evaluate thermodynamic properties
-    h_in : float
-        Enthalpy at the start of the compression process.
-    p_in : float
-        Pressure at the start of the compression process.
-    p_out : float
-        Pressure at the end of the compression process.
-    efficiency : float
[…]
