+++ turboflow/cycles/cycle_optimization_old.py
@@ -1,837 +1,0 @@
-# import os
-# import time
-# import yaml
-# import copy
-# import datetime
-# import threading
-# import numpy as np
-# import pandas as pd
-# import matplotlib.pyplot as plt
-
-# from .. import pysolver_view as psv
-# from .. import utilities
-# from .. import fluid_properties as props
-
-# from . import brayton_recuperated
-# from . import brayton_split_compression
-
-# COLORS_MATLAB = utilities.COLORS_MATLAB
-# LABEL_MAPPING = {
-#     "s": "Entropy [J/kg/K]",
-#     "T": "Temperature [K]",
-#     "p": "Pressure [Pa]",
-#     "h": "Enthalpy [J/kg]",
-#     "d": "Density [kg/m$^3$]",
-#     "a": "Speed of sound [m/s]",
-#     "Z": "Compressibility factor [-]",
-#     "heat": "Heat flow rate [W]",
-# }
-
-# # Cycle configurations available
-# CYCLE_TOPOLOGIES = {
-#     "recuperated": brayton_recuperated.evaluate_cycle,
-#     "split_compression": brayton_split_compression.evaluate_cycle,
-#     "recompression": brayton_split_compression.evaluate_cycle,
-# }
-
-# GRAPHICS_PLACEHOLDER = {
-#     "process_lines": {},
-#     "state_points": {},
-#     "pinch_point_lines": {},
-# }
-
-
-# class ThermodynamicCycleProblem(psv.OptimizationProblem):
-#     """
-#     A class to represent a thermodynamic cycle optimization problem.
-
-#     This class provides functionalities to load and update the configuration for
-#     the thermodynamic cycle, and to perform interactive plotting based on the current
-#     configuration.
-
-#     Attributes
-#     ----------
-#     plot_initialized : bool
-#         Flag to indicate if the plot has been initialized.
-#     constraints : dict
-#         Dictionary holding the constraints of the problem.
-#     fixed_parameters : dict
-#         Dictionary holding the fixed parameters of the problem.
-#     design_variables : dict
-#         Dictionary holding the current values of the design variables.
-#     lower_bounds : dict
-#         Dictionary holding the lower bounds of the design variables.
-#     upper_bounds : dict
-#         Dictionary holding the upper bounds of the design variables.
-#     keys : list
-#         List of keys (names) of the design variables.
-#     x0 : np.ndarray
-#         Initial guess for the optimization problem.
-
-#     Methods
-#     -------
-#     update_config(configuration):
-#         Update the problem's configuration based on the provided dictionary.
-#     load_config_from_file(configuration_file):
-#         Load and update the problem's configuration from a specified file.
-#     plot_cycle_interactive(configuration_file, update_interval=0.20):
-#         Perform interactive plotting, updating the plot based on the configuration file.
-#     """
-
-#     def __init__(self, configuration, out_dir=None):
-#         """
-#         Constructs all the necessary attributes for the ThermodynamicCycleProblem object.
-
-#         Parameters
-#         ----------
-#         config : dict
-#             Dictionary containing the configuration for the thermodynamic cycle problem.
-#         """
-#         # Initialize variables
-#         self.figure = None
-#         self.figure_TQ = None
-#         self.configuration = copy.deepcopy(configuration)
-#         self.update_configuration(self.configuration)
-#         self.graphics = copy.deepcopy(GRAPHICS_PLACEHOLDER)
-
-#         # Create fluid object to calculate states used for the scaling of design variables
-#         self.fluid = props.Fluid(**self.fixed_parameters["working_fluid"])
-#         self._calculate_special_points()
-
-#         # Define filename with unique date-time identifier
-#         if out_dir == None:
[…]

[…]
