+++ examples/R125/R125_config.yaml
@@ -1,0 +1,163 @@
+turbomachinery: axial_turbine
+operation_points: 
+  fluid_name: R125       # Fluid
+  T0_in: 428.15            # Inlet total pressure (Inlet total temperature)
+  p0_in: 36.18e5           # Inlet total temperature (Inlet total pressure)
+  p_out : 15.69e5          # Exit static pressure
+  omega: 3810              # Flow angle at first stator (Rotational speed)
+  alpha_in: 0              # Rotational speed (Flow angle at first stator)
+
+
+simulation_options: 
+  deviation_model : aungier 
+  choking_criterion : critical_mach_number
+  rel_step_fd: 1e-4 
+  loss_model:  
+    model: benner  
+    loss_coefficient: stagnation_pressure  
+    inlet_displacement_thickness_height_ratio: 0.011  
+  
+performance_analysis :
+  performance_map: 
+    fluid_name: R125 
+    T0_in: 428.15
+    p0_in: 36.18e5 
+    p_out: 36.18e5/np.linspace(1.6, 4.5, 40)
+    omega: 3810 
+    alpha_in: 0 
+  solver_options:
+    method: hybr  
+    tolerance: 1e-8  
+    max_iterations: 100 
+    derivative_method: 2-point  
+    derivative_abs_step: 1e-6 
+    print_convergence: True
+    plot_convergence: False
+  initial_guess :
+      efficiency_tt : [0.9, 0.8] 
+      efficiency_ke : [0.2, 0.1]
+      ma_1 : [0.8, 0.8]
+      ma_2 : [0.8, 0.8]
+
+geometry:
+  cascade_type: ["stator", "rotor"]
+  radius_hub_in: [0.029655809872, 0.029655809872]
+  radius_hub_out: [0.029655786328, 0.028352202799]
+  radius_tip_in: [0.049426352799, 0.049426376343]
+  radius_tip_out: [0.049426376343, 0.050729959872]
+  pitch: [0.005651993855, 0.006628034553]
+  chord: [0.010647308861, 0.010623976311]
+  stagger_angle: [+40.00, -45.903674771656]
+  opening: [0.000981458433, 0.001512676769]
+  leading_edge_angle : [0.00, -15.00]
+  leading_edge_wedge_angle : [45.00, 45.00]
+  leading_edge_diameter : [0.01200896, 0.01200896]
+  trailing_edge_thickness : [4.9072922e-05, 7.5633838e-05]
+  maximum_thickness : [0.002129461772, 0.001883200453]
+  tip_clearance: [0.00, 0.0005]
+  throat_location_fraction: [1, 1]
+  
+
+design_optimization:
+  objective_function: 
+    variable : overall.efficiency_ts
+    scale : 1
+    type : maximize
+  variables :
+    specific_speed : # Limited 
+      value : 0.713756
+      lower_bound : 0.5
+      upper_bound : 1.5
+    blade_jet_ratio : # Limited
+      value : 0.466501
+      lower_bound : 0.3
+      upper_bound : 0.8
+    hub_tip_ratio_in :
+      value : [0.715999, 0.715999]
+      lower_bound : [0.6, 0.6]
+      upper_bound : [0.9, 0.9]
+    hub_tip_ratio_out :
+      value : [0.715999, 0.674840]
+      lower_bound : [0.6, 0.6]
+      upper_bound : [0.9, 0.9]
+    aspect_ratio :
+      value : [1.285550, 1.402149]
+      lower_bound : [1.0, 1.0]
+      upper_bound : [2.0, 2.0]
+    pitch_chord_ratio :
+      value : [0.699312, 0.584804]
+      lower_bound : [0.5, 0.5]
+      upper_bound : [1.10, 1.10]
+    trailing_edge_thickness_opening_ratio : 
[…]
