+++ turboflow/__init__.py
@@ -19,7 +19,6 @@ from .config_validation import *
 # Import subpackages
 from .pysolver_view import *
 from .properties import *
-from .cycles import *
 from .utilities import *
 
 from .axial_turbine import compute_performance, compute_optimal_turbine
