### 2025-02-02 06a64fc
add logger funtion with examples
