+++ dev_projects/automatic_differentiation/plot_convergence_history.py
@@ -7,6 +7,8 @@ import matplotlib.cm as cm
 import seaborn as sns
 import os
 
+tf.set_plot_options()
+
 algorthims = False
 pop_size = True
 
@@ -19,32 +21,38 @@ os.makedirs(figures_dir, exist_ok=True)
 output_dir = "output"
 
 # Load object
-filename_jax_slsqp = os.path.join(output_dir, '1stage_jax_slsqp.pkl')
-filename_FD_3_slsqp = os.path.join(output_dir, '1stage_FD_1e-3_slsqp.pkl')
-filename_FD_6_slsqp = os.path.join(output_dir, '1stage_FD_1e-6_slsqp.pkl')
-filename_FD_9_slsqp = os.path.join(output_dir, '1stage_FD_1e-9_slsqp.pkl')
+filename_jax_slsqp = os.path.join(output_dir, './case_46/two_stage_config_2024-12-01_14-31-43.pkl')
+# filename_FD_3_slsqp = os.path.join(output_dir, './case_17/one_stage_config_latest.pkl')
+# filename_FD_6_slsqp = os.path.join(output_dir, './case_18/one_stage_config_latest.pkl')
+# filename_FD_9_slsqp = os.path.join(output_dir, './case_19/one_stage_config_latest.pkl')
+# filename_FD_12_slsqp = os.path.join(output_dir, './case_20/one_stage_config_latest.pkl')
 
-filename_FD_3_ipopt = os.path.join(output_dir, '1stage_FD_1e-3_ipopt.pkl')
-filename_FD_6_ipopt = os.path.join(output_dir, '1stage_FD_1e-6_ipopt.pkl')
-filename_FD_9_ipopt = os.path.join(output_dir, '1stage_FD_1e-9_ipopt.pkl')
+filename_jax_ipopt = os.path.join(output_dir, './case_51/two_stage_config_2024-12-01_17-52-40.pkl')
+# filename_FD_3_ipopt = os.path.join(output_dir, '1stage_FD_1e-3_ipopt.pkl')
+# filename_FD_6_ipopt = os.path.join(output_dir, '1stage_FD_1e-6_ipopt.pkl')
+# filename_FD_9_ipopt = os.path.join(output_dir, '1stage_FD_1e-9_ipopt.pkl')
 
-filename_FD_3_snopt = os.path.join(output_dir, '1stage_FD_1e-3_snopt.pkl')
-filename_FD_6_snopt = os.path.join(output_dir, '1stage_FD_1e-6_snopt.pkl')
-filename_FD_9_snopt = os.path.join(output_dir, '1stage_FD_1e-9_snopt.pkl')
+filename_jax_snopt = os.path.join(output_dir, './case_56/two_stage_config_2024-12-01_18-39-40.pkl')
+# filename_FD_3_snopt = os.path.join(output_dir, '1stage_FD_1e-3_snopt.pkl')
+# filename_FD_6_snopt = os.path.join(output_dir, '1stage_FD_1e-6_snopt.pkl')
+# filename_FD_9_snopt = os.path.join(output_dir, '1stage_FD_1e-9_snopt.pkl')
 
 
 solver_jax_slsqp = tf.load_from_dill(filename_jax_slsqp)
-solver_FD_3_slsqp = tf.load_from_dill(filename_FD_3_slsqp)
-solver_FD_6_slsqp = tf.load_from_dill(filename_FD_6_slsqp) 
-solver_FD_9_slsqp = tf.load_from_dill(filename_FD_9_slsqp)
+# solver_FD_3_slsqp = tf.load_from_dill(filename_FD_3_slsqp)
[…]
