+++ dev_projects/automatic_differentiation/validation/plot_validation_map.py
@@ -1,0 +1,100 @@
+import pandas as pd
+import matplotlib.pyplot as plt
+import numpy as np
+import turboflow as tf
+
+
+tf.set_plot_options()
+
+df = pd.read_excel("./output_map/performance_latest.xlsx", sheet_name= "overall")
+df_exp = pd.read_excel("./validation_cases_summary.xlsx")
+
+markerlist = ["o", "+", "*", "s"]
+# fig, ax = plt.subplots()
+# for omega in df["angular_speed"][df["angular_speed"] > 0.6*1627]:
+#     ax.plot(df["PR_ts"][df["angular_speed"]==omega], df["efficiency_ts"][df["angular_speed"]==omega])
+#     ax.plot(df_exp["pressure_ratio_ts"][np.isclose(df_exp["omega"], omega)], df_exp["efficiency_ts"][np.isclose(df_exp["omega"], omega)], linestyle= "None", marker="o")
+#     # ax.set_ylim([2.55, 2.85])
+
+# fig, ax = plt.subplots()
+# for omega in df["angular_speed"][df["angular_speed"] > 0.6*1627]:
+#     ax.plot(df["PR_ts"][df["angular_speed"]==omega], df["torque"][df["angular_speed"]==omega])
+#     ax.plot(df_exp["pressure_ratio_ts"][np.isclose(df_exp["omega"], omega)], df_exp["torque"][np.isclose(df_exp["omega"], omega)], linestyle= "None", marker="o")
+#     # ax.set_ylim([2.55, 2.85])
+
+# fig, ax = plt.subplots()
+# for omega in df["angular_speed"][df["angular_speed"] > 0.6*1627]:
+#     ax.plot(df["PR_ts"][df["angular_speed"]==omega], df["mass_flow_rate"][df["angular_speed"]==omega])
+#     ax.plot(df_exp["pressure_ratio_ts"][np.isclose(df_exp["omega"], omega)], df_exp["mass_flow_rate"][np.isclose(df_exp["omega"], omega)], linestyle= "None", marker="o")
+#     # ax.set_ylim([2.55, 2.85])
+
+# Create a figure and axis
+fig, ax = plt.subplots()
+
+# Define a color palette with shades of red, orange, and brown
+# You can manually define these colors or generate them using matplotlib's color options
+# colors = ['#D7301F', '#FC8D59', '#FEE08B', '#D8B365']  # Red, Orange, Yellow, Dark Brown
+colors = plt.get_cmap('plasma_r')(np.linspace(0.2, 1.0, 4))
+num_omegas = len(colors)
+
+# Lists to store the handles and labels for the legend
+handles_sim = []  # Handles for simulation lines
+labels_sim = []   # Labels for simulation lines
+handles_exp = []  # Handles for experimental points
+labels_exp = []   # Labels for experimental points
+
+# Assuming you already have `df` and `df_exp` for the data
+unique_omegas = df["angular_speed"][df["angular_speed"] > 0.6 * 1627].unique()
+
+# Loop through omega values
+for idx, omega in enumerate(unique_omegas):
+    
+    # Filter data for the current omega
+    sim_x = df["PR_ts"][df["angular_speed"] == omega]
+    sim_y = df["efficiency_ts"][df["angular_speed"] == omega]
+    
+    # Plot the simulation line with the chosen color
+    line, = ax.plot(sim_x, sim_y, color=colors[idx], label=f'{int(np.round((omega/1627)*100))}')
+    
+    # Add the line handle and label to the simulation legend list
+    handles_sim.append(line)
+    labels_sim.append(f'{int(np.round((omega/1627)*100))}')
+    
+    # Experimental data for the current omega
+    exp_x = df_exp["pressure_ratio_ts"][np.isclose(df_exp["omega"], omega)]
[…]
