+++ validation_cases/axial_turbine/Kofskey1972_one_stage/quantify_error.py
@@ -2,12 +2,15 @@ import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
 
-filename_exp = "./experimental_data_kofskey1972_1stage_raw.xlsx"
-filename_sim = "output/performance_analysis_2024-03-14_01-52-41.xlsx"
+# Set filenames
+filename_exp = "experimental_data/experimental_data_kofskey1972_1stage_raw.xlsx"
+filename_sim = "output/experimental_points.xlsx"
 
+# Set design point angular speed
+omega_des = 1627
+
+# Define percentage limits
 limits = [2.5, 5, 10]
-sheets = ["Mass flow rate", "Torque", "alpha_out"]
-
 
 def count(values, lim):
     below_lim_count = np.count_nonzero(abs(values) < lim)
@@ -19,24 +22,25 @@ def count(values, lim):
     percentage_below_lim = (below_lim_count / total_count) * 100
     return percentage_below_lim
 
-
+# Load simulation data
 speed_percent = np.flip([110, 100, 90, 70])
 data_sim = pd.read_excel(filename_sim, sheet_name=["overall"])
 data_sim = data_sim["overall"]
-
-data_exp = pd.read_excel(filename_exp, sheet_name=sheets)
-for sheet in sheets:
-    data_exp[sheet] = data_exp[sheet][data_exp[sheet]["omega"].isin(speed_percent)]
-
 data_mass_flow = data_sim[0:37]
 data_torque = data_sim[37:85]
 data_eta = data_sim[85:170]
 data_alpha = data_sim[170:]
 
+# Load experimental data
+sheets = ["Mass flow rate", "Torque", "alpha_out"]
+data_exp = pd.read_excel(filename_exp, sheet_name=sheets)
+for sheet in sheets:
+    data_exp[sheet] = data_exp[sheet][data_exp[sheet]["omega"].isin(speed_percent)]
+
+# Initalize dictionary to store results
 mass_flow_rate_error = {}
 torque_error = {}
 alpha_error = {}
-
 mass_flow_rate_count = {}
 torque_count = {}
 alpha_count = {}
@@ -48,8 +52,8 @@ for speed in speed_percent:
         data_exp["Mass flow rate"]["omega"] == speed
     ]["m"].values
     mass_flow_rate_sim = data_mass_flow[
-        (data_mass_flow["speed_percent"] > speed - 1)
-        & (data_mass_flow["speed_percent"] < speed + 1)
+        (data_mass_flow["angular_speed"] > speed/100*omega_des - 1)
+        & (data_mass_flow["angular_speed"] < speed/100*omega_des + 1)
     ]["mass_flow_rate"].values
     mass_flow_rate_error[speed] = (
         (mass_flow_rate_exp - mass_flow_rate_sim) / mass_flow_rate_exp * 100
@@ -65,8 +69,8 @@ for speed in speed_percent:
         "Torque"
     ].values
     torque_sim = data_torque[
-        (data_torque["speed_percent"] > speed - 1)
-        & (data_torque["speed_percent"] < speed + 1)
+        (data_torque["angular_speed"] > speed/100*omega_des - 1)
+        & (data_torque["angular_speed"] < speed/100*omega_des + 1)
     ]["torque"].values
     torque_error[speed] = (torque_exp - torque_sim) / torque_exp * 100
     torque_count[speed] = [
@@ -85,8 +89,8 @@ for speed in speed_percent:
     )
     alpha_sim = np.cos(
         data_alpha[
-            (data_alpha["speed_percent"] > speed - 1)
-            & (data_alpha["speed_percent"] < speed + 1)
+            (data_alpha["angular_speed"] > speed/100*omega_des - 1)
[…]
