### 2025-02-02 5a3cc7e
limit scope to axial turbines
