+++ _sources/source/tutorials.rst.txt
@@ -6,7 +6,7 @@ Tutorials
 ===============
 
 Welcome to the TurboFlow Tutorial Guide. This guide offers step-by-step instructions for using TurboFlow in mean-line 
-modeling of turbomachinery. The tutorials focus on the two main aspects of the software: performance prediction and design optimization, 
+modeling of axial turbines. The tutorials focus on the two main aspects of the software: performance prediction and design optimization, 
 each illustrated with detailed examples. Additionally, an overview of the various configuration file options is included 
 to help you fully utilize TurboFlow's capabilities in your projects.
 
