### 2025-02-03 ffd4bfe
update docs and examples
