### 2025-02-03 65058ac
improve pickling function
