### 2025-02-03 2e6eb97
add affiliated papers and author contributions
