+++ _sources/source/publications.rst.txt
@@ -1,0 +1,25 @@
+.. _publications:
+
+Affiliated Papers and Author Contributions
+==========================================
+
+This section lists the scientific papers affiliated with the code and provides an overview of the contributions made by the developers and authors.
+
+Affiliated Papers
+-----------------
+
+The following paper is affiliated with the code, having been used for validation of the model:
+
+- **Equation-Oriented Meanline Method for Axial Compressors and Turbines**  
+  *L. B. Anderson, R. Agromayor, F. Haglind, L. O. Nord*  
+  [DOI: 10.1115/1.4057239](https://asmedigitalcollection.asme.org/turbomachinery/article/147/4/041002/1206960/Equation-Oriented-Meanline-Method-for-Axial)
+
+Author Contributions
+---------------------
+
+The development of the code has been a collaborative effort involving multiple contributors:
+
+- **Lasse B. Anderson**: Main contributor to code and concept.
+- **Roberto Agromayor**: Contributor to code and concept.
+- **Fredrik Haglind**: Supervision and revision.
+- **Lars O. Nord**: Supervision and revision.
