+++ validation_cases/axial_turbine/Kofskey1972_two_stage/Kofskey1972_2stage.yaml
@@ -9,7 +9,7 @@ operation_points:
 
 
 simulation_options:
-  choking_criterion : critical_mass_flow_rate 
+  choking_criterion : critical_mach_number 
   deviation_model : aungier
   blockage_model : 0.00  
   rel_step_fd: 1e-3
@@ -29,9 +29,9 @@ performance_analysis :
     fluid_name: air
     T0_in: 298.9
     p0_in: 12.4e4
-    p_out: 12.4e4/np.linspace(2.0, 6.0, 5)
-    # omega: 1627*np.array([0.7, 0.9, 1.0, 1.1])
-    omega: 1627*1.1
+    p_out: 12.4e4/np.linspace(2.0, 6.0, 60)
+    omega: 1627*np.array([0.7, 0.9, 1.0, 1.1])
+    # omega: 1627*np.array([1.1])
     alpha_in: 0 
   solver_options:
     method: hybr
