+++ paper/paper.bib
@@ -9,6 +9,18 @@
 	date = {2014},
 }
 
+@article{anderson_equation-oriented_2024,
+	title = {Equation-oriented meanline method for axial turbine performance prediction under choking conditions},
+	volume = {147},
+	issn = {0889-504X},
+	url = {https://doi.org/10.1115/1.4066741},
+	doi = {10.1115/1.4066741},
+	number = {41002},
+	journaltitle = {Journal of Turbomachinery},
+	author = {Anderson, Lasse B. and Agromayor, Roberto and Parisi, Simone and Haglind, Fredrik and Nord, Lars O.},
+	date = {2024-10-25},
+}
+
 @book{macchi_organic_2017,
 	title = {Organic rankine cycle ({ORC}) power systems : technologies and applications},
 	isbn = {0-08-100511-3},
