+++ paper/paper.md
@@ -36,11 +36,10 @@ Meanline modeling is a fundamental approach used in the design and analysis of a
 
 TurboFlow is a Python package for meanline modeling of axial turbines, providing a comprehensive framework for on- and off-design performance analysis and design optimization. It employs an equation-oriented model formulation, making it compatible with gradient-based equation solvers and optimization algorithms for efficient computations. The package features a modular architecture that allows for seamless integration of various submodels, enabling users to 
 select and combine different models for calculating losses, flow angles and choking, tailoring the analysis to specific needs. The structure also facilitates the implementation of other submodels for these purposes. 
-TurboFlow provides access to advanced equations of state for real gas fluid properties by interfacing to the CoolProp library. The accuracy and computational robustness of the implemented models have been demonstrated through comprehensive validation against experimental data.  
+TurboFlow provides access to advanced equations of state for real gas fluid properties by interfacing to the CoolProp library. The accuracy and computational robustness of the implemented models have been demonstrated through comprehensive validation against experimental data [@anderson_equation-oriented_2024].  
 
-TurboFlow comes with comprehensive documentation, including installation guides, tutorials, model descriptions and a complete API reference. 
-This extensive resource ensures that users can easily learn how to use the package and apply it effectively in their projects. For more details, visit the [documentation pages](https://turbo-sim.github.io/turboflow/).
-Additionally, the package includes preconfigured examples that demonstrate performance analysis and design optimization. These examples serve as practical guides for users to apply TurboFlow to their own projects. 
+TurboFlow comes with comprehensive documentation, including installation guides, tutorials, model descriptions and a complete API reference. This extensive resource ensures that users can easily learn how to use the package and apply it effectively in their projects. For more details, visit the [documentation pages](https://turbo-sim.github.io/turboflow/).
+Additionally, the package includes preconfigured examples that demonstrate performance analysis and design optimization. These examples serve as practical guides for users to apply TurboFlow to their own projects. Additionally, these examples showcase the post-processing capabilities, including plotting, logging, and export utilities for result interpretation and analysis.
 
 The package source code is hosted in a [Github repository](https://github.com/turbo-sim/TurboFlow) [@turboflow]. Through Github Actions, an automated test suite is included, which checks the functionality of the performance analysis and design optimization, as well as all submodels. It enables continuous integration, ensuring that code changes are systematically tested and validated. This comprehensive testing framework provides confidence that the code works as expected, maintaining the reliability of the package with each update.
 
