+++ tests/regression_data_windows/performance_analysis_ainley_mathieson.yaml
@@ -7,12 +7,12 @@ operation_points:
   omega: 1627.0
   alpha_in: 0.0
 simulation_options:
-  deviation_model: aungier
+  deviation_model: ainley_mathieson
   blockage_model: 0.0
   choking_criterion: critical_mass_flow_rate
   rel_step_fd: 0.0001
   loss_model:
-    model: kacker_okapuu
+    model: benner
     loss_coefficient: stagnation_pressure
     custom_value: 0.0
     inlet_displacement_thickness_height_ratio: 0.011
@@ -273,3 +273,4 @@ design_optimization:
     update_on: gradient
     options: {}
   radius_type: constant_mean
+  multistarts: 0
