+++ tests/regression_data_windows/performance_analysis_critical_mach_number.yaml
@@ -9,10 +9,10 @@ operation_points:
 simulation_options:
   deviation_model: aungier
   blockage_model: 0.0
-  choking_criterion: critical_mass_flow_rate
+  choking_criterion: critical_mach_number
   rel_step_fd: 0.0001
   loss_model:
-    model: moustapha
+    model: benner
     loss_coefficient: stagnation_pressure
     custom_value: 0.0
     inlet_displacement_thickness_height_ratio: 0.011
@@ -273,3 +273,4 @@ design_optimization:
     update_on: gradient
     options: {}
   radius_type: constant_mean
+  multistarts: 0
