+++ .github/workflows/generate_regression_data_ubuntu.yaml
@@ -40,7 +40,7 @@ jobs:
         python tests/generate_regression_data.py  # Adjust this to your specific script
 
     - name: Upload regression data
-      uses: actions/upload-artifact@v3
+      uses: actions/upload-artifact@v4
       with:
         name: regression-data
         path: tests/regression_data_linux/**
