+++ tests/regression_data_linux/performance_analysis_critical_mass_flow_rate_linux.yaml
@@ -273,3 +273,4 @@ design_optimization:
     update_on: gradient
     options: {}
   radius_type: constant_mean
+  multistarts: 0
