### 2025-02-06 #7588 @lasseband
Hi @aromei,

Thank you for taking the time to carefully review our code and for your thoughtful comments. We have done our best to address each of them, and you can find our replies below.

### General checks
- ***License: I am not entirely sure, but shouldn’t the copyright belong to all authors rather than being assigned only to the main author?***
	- All authors added in the lisence file.
### Documentation
- ***Community guidelines: I am uncertain about this point in the reviewer’s checklist, as there are no clear guidelines provided. However, since the software is hosted on GitHub, I believe the community can use this platform to contribute to the code directly.***
	- In the develop guidelines section in the documentation you can find the following information:
		- Installation instruction for developers.
		- Pull request guidelines.
		- How to add tests.
		- How to report an issue.
		- How to publish new versions to PyPI.
- ***Installation instructions: about dependencies, I needed to install pytest to run the test.***
	- pytest is installed through poetry in the developer installation guide.
	- Developer installation guide is updated to ensure that development dependencies are included.
- ***Functionality documentation: It is unclear which types of turbomachinery the software can handle. Based on the documentation, it appears to be limited to axial turbines. But, the code also contains folders related to centrifugal compressors, which are not documented.***
	- Within this submission, axial turbines are the only turbomachinery considered. The paper/documentation and README file is updated to clearly state this.
	- The authors have started incorporating new turbomachinery, e.g. centrifugal compressors. However, this will not be a part of this submission, but hopefully fully available in later versions.
- ***Contribution and authorship: Two of the four listed authors have not made any commits to the project, while there are two contributors who do not appear on the author list. I just wanted to double-check that the author list accurately reflects the contributions made to the software.***
	- Two of the authors are the supervisors for the main contributors to the code (Lasse B. Anderson and Roberto Agromayor). Srinivas Diwanji is contributing to the code in another project and is only working on a branch of the code. Daniel S. Katz is from JOSS.
	- Author contributions added in the documentation
### Software paper
- ***Summary: The authors can spend a few sentences to clearly state what is a meanline code to help a non-specialist audience.***
	- Sentences added.
- ***Summary: The authors state that the software includes meanline modeling for turbomachinery. Since "turbomachinery" is a broad term, it raises the question of whether axial compressors and radial turbines are also supported. Overall, the summary lacks specificity and does not accurately reflect the software's capabilities.***
	- Paper changed to specify that the code is for meanline modelling of axial turbines, not turbomachinery in general.
	- The plotting, logging and export capabilities noted in the summary.
- ***Summary: Link to ‘documentation pages’ is not working***
	- Link fixed.
- ***Double-check the text with a grammar check, because I found spurious typos (e.g., line 14 strcuture).***
	- Checked.
### General comments on code
- ***I would recommend excluding the dev_projects folder from the main repository, as it seems to contain untested or under-development features, which may create confusion.***
	- Dev project moved to separate branch named "feature-dev".
- ***The code architecture is somewhat unclear, as it lacks common coding conventions such as a src folder and a docs folder. Additionally, some folders and files, such as .vscode, could be excluded and added to the .gitignore file to reduce clutter.***
	- In Python projects, the convention is to name the main source code folder after the package name (i.e. `turboflow`) rather than using a generic `src` folder.
	- docs folder is implemented.
	- .vscode file added to .gitignore.
- ***Overall, it is unclear whether the software includes additional functionalities beyond the meanline code for axial turbines. The documentation focuses solely on axial turbines, yet the example folder contains a case for a centrifugal compressor. Additionally, the src folder (turboflow) includes directories related to cycle calculations, which differ from the scope of the software (meanline code for turbomachinery).***
	- See documentation answer. The scope of the code is limited to axial turbines, which should be clearly stated now.
	- Cycle functions removed.
- ***It seems that metastable properties are implemented in fluid_properties.py. How can the code be instructed to consider either metastable or equilibrium properties? It does not appear that there is a line in the configuration file to specify this. Moreover, if I attempt to run fluid_properties.py to execute the calculation example under name == "main", it returns an error, likely related to issues with relative or absolute paths.***
	- The code uses equilibrium states only for this version. However, we are exploring the options to offer metastable properties in future versions.
	- Example function removed.
- ***The code does not generate a proper log file. While printing everything to the console is fine, it would be helpful to also have an option to save the output to a plain log file. This could be controlled via a flag in the input file to specify whether logging is desired. The same suggestion applies to the results.***
	- A logging feature is implemented.
	- Example of logging added in example folder and tutorial in documentation.
- ***I encountered issues running the Kofskey1974_stator validation test, which gave me two validation errors related to the config file.***
	- Redundant script removed.
	- This was simply a test script for debugging during the model development, and does not longer have relevance to the code.
- ***Please organize the folders more systematically, as there are numerous scattered files. For example, some figures are not placed in the figure folder within the validation tests, and there are old or redundant folders within other directories.***
	- Repository cleaned. With dev_scripts, cycles and .vscode scripts removed, the repository is cleaner. The validation folder is also cleaned and updated.
- ***While running the example for the two-stage turbine and plotting the results, I observed an unusual trend in the incidence loss, even though the plot matches the one provided in the repository. According to the log summary, the simulation was successful for all 40 points. Is this behavior (incidence loss going to zero for two distinct operating points) expected?***
	- Great catch! There was a slight mistake in the incidence loss calculation, with an absolute sign causing issues for this particular case. This had no impact on the other examples/validation cases.
- ***The code is intended to be general regarding thermodynamic modeling. However, in the calculation of the critical Mach number, it relies on an equation specific to perfect-gas modeling, as demonstrated by the use of the specific heat ratio. A similar dependency can be seen in the get_flow_capacity function.***
	- The critical mach number choking criterion is only one criterion for choking. Two other criterions are implemented (see documentation), and the critical_mass_flow_rate criterion is general and does not rely on perfect-gas relations.
- ***Can you confirm whether the plotting utility plot_axial_radial_plane works only for optimization? I attempted to use it with the two-stage turbine example in performance analysis mode to plot the meridional channel, but it returned an error. It would be helpful if the meridional channel could be plotted regardless of the specific solver used.***
	- It works for performance analysis. Example of use is added in example and tutorial in docs.
- ***I suggest expanding the examples available. Currently, the repository includes a single-stage and a two-stage turbine, both operating with air under ideal gas conditions. It would be valuable to include a case featuring a turbine operating with a fluid where non-ideal effects play a significant role.***
	- New case added using R125 as working fluid. The case features expansion from supercritical conditions, exhibiting significant non-ideal gas effects.
- ***Does the code handle supersonic cascades? Based on the documentation and the code itself, it seems that it does not, as there are no mentions of shock losses, supersonic deviation, or unique incidence. If this is the case, the software paper and documentation should explicitly report this limitation.***
	- The code does not handle cascades where the inlet and exit is supersonic.
	- It is stated in the documentation that the code  does not handle these cases.
	- The cascades can have supersonic exit, where shock losses are implemented through the loss model (see loss models in documentation).
