### 2025-02-21 #7588 @salehisaeed
First of all, I would like to acknowledge the authors' contribution to the field of **turbomachinery**. The package is well-structured, and I appreciate the effort in providing both open-source and proprietary references for context. The documentation is extensive, clear, and the provided examples cover a range of applications, which is very helpful for users looking to get started quickly.

That being said, I encountered some issues while testing the package, which I believe should be addressed to improve usability and ensure a smoother user experience.

I also want to acknowledge the review and comments provided by the other reviewer. Some of their points align with mine, and I will not repeat them here. Instead, I will focus on additional aspects that I encountered during my review.

---

## Installation and Environment
- Installed the package using `pip` on a **virtual environment in Python 3.11**. The installation went through without any issues.
- Suggestion: Avoid using **spaces in file and folder names** (for example in `validation_cases` and `examples`).

## Issues Encountered

Maybe I am missing something here, but for some reason, I was not able to run any of the provided tutorials in the `examples`. Here is a summary:

### 1. Compressor Example (`compressor_main.py`)
- The YAML configuration file does not seem to be read correctly. Possible missing information in the configuration file? Error encountered:

  ```
  5 validation errors for CentrifugalCompressor
  geometry.impeller.length_axial
    Field required [type=missing]
  geometry.impeller.length_meridional
    Field required [type=missing]
  geometry.impeller.impeller_length
    Extra inputs are not permitted [type=extra_forbidden]
  geometry.vaneless_diffuser.width_out
    Field required [type=missing]
  geometry.vaneless_diffuser.wall_divergence
    Extra inputs are not permitted [type=extra_forbidden]
  ```
  - Likely, the schema expected by `turboflow` does not match the provided YAML file.

### 2. The rest of examples, namely, `one-stage_main.py`, `R125_main.py`, and `two-stage_main.py` gave the following error:

  ```
  TypeError: create_logger() got an unexpected keyword argument 'to_console'
  ```
  - It seems `to_console` is no longer a valid argument in `tf.create_logger()`.

### 3. Validation Case (`Kofskey1972_one_stage`)
- The case ran smoothly, but the **plotting script failed** due to a misspelled filename.
  - Issue: The code references `experimental_data_Kofskey1972_1stage_interpolated.xlsx`, but the actual file name is spelled with lower case "k".

- I did not run the rest of the validation cases.

**Request:** Kindly double-check all provided examples to ensure they run without issues.

---

## General Feedback on the Paper
- The paper is well-written and maintains a clear and logical flow. The **acknowledgment of existing open-source and proprietary alternatives** is well done. However, the explanation of how this package differs from **other open-source meanline solvers** is a bit too general. Specifically, the differences between this package and **Brind (2024)** are not explicitly explained. Brind (2024) also provides a **Python-based open-source meanline simulation package** using an **equation-oriented approach** with **design capabilities**. It would be beneficial to include a clearer comparison between the two. What are the gaps in the previosu package that are filled with the current development?

## Additional Features
1. **Meridional Flow Contours**
   - In meanline (1D) simulation workflows, it is often possible to extract **contours of meridional flow**. Is that feature available in this package? It would be interesting for users to see such results.

2. **CFD Simulation Curves**
   - Such meanline simulations and design processes can always be a solid step before a full 3D CFD analysis. So, I was wondering if it is possible to extract the designed geometrical curves by **turboflow** for input to CFD mesh software (like TurboGrid), i.e., hub curve, shroud curve, and a number of blade profiles?

## Final Thoughts
Overall, this package is a **valuable addition** to the open-source turbomachinery community. Addressing the above issues will enhance usability and clarity for new users. **I will come back to the examples once we have resolved the issues.**
