### 2025-03-06 d695ef5
added optimization cases
