+++ dev_projects/optimization_algos/engineering_case.py
@@ -1,0 +1,114 @@
+
+import turboflow as tf
+import matplotlib.cm as cm
+import matplotlib.pyplot as plt
+import numpy as np
+from mpl_toolkits.axes_grid1.inset_locator import inset_axes, InsetPosition
+
+save_figs = False
+
+# Load files 
+filename_80 = "output/real_case_80p_2025-02-22_09-00-48.pkl" # 80 % centrifugal stress limit
+filename_85 = "output/real_case_85p_2025-02-22_08-59-17.pkl" # 85 % centrifugal stress limit
+filename_90 = "output/real_case_90p_2025-02-22_08-57-52.pkl" # 90 % centrifugal stress limit
+filename_95 = "output/real_case_95p_2025-02-22_08-56-22.pkl" # 95 % centrifugal stress limit
+filename_100 = "output/real_case_100p_2025-02-22_08-54-50.pkl" # 100 % centrifugal stress limit
+filename_ig = "output/initial_guess_2024-09-20_16-12-35.pkl" # Initial guess
+solver_80 = tf.load_from_pickle(filename_80)
+solver_85 = tf.load_from_pickle(filename_85)
+solver_90 = tf.load_from_pickle(filename_90)
+solver_95 = tf.load_from_pickle(filename_95)
+solver_100 = tf.load_from_pickle(filename_100)
+initial_guess = tf.load_from_pickle(filename_ig)
+
+# Get initial guess geometry
+ig_geometry = initial_guess.geometry.to_dict(orient = "list")
+ig_geometry["number_of_cascades"] = 2
+
+# Plot options
+colors = plt.get_cmap('magma')(np.linspace(0.1, 0.9, 2))
+
+# Plot axial-radial plane
+fig, ax = tf.plot_axial_radial_plane(ig_geometry, label = "Baseline", blade_color = "grey", plot_casing=False)
+fig, ax = tf.plot_axial_radial_plane(solver_90.problem.geometry, fig = fig, ax = ax, label = "Optimized", blade_color=colors[0], linestyle = "--", plot_rotation_ax= False, plot_casing=False)
+ylims = ax.get_ylim()
+ax.legend(handletextpad=0.5, frameon=False, loc = "lower left")
+ax.set_ylim([ylims[0], ylims[1]*1.15])
+
+# Plade profile as inset axis
+axins = inset_axes(ax, width="100%", height="100%", loc = "lower left")
+ip = InsetPosition(ax, [1.2, -0.1, 1.0, 1.0])  
+axins.set_axes_locator(ip)
+axins.set_aspect('equal', 'box')
+axins.spines['top'].set_visible(False)
+axins.spines['right'].set_visible(False)
+axins.spines['left'].set_visible(False)
+axins.spines['bottom'].set_visible(False)
+axins.yaxis.set_ticks([])
+axins.xaxis.set_ticks([])
+axins.grid(False)
+axins.tick_params(axis='x', which='both', top=False, labeltop=False)
+axins.tick_params(axis='y', which='both', right=False, labelright=False)
+axins.set_xlabel("Axial direction", labelpad=5)
+axins.set_ylabel("Tangential direction", labelpad=5)
+repeats = 3
+fig2, ax2 = tf.plot_view_axial_tangential(initial_guess.results, ig_geometry, fig = fig, ax = axins, repeats= repeats, color = "grey")
+fig2, ax2 = tf.plot_view_axial_tangential(solver_90.problem.results, solver_90.problem.geometry, fig = fig2, ax = ax2, repeats= repeats, color = colors[0])
+xlims = axins.get_xlim()
+ylims = axins.get_ylim()
+plt.annotate(
+        "",
[…]

[…]
