### 2025-03-06 6b31354
add centrifugal stress calculation
