+++ turboflow/plot_functions.py
@@ -385,7 +385,7 @@ def savefig_in_formats(
         fig.savefig(f"{path_without_extension}{ext}", bbox_inches="tight")
 
 
-def plot_axial_radial_plane(geometry):
+def plot_axial_radial_plane(geometry, linestyle = "-", blade_color = "k", plot_rotation_ax = True, plot_casing = True, label = "", fig = None, ax = None):
     r"""
     Plot the turbine geometry in an axial-radial plane.
 
@@ -447,60 +447,78 @@ def plot_axial_radial_plane(geometry):
     )
 
     # Initialize figure and axis object
-    fig, ax = plt.subplots()
+    if fig == None:
+        fig, ax = plt.subplots()
+        ax.grid(False)
+        ax.tick_params(which="both", direction="out", right=False)
+        ax.tick_params(which="minor", left=False)
+        ax.set_xticks([])
+        ax.spines["top"].set_visible(False)
+        ax.spines["right"].set_visible(False)
+        ax.spines["bottom"].set_visible(False)
+        ax.set_xlabel("Axis of rotation")
+        ax.set_ylabel("Radius")
+        ax.set_aspect("equal")
+        ax.set_ylim([0, 1.5 * max(y_tip) * ct + max(y_tip)])
 
-    # Plot upper casing
-    tip_clearance_extended = np.array(
-        [val for pair in zip(tip_clearance, tip_clearance) for val in pair]
-    )  # Extend tip clearance vector
-    casing_thickness = max(y_tip) * ct
-    x_casing = x
-    x_casing = np.insert(x_casing, 0, x[0] - (cl - 1) * x[-1] / 2)
-    x_casing = np.append(x_casing, x[-1] + (cl - 1) * x[-1] / 2)
+    if plot_casing:
+        # Plot upper casing
+        tip_clearance_extended = np.array(
+            [val for pair in zip(tip_clearance, tip_clearance) for val in pair]
+        )  # Extend tip clearance vector
+        casing_thickness = max(y_tip) * ct
+        x_casing = x
+        x_casing = np.insert(x_casing, 0, x[0] - (cl - 1) * x[-1] / 2)
+        x_casing = np.append(x_casing, x[-1] + (cl - 1) * x[-1] / 2)
 
-    y_casing_lower = y_tip + tip_clearance_extended
-    y_casing_lower = np.insert(y_casing_lower, 0, y_casing_lower[0])
-    y_casing_lower = np.append(y_casing_lower, y_casing_lower[-1])
-    y_casing_upper = np.ones(len(x_casing)) * (max(y_tip) + casing_thickness)
+        y_casing_lower = y_tip + tip_clearance_extended
+        y_casing_lower = np.insert(y_casing_lower, 0, y_casing_lower[0])
+        y_casing_lower = np.append(y_casing_lower, y_casing_lower[-1])
+        y_casing_upper = np.ones(len(x_casing)) * (max(y_tip) + casing_thickness)
 
-    ax.plot(x_casing, y_casing_lower, "k")
-    ax.plot(x_casing, y_casing_upper, "k")
-    ax.plot([x_casing[0], x_casing[0]], [y_casing_lower[0], y_casing_upper[0]], "k")
-    ax.plot([x_casing[-1], x_casing[-1]], [y_casing_lower[-1], y_casing_upper[-1]], "k")
-    ax.fill_between(x_casing, y_casing_lower, y_casing_upper, color="0.5")
+        ax.plot(x_casing, y_casing_lower, "k")
+        ax.plot(x_casing, y_casing_upper, "k")
+        ax.plot([x_casing[0], x_casing[0]], [y_casing_lower[0], y_casing_upper[0]], "k")
[…]
