+++ examples/two-stage_turbine/two-stage_main.py
@@ -9,7 +9,7 @@ import os
 import turboflow as tf
 
 # Define mode
-MODE = "performance_map" 
+MODE = "performance_analysis" 
 
 # Load configuration file
 CONFIG_FILE = os.path.abspath("two-stage_config.yaml")
@@ -36,6 +36,7 @@ if MODE == "performance_analysis":
     # Plot turbine geometry and velocity triangles
     fig, ax = tf.plot_functions.plot_axial_radial_plane(solvers[0].problem.geometry)
     fig, ax = tf.plot_functions.plot_velocity_triangles_planes(solvers[0].problem.results["plane"])
+    fig, ax = tf.plot_functions.plot_view_axial_tangential(solvers[0].problem)
 
 elif MODE == "performance_map":
 
