+++ examples/centrifugal_compressor/compressor_config.yaml
@@ -68,11 +68,12 @@ geometry:
     leading_edge_angle : (43.6+60)/2
     trailing_edge_angle : -24.5
     number_of_blades : 12
-    impeller_length : 45e-3
+    length_axial : 45e-3
+    length_meridional : 47e-3
     tip_clearance : 0.2e-3
     area_throat_ratio : 0.5
   vaneless_diffuser :
-    wall_divergence : 0
+    width_out : 6.29e-3
     radius_out : 150.22/2*1e-3
     radius_in : 143/2*1e-3
     width_in : 6.29e-3
