+++ paper/paper.md
@@ -73,8 +73,7 @@ The few open-source meanline models that do exist come with significant limitati
 | [@agromayor_preliminary_2019]      | 2019 | MATLAB (proprietary) | Equation-oriented     | Design optimization  |
 | [@brind_data-driven_2024]          | 2024 | Python               | Equation-oriented     | Design optimization  |
 
-The use of diverse programming languages, such as MATLAB and FORTRAN77, presents accessibility and compatibility issues. MATLAB-based models are proprietary, which limit their accessibility to those with MATLAB licenses.
-While legacy languages like FORTRAN77 might be more accessible, they fall short in terms of modern features and extensive community support. Consequently, models developed with these languages are less efficient to develop and less attractive to potential contributors, hampering development and collaboration. Furthermore, models adopting a sequential model formulation, solve sets of model equations sequentially through multiple nested iterations. This approach can lead to unreliable convergence and prolonged execution times due to the high number of model equations evaluations required. In contrast, an equation-oriented model formulation solves a larger set of equations simultaneously, enhancing reliability and computational efficiency. Lastly, the functionalities provided by existing models differ, with some focusing solely on performance analysis and others on design optimization, yet no single open-source model offers a comprehensive solution for both. 
+The use of diverse programming languages, such as MATLAB and FORTRAN77, presents accessibility and compatibility issues. MATLAB-based models, such as those by [@genrup_reduced-order_2005] and [@agromayor_preliminary_2019], are proprietary, which limit their accessibility to those with MATLAB licenses. While legacy languages like FORTRAN77, as used by [@denton_multallopen_2017], might be more accessible, they fall short in terms of modern features and extensive community support. Consequently, models developed with these languages are less efficient to develop and less attractive to potential contributors, hampering development and collaboration. Furthermore, models adopting a sequential model formulation, solve sets of model equations sequentially through multiple nested iterations. This approach can lead to unreliable convergence and prolonged execution times due to the high number of model equations evaluations required. In contrast, an equation-oriented model formulation solves a larger set of equations simultaneously, enhancing reliability and computational efficiency. Lastly, the functionalities provided by existing models differ, with some focusing solely on performance analysis and others on design optimization. For instance, [@brind_data-driven_2024] employs an equation-oriented formulation in Python but is specifically tailored for design optimization through an inverse design approach, making it unsuitable for predicting the performance of a given geometry. As a result, no single open-source model currently offers a comprehensive solution for both.
 
 TurboFlow addresses these gaps with a robust, open-source framework for meanline modeling. It combines performance analysis and design optimization within a flexible, modular 
 architecture, accommodating various submodels seamlessly. This flexibility allows for the integration of new submodels, giving users the options to tailor the analysis for their application. 
