+++ paper/paper.bib
@@ -131,7 +131,6 @@
 	author = {Brind, James},
 	date = {2024},
 	langid = {english},
-	file = {Brind - 2024 - Data-driven radial compressor design space mapping.pdf:C\:\\Users\\lasseba\\Zotero\\storage\\3JE7R6FF\\Brind - 2024 - Data-driven radial compressor design space mapping.pdf:application/pdf},
 }
 
 
