### 2025-03-17 72b54c7
add plotting function for blade profiles
