+++ turboflow/plot_functions.py
@@ -842,8 +842,6 @@ def plot_velocity_triangles_planes(plane):
 
     return fig, ax
 
-import matplotlib.pyplot as plt
-
 def combine_multiple_figures(figures, labels=None, markers=None, linestyles=None, colors=None, y_labels = None):
     """
     Combine multiple matplotlib figures into a single figure with custom labels, markers, linestyles, and colors.
@@ -1152,7 +1150,349 @@ def plot_convergence_process(data, plot_together=True, colors=None, linestyles=N
     return figs, ax
 
 
+def NURBS(P, p, U, u, W):
+    # ...
+    u = np.atleast_1d(u)
+    W = np.atleast_1d(W)
 
+    Nt = len(u)
+    dim, nn = P.shape  # nn = n+1 control points
+    n = nn - 1
+    m = n + p + 1      # total B-spline segments for initial p=0
+
+    # ** We now allocate with 'm' columns instead of 'nn': **
+    N = np.zeros((Nt, m))
+
+    # First step (p=0)
+    for i in range(m):
+        N[:, i] = ((u >= U[i]) & (u <= U[i+1])).astype(float)
+
+    current_p = 1
+    while m > (n + 1):
+        m -= 1
+        N_bis = np.zeros((Nt, m))
+
+        for i in range(m):
+            denom1 = (U[i + current_p] - U[i])
+            denom2 = (U[i + current_p + 1] - U[i + 1])
+            
+            if denom1 != 0:
+                n1 = ((u - U[i]) / denom1) * N[:, i]
+            else:
+                n1 = 0.0
+
+            if denom2 != 0:
+                n2 = ((U[i + current_p + 1] - u) / denom2) * N[:, i + 1]
+            else:
+                n2 = 0.0
+
+            n1 = np.nan_to_num(n1)
+            n2 = np.nan_to_num(n2)
+            N_bis[:, i] = n1 + n2
+
+        N = N_bis
+        current_p += 1
+
+    # At this point, N has shape (Nt, n+1).
+    # Compute rational basis
+    W_mat = np.tile(W, (Nt, 1))
+    numerator = N * W_mat
+    denominator = np.sum(numerator, axis=1, keepdims=True)
+    denominator = np.where(denominator == 0.0, 1e-14, denominator)
+    R = numerator / denominator
+
+    # Evaluate final curve in R^dim
+    r = R @ P.T  # shape (Nt, dim)
+
+    return r, R, N
+
+
+
+def plot_blade_profile(blade_parameters):
+    """
+    Python version of the MATLAB function [x, y] = plot_blade_profile(blade_parameters).
+
+    Given a dictionary `blade_parameters` with:
+      chord, stagger, theta_in, theta_out,
+      wedge_in, wedge_out, radius_in, radius_out,
+      d1, d2, d3, d4
+    it computes the shape of a 2D blade profile (x, y) by connecting
+    four NURBS segments:
+      (1) Leading edge arc (between points 1 and 2)
+      (2) Spline from point 2 to point 3
+      (3) Trailing edge arc (between points 3 and 4)
+      (4) Spline from point 4 back to point 1
+    
+    Returns:
+    --------
+    x, y : 1D numpy arrays
+        The concatenated coordinates of the blade in the plane.
+    """
+
+    # -- 1) Unpack parameters --
+    chord      = blade_parameters["chord"]        # Blade chord
+    stagger    = blade_parameters["stagger"]      # Stagger angle
+    theta_in   = blade_parameters["theta_in"]     # Inlet camber angle
+    theta_out  = blade_parameters["theta_out"]    # Outlet camber angle
[…]
