### 2025-03-19 58e0969
update compressor dev scripts
