+++ dev_projects/centrifugal_compressor/validation_plots.py
@@ -6,6 +6,14 @@ import numpy as np
 
 impeller = "A" # A, O
 save_figs = True
+fontsize = 16
+plt.rcParams['font.size'] = fontsize  # Replace 14 with the desired font size
+plt.rcParams.update({
+    'axes.labelsize': fontsize,    # Font size for x and y labels
+    'xtick.labelsize': fontsize,   # Font size for x tick labels
+    'ytick.labelsize': fontsize,   # Font size for y tick labels
+    'legend.fontsize': fontsize,   # Font size for legends
+})
 
 files_exp = {
     "A" : "validation_cases\Eckardt\eckardt_impeller_A\eckardt_impeller_A.xlsx",
@@ -34,9 +42,33 @@ files_sim_incl =  {
     }
 }
 
+files_sim_condensed = {
+    "A" : {
+        "oh" : {
+            "PR" : "output\performance_analysis_2024-11-27_12-17-50.xlsx",
+            "eta" : "output\performance_analysis_2024-11-27_12-22-24.xlsx"
+        },
+        "zhang" : {
+
+            "PR" : "output\performance_analysis_2024-12-03_11-06-19.xlsx",
+            "eta" : "output\performance_analysis_2024-12-03_11-04-49.xlsx"
+        }
+    },
+    "O" : {
+        "oh" : {
+            "PR" : "output\performance_analysis_2024-11-27_12-24-47.xlsx",
+            "eta" : "output\performance_analysis_2024-11-27_12-24-47.xlsx"
+        },
+        "zhang" : {
+            "PR" : "output\performance_analysis_2024-12-03_11-02-32.xlsx",
+            "eta" : "output\performance_analysis_2024-12-03_11-02-32.xlsx" 
+        }
+    }
+}
+
 filename_exp = files_exp[impeller]
-filename_oh = files_sim_incl[impeller]["oh"]
-filename_zhang = files_sim_incl[impeller]["zhang"]
+filename_oh = files_sim_condensed[impeller]["oh"]["PR"]
+filename_zhang = files_sim_condensed[impeller]["zhang"]["PR"]
 
 data_exp = tf.load_data(filename_exp, ["pr_tt", "eta_tt"])
 data_oh = tf.load_data(filename_oh)
@@ -44,8 +76,8 @@ data_zhang = tf.load_data(filename_zhang)
 
 # Plot simulation data
 omegas = [10e3, 12e3, 14e3, 16e3]
-color_map = 'viridis'
-colors = [plt.get_cmap(color_map)(i / (len(omegas) - 1)) for i in range(len(omegas))]
+color_map = 'magma'
+colors = plt.get_cmap(color_map)(np.linspace(0.2, 0.8, len(omegas)))
 subsets = ["omega"] + list(2*np.pi/60*np.array(omegas))
 
 # Mass flow rate
@@ -86,11 +118,11 @@ linestyle_handles = [
 
 # ax2.legend(np.array(omegas)/1e3, ncols = 1, loc = 'best')
 color_legend = ax1.legend(handles = color_handles)
-linestyle_legend = ax1.legend(handles=linestyle_handles, loc='upper center', bbox_to_anchor=(0.5, 1.15), ncol=2, frameon=False)
+linestyle_legend = ax1.legend(handles=linestyle_handles, loc='upper center', bbox_to_anchor=(0.5, 1.2), ncol=2, frameon=False)
 ax1.add_artist(color_legend)
 # ax1.add_artist(linestyle_legend)
-# ax1.set_xlim([2.0, 8.0])
-ax1.set_ylim([1.2, 2.4])
+ax1.set_xlim([2.1, 7.1])
+ax1.set_ylim([1.2, 2.6])
[…]
