### 2025-03-24 #7588 @lasseband
Hi @salehisaeed,

Thank you for reviewing our code and for your helpful comments. We have finally managed to address the identified issues, and our responses are provided below.

- **Suggestion: Avoid using spaces in file and folder names (for example in `validation_cases` and `examples`).**
	* Folders and files are renamed to avoid any spaces.
- **Compressor example: The YAML configuration file does not seem to be read correctly. Possible missing information in the configuration file?**
	- This issue arose because the PyPI release was not updated after the previous revision. The release is now updated, and the examples should work correctly.
- **Rest of examples: It seems `to_console` is no longer a valid argument in `tf.create_logger()`.**
	- This error comes from that the PyPI release was not updated after the previous revision. It is now updated, and the examples should work properly.
- **Validation case (`Kofskey1972_one_stage`): The case ran smoothly, but the plotting script failed due to a misspelled filename.**
	- The filename is corrected.
- **Request: Kindly double-check all provided examples to ensure they run without issues**
	- All examples have been double-checked and should now run without issues.
- **General Feedback on the Paper: The explanation of how this package differs from other open-source meanline solvers is a bit too general. Specifically, the differences between this package and Brind (2024) are not explicitly explained. Brind (2024) also provides a Python-based open-source meanline simulation package using an equation-oriented approach with design capabilities. It would be beneficial to include a clearer comparison between the two. What are the gaps in the previous package that are filled with the current development?**
	- The section comparing TurboFlow with existing open-source software has been revised to provide a clearer explanation of how TurboFlow differs from other alternatives:

	  *The use of diverse programming languages, such as MATLAB and FORTRAN77, presents accessibility and compatibility issues. MATLAB-based models, such as those by [@genrup_reduced-order_2005] and [@agromayor_preliminary_2019], are proprietary, which limit their accessibility to those with MATLAB licenses. While legacy languages like FORTRAN77, as used by [@denton_multallopen_2017], might be more accessible, they fall short in terms of modern features and extensive community support. Consequently, models developed with these languages are less efficient to develop and less attractive to potential contributors, hampering development and collaboration. Furthermore, models adopting a sequential model formulation, solve sets of model equations sequentially through multiple nested iterations. This approach can lead to unreliable convergence and prolonged execution times due to the high number of model equations evaluations required. In contrast, an equation-oriented model formulation solves a larger set of equations simultaneously, enhancing reliability and computational efficiency. Lastly, existing models vary in functionality. The model from [@genrup_reduced-order_2005] is dedicated to performance analysis, whereas [@agromayor_preliminary_2019] focuses on design optimization. On the other hand, [@brind_data-driven_2024] and [@denton_multallopen_2017] employ an inverse design methodology, where the geometry is derived from specified performance parameters (e.g., flow coefficient, loading coefficient, degree of reaction), rendering them unsuitable for performance prediction for a given geometry. Consequently, no single open-source model currently provides a comprehensive solution that integrates both performance analysis and design optimization.*
- **Additional Features: Meridional Flow Contours**
	- A function for plotting the blade profile has been added to the package. This function utilizes B-splines to visualize the blade profile based on the geometrical variables available from the meanline analysis.
	- The plotting functionality has also been included in the tutorial section of the documentation.
- **Additional Features: CFD Simulation Curves**
	- We agree that the ability to extract coordinates at the hub and shroud for CFD input would be a valuable addition. While we hope to implement this in the future, it will not be included in the JOSS publication.
