### 2025-03-29 6bed734
fix typo in plot y-label
