+++ examples/R125/R125_plots.py
@@ -29,7 +29,7 @@ fig1, ax1 = tf.plot_functions.plot_lines(
     x_key="PR_ts",
     y_keys=["efficiency_ts"],
     xlabel="Total-to-static pressure ratio [$p_{0, \mathrm{in}}/p_\mathrm{out}$]",
-    ylabel="Mass flow rate [kg/s]",
+    ylabel="Efficiency [%]",
     colors='k',
     filename = 'mass_flow_rate',
     outdir = "figures",
