+++ pyproject.toml
@@ -21,6 +21,9 @@ PyYAML = "^6.0.1"
 toml = "^0.10.2"
 pydantic = "^2.7.2"
 cycler = "^0.12.1"
+jinja2 = "^3.1.6"
+urllib3 = "^2.3.0"
+certifi = "^2025.1.31"
 
 [tool.poetry.group.dev.dependencies]
 ipython = "^8.24.0"
