### 2025-03-29 40c859d
use slsqp instead of ipopt
