+++ examples/one-stage_turbine/one-stage_config.yaml
@@ -154,8 +154,8 @@ design_optimization:
 
   radius_type : "constant_mean"
   solver_options: 
-    library: pygmo 
-    method: ipopt
+    library: scipy 
+    method: slsqp
     derivative_method: 2-point 
     derivative_abs_step: 1e-6 
     print_convergence: True 
