+++ dev_projects/automatic_differentiation/property_calculations/perfect_gas_test/perfect_gas_props_functions.py
@@ -1,0 +1,309 @@
+# This script contains perfect gas equations for different input cases 
+# This is used to calculate the thermodynamic properties and is imported in perfect_gas_props script
+
+import jax.numpy as jnp
+from jax import grad
+import jax
+jax.config.update("jax_enable_x64", True)
+
+
+# Constants
+R = 287.0  # Specific gas constant for air (J/(kg*K))
+gamma = 1.41  # Specific heat ratio for air
+T_ref = 288.15  # Reference temperature (K)
+P_ref = 101306.33  # Reference pressure (Pa)
+s_ref = 1659.28  # Reference entropy (J/(kg*K))
+myu_ref = 1.789e-5  # Reference dynamic viscosity (Kg/(m*s))
+S_myu = 110.56  # Sutherland's constant for viscosity (K)
+k_ref = 0.0241  # Reference thermal conductivity (W/(m*K))
+S_k = 194  # Sutherland's constant for thermal conductivity (K)
+
+PROPERTY_ALIAS = {
+    # "P": "p",
+    # "rho": "rhomass",
+    # "d": "rhomass",
+    # "u": "umass",
+    # "h": "hmass",
+    # "s": "smass",
+    # "cv": "cvmass",
+    # "cp": "cpmass",
+    "speed_sound": "a",
+    # "Z": "compressibility_factor",
+    # "mu": "viscosity",
+    # "k": "conductivity",
+} 
+
+def specific_heat(R, gamma):
+        """Calculate specific heats at constant pressure and volume."""
+        cp = (gamma * R) / (gamma - 1)  # J/(kg*K)
+        cv = cp / gamma  # J/(kg*K)
+        return cp, cv
+
+def calculate_properties_hs(h, s):
+    """Calculate all thermodynamic properties given enthalpy and Pressure."""
+
+    def temperature(h, s):
+        """Calculate temperature from enthalpy."""
+        cp, cv = specific_heat(R, gamma)
+        return jnp.max(jnp.asarray([1.0, h / cp])) # Temperature in Kelvin
+        # return h / cp
+
+    def pressure(h, s):
+        """Calculate pressure using enthalpy and entropy."""
+        T = temperature(h, s)
+        cp, cv = specific_heat(R, gamma)
+        return P_ref * jnp.exp(((cp * jnp.log(T / T_ref)) - (s - s_ref)) / R)
+
+    def density(h, s):
+        """Calculate density using enthalpy and entropy."""
+        T = temperature(h, s)
+        P = pressure(h, s)
+        return P / (R * T)  # Density in kg/m^3
+
+    def viscosity(h, s):
+        """Calculate dynamic viscosity using Sutherland's formula."""
+        T = temperature(h, s)
+        return myu_ref * ((T / T_ref) ** (3/2)) * ((T_ref + S_myu) / (T + S_myu))
+
+    def thermal_conductivity(h, s):
+        """Calculate thermal conductivity using Sutherland's formula."""
+        T = temperature(h, s)
+        return k_ref * ((T / 273) ** (3/2)) * ((273 + S_k) / (T + S_k))
+
+    def speed_of_sound(h, s):
+        T = temperature(h, s)
+        return jnp.sqrt(gamma*R*T) # Speed of sound in m/s
+
+    properties = {"T": temperature(h, s),
+            "p": pressure(h, s),
+            "d": density(h, s), 
+            "h": h,
+            "s": s,
+            "mu": viscosity(h, s), 
+            "k": thermal_conductivity(h, s), 
+            "a": speed_of_sound(h, s),
+            "gamma": gamma}
+    
+    # # Add properties as aliases
+    # for key, value in PROPERTY_ALIAS.items():
+    #     properties[key] = properties[value]
+
+    # Add properties as aliases
[…]
