### 2025-04-23 abd6119
add restart optimization script
