+++ dev_projects/optimization_algos/kofskey1972_1stage.yaml
@@ -153,13 +153,16 @@ design_optimization:
       normalize : True
   radius_type : "constant_mean"
   solver_options: 
-    library: pygmo   
-    method: snopt
+    library: pygmo    
+    method: ipopt
     derivative_method: 2-point 
     derivative_abs_step: 1e-6 
     print_convergence: True 
     plot_convergence: False 
     update_on: "gradient" 
+    options : 
+      tau_min : 0.001
+      mu_init : 0.001
 
 
   
