### 2025-05-26 cade9a8
add pysolverview as external package, improve multistart functionality, fix incompatibilities
