+++ turboflow/axial_turbine/design_optimization.py
@@ -4,8 +4,11 @@ import CoolProp as cp
 import datetime
 import os
 import yaml
+import copy
 import pickle
-from .. import pysolver_view as psv
+import pysolver_view as psv
+
+
 from .. import utilities as utils
 from . import geometry_model as geom
 from . import flow_model as flow
@@ -48,6 +51,7 @@ def compute_optimal_turbine(
     out_filename=None,
     out_dir="output",
     export_results=True,
+    x0=None,
 ):
     r"""
     Calculate the optimal turbine configuration based on the specified optimization problem.
@@ -69,78 +73,124 @@ def compute_optimal_turbine(
     """
 
     # Initialize problem object
+
     problem = CascadesOptimizationProblem(config)
 
+    # Determine initial guesses
+    if x0 is not None:
+        if isinstance(x0, np.ndarray):
+            initial_guesses = [x0]
+        elif isinstance(x0, list) and all(isinstance(x, np.ndarray) for x in x0):
+            initial_guesses = x0
+        else:
+            raise ValueError("x0 must be a NumPy array or a list of NumPy arrays.")
+    else:
+        ig = problem.initial_guesses
+        if isinstance(ig, np.ndarray):
+            initial_guesses = [ig]
+        elif isinstance(ig, list) and all(isinstance(x, np.ndarray) for x in ig):
+            initial_guesses = ig
+        else:
+            raise ValueError("problem.initial_guesses must be a NumPy array or a list of NumPy arrays.")
+
     # Perform initial function call to initialize problem
-    # This populates the arrays of equality and inequality constraints
-    # TODO: it might be more intuitive to create a new method called initialize_problem() that generates the initial guess and evaluates the fitness() function with it
-    problem.fitness(problem.initial_guesses[0])
+    problem.fitness(initial_guesses[0])
+
+
+    # # Perform initial function call to initialize problem
+    # # This populates the arrays of equality and inequality constraints
+    # # TODO: it might be more intuitive to create a new method called initialize_problem() that generates the initial guess and evaluates the fitness() function with it
+    # initial_guesses = x0 if x0 is not None else problem.initial_guesses
+    # problem.fitness(initial_guesses[0])
 
     # Load solver configuration
     solver_config = config["design_optimization"]["solver_options"]
-
     solver_container = SolverContainer()
     success = []
-    for ig in problem.initial_guesses:
+    out_dir_base = copy.copy(out_dir)
+    for i, ig in enumerate(initial_guesses):
         # Initialize solver object using keyword-argument dictionary unpacking
         solver = psv.OptimizationSolver(problem, **solver_config)
         
-         # Solve optimization problem for initial guess x0
+        # Solve optimization problem for initial guess x0
         solver.solve(ig)
         solver_container.solver_container.append(solver)
         success.append(solver.success)
 
-    dfs = {
-        "operation point": pd.DataFrame(
-            {key: pd.Series(val) for key, val in problem.boundary_conditions.items()}
-        ),
-        "overall": pd.DataFrame(problem.results["overall"], index=[0]),
-        "plane": problem.results["plane"],
-        "cascade": problem.results["cascade"],
-        "stage": problem.results["stage"],
-        "geometry": pd.DataFrame(
-            {key: pd.Series(val) for key, val in problem.geometry.items()}
-        ),
-        "solver": pd.DataFrame(
-            {
-                "completed": pd.Series(True, index=[0]),
-                "success": pd.Series(solver.success, index=[0]),
-                "message": pd.Series(solver.message, index=[0]),
-                "grad_count": solver.convergence_history["grad_count"],
-                "func_count": solver.convergence_history["func_count"],
-                "func_count_total": solver.convergence_history["func_count_total"],
[…]
