+++ turboflow/pysolver_view/nonlinear_system.py
@@ -1,620 +1,0 @@
-import os
-import time
-import copy
-import logging
-import warnings
-import numpy as np
-import matplotlib.pyplot as plt
-
-from datetime import datetime
-from abc import ABC, abstractmethod
-from matplotlib.ticker import MultipleLocator
-from scipy.optimize import root
-
-from . import numerical_differentiation
-from .pysolver_utilities import savefig_in_formats
-
-SOLVER_OPTIONS = ["hybr", "lm"]
-
-
-class NonlinearSystemSolver:
-    r"""
-    Solver class for nonlinear systems of equations.
-
-    The solver is designed to handle system of nonlinear equations of the form:
-
-    .. math::
-
-        F(x) = 0
-
-    where :math:`F: \mathbb{R}^n \rightarrow \mathbb{R}^n` is a vector-valued
-    function of the vector :math:`x`.
-
-    The class interfaces with the `root` method from `scipy.optimize` to solve the equations and provides a structured framework for initialization, solution monitoring, and post-processing.
-
-    Parameters
-    ----------
-    problem : NonlinearSystemProblem
-        An instance of a problem defining the system of equations to be solved.
-    method : str, optional
-        Method to be used by scipy's root for solving the nonlinear system. Available solvers are:
-
-        - :code:`hybr`:  Uses MINPACK's 'hybrd' method, which is is a modification of Powell's hybrid method.
-        - :code:`lm`: The Levenberg-Marquardt method, which blends the steepest descent and the Gauss-Newton methods.
-
-        The choice between :code:`hybr` and :code:`lm` largely depends on the specifics of the problem at hand.
-        The :code:`hybr` usually requires less gradient evaluations and it is often faster when analytic gradients are not available.
-        It is advisable to experiment with both methods to determine the most appropriate choice for a given problem.
-
-        Defaults to 'hybr'.
-    tol : float, optional
-        Tolerance for the solver termination. Defaults to 1e-9.
-    max_iter : integer, optional
-        Maximum number of function evaluations for the solver termination. Defaults to 100.
-    options : dict, optional
-        Additional options to be passed to scipy's root.
-    derivative_method : str, optional
-        Finite difference method to be used when the problem Jacobian is not provided. Defaults to '2-point'
-    derivative_abs_step : float, optional
-        Finite difference absolute step size to be used when the problem Jacobian is not provided. Defaults to 1e-6
-    print_convergence : bool, optional
-        If True, displays the convergence progress. Defaults to True.
-    plot_convergence : bool, optional
-        If True, plots the convergence progress. Defaults to False.
-    logger : logging.Logger, optional
-        Logger object to which logging messages will be directed. Defaults to None.
-    update_on : str, optional
-        Specifies if the convergence report should be updated on a new function evaluations ("function") or on gradient evaluations ("gradient"). Defaults to "function".
-
-
-    Methods
-    -------
-    solve(x0)
-        Solve the system of nonlinear equations using the specified initial guess `x0`.
-    residual(x)
-        Evaluate the vector of residuals of the at a given point `x`.
-    gradient(x)
-        Evaluate the Jacobian of the system at a given point `x`.
-    print_convergence_history()
-        Print the convergence history of the problem.
-    plot_convergence_history()
-        Plot the convergence history.
-
-    """
-
-    def __init__(
-        self,
-        problem,
-        method="hybr",
-        tolerance=1e-6,
-        max_iterations=100,
-        options={},
-        derivative_method="2-point",
-        derivative_abs_step=1e-6,
-        print_convergence=True,
-        plot_convergence=False,
-        plot_scale="log",
-        logger=None,
-        update_on="function",
-        callback_func=None,
-    ):
-        # Initialize class variables
[…]
