### 2025-05-26 8356c50
update kofskey config for multistart
