### 2025-05-26 50f3ad0
change choking criterion clipping to make it more robust
