+++ dev_projects/optimization_algos/main.py
@@ -12,8 +12,8 @@ import turboflow as tf
 CASE = 2
 
 # Load configuration file
-CONFIG_FILE = os.path.abspath("kofskey_constrained.yaml")
-# CONFIG_FILE = os.path.abspath("kofskey1972_1stage.yaml")
+# CONFIG_FILE = os.path.abspath("kofskey_constrained.yaml")
+CONFIG_FILE = os.path.abspath("kofskey1972_1stage.yaml")
 config = tf.load_config(CONFIG_FILE, print_summary=False)
 
 if CASE == 0:
@@ -31,7 +31,7 @@ elif CASE == 1:
 elif CASE == 2:
     operation_points = config["operation_points"]
     solver = tf.compute_optimal_turbine(config, export_results=False)
-    tf.save_to_pickle(solver, filename = f"real_case_90p", out_dir = "output")
+    tf.save_to_pickle(solver, filename = f"pickle_multistart_slsqp", out_dir = "output")
 
 elif CASE == 3:
 
