+++ dev_projects/optimization_algos/kofskey1972_1stage.yaml
@@ -62,6 +62,84 @@ design_optimization:
     scale : 1
     type : maximize
   variables :
+    # v_in:
+    #   value: 0.227150609503
+    #   lower_bound: 0.001
+    #   upper_bound: 0.5
+    # w_out:
+    #   value: [0.756016831137, 0.714178597145]
+    #   lower_bound: [0.1, 0.1]
+    #   upper_bound: [1.0, 1.0]
+    # s_out:
+    #   value: [0.023935947025, 0.072366299371]
+    #   lower_bound: [0.0, 0.0]
+    #   upper_bound: [0.32, 0.32]
+    # beta_out:
+    #   value: [0.865041724256, 0.16150395975]
+    #   lower_bound: [0.72, 0.06]
+    #   upper_bound: [0.94, 0.28]
+    # v_crit_in:
+    #   value: [0.4, 0.4]
+    #   lower_bound: [0.1, 0.1]
+    #   upper_bound: [1.0, 1.0]
+    # beta_crit_throat:
+    #   value: [0.83, 0.17]
+    #   lower_bound: [0.72, 0.06]
+    #   upper_bound: [0.94, 0.28]
+    # w_crit_throat:
+    #   value: [0.954641310154, 0.89015497855]
+    #   lower_bound: [0.1, 0.1]
+    #   upper_bound: [1.0, 1.0]
+    # s_crit_throat:
+    #   value: [0.040063613132, 0.096142058811]
+    #   lower_bound: [0.0, 0.0]
+    #   upper_bound: [0.32, 0.32]
+    # specific_speed:
+    #   value: 0.713756
+    #   lower_bound: 0.5
+    #   upper_bound: 1.5
+    # blade_jet_ratio:
+    #   value: 0.466501
+    #   lower_bound: 0.3
+    #   upper_bound: 0.8
+    # hub_tip_ratio_in:
+    #   value: [0.715999, 0.715999]
+    #   lower_bound: [0.6, 0.6]
+    #   upper_bound: [0.9, 0.9]
+    # hub_tip_ratio_out:
+    #   value: [0.715999, 0.67484]
+    #   lower_bound: [0.6, 0.6]
+    #   upper_bound: [0.9, 0.9]
+    # aspect_ratio:
+    #   value: [1.28555, 1.402149]
+    #   lower_bound: [1.0, 1.0]
+    #   upper_bound: [2.0, 2.0]
+    # pitch_chord_ratio:
+    #   value: [0.699312, 0.584804]
+    #   lower_bound: [0.5, 0.5]
+    #   upper_bound: [1.1, 1.1]
+    # trailing_edge_thickness_opening_ratio:
+    #   value: [0.066889, 0.068007]
+    #   lower_bound: [0.05, 0.05]
+    #   upper_bound: [0.2, 0.2]
+    # leading_edge_angle:
+    #   value: [0.5, 0.66444444]
+    #   lower_bound: [0.11111111, 0.41666667]
+    #   upper_bound: [0.58333333, 0.88888889]
+    # gauging_angle:
+    #   value: [0.86601513, 0.16024573]
+    #   lower_bound: [0.80555556, 0.08333333]
+    #   upper_bound: [0.91666667, 0.19444444]
+    # throat_location_fraction:
+    #   value: [0.833333333333, 0.833333333333]
+    # leading_edge_diameter:
+    #   value: [0.00254, 0.00162]
+    # leading_edge_wedge_angle:
+    #   value: [50.0, 50.0]
+    # tip_clearance:
+    #   value: [0.0, 0.0003]
+    # cascade_type:
+    #   value: [stator, rotor]
     specific_speed : # Limited 
       value : 0.713756
       lower_bound : 0.5
@@ -153,17 +231,14 @@ design_optimization:
       normalize : True
[…]
