+++ dev_projects/automatic_differentiation/property_calculations/Lookup_tables/generate_lookup_table.py
@@ -1,0 +1,388 @@
+#%% Initialization
+
+import os,psutil #Look at the number of physical cores.
+
+#NOTE: On modern machines might want to use the number of high performance
+#cores, because different core architectures can cause problems in parallel
+#processing
+
+NCORES=psutil.cpu_count(logical=False) #Before loading jax, force it to see the CPU count we want
+os.environ["XLA_FLAGS"]="--xla_force_host_platform_device_count=%d"%NCORES
+import jax
+jax.config.update("jax_enable_x64", True) #By default jax uses 32 bit, for scientific computing we need 64 bit precision
+import jax.numpy as jnp
+import pickle
+import pandas as pd
+
+#To read more regarding the automatic parallelization in jax
+#https://jax.readthedocs.io/en/latest/sharded-computation.html
+
+from jax.experimental import mesh_utils
+from jax.sharding import Mesh, PartitionSpec , NamedSharding
+
+import CoolProp.CoolProp as cp
+import matplotlib.pyplot as plt
+from matplotlib import rcParams
+from matplotlib.lines import Line2D
+import numpy as np
+import time
+from jax_bicubic_HEOS_interpolation_1 import (
+    compute_bicubic_coefficients_of_ij,
+    bicubic_interpolant,
+    inverse_interpolant_scalar_DP,
+    inverse_interpolant_scalar_hD,
+)
+import turboflow as tf
+
+float64=jnp.dtype('float64')
+complex128=jnp.dtype('complex128')
+
+#%% Grid and Fluid definition
+
+# Defining the Grid
+N=50       # Grid size in h direction
+M=50      # Grid size in log(P) direction
+name="CO2" # Fluid selection
+
+# Grid boundaries
+hmin=200000  # units - J/kg
+hmax=600000 # units - J/kg
+Pmin=20*1e5 # units - Pa
+Pmax=200*1e5 # units - Pa
+Lmin=jnp.log(Pmin) # Log of P
+Lmax=jnp.log(Pmax) # Log of P
+
+# Number of random points to test the error
+Npoints=50000
+
+#Number of times to repeat the computation. used to stabilize the timing of  
+#very fast computations when using time.time()
+Nrepeats=int(1e6/Npoints)
+
+fluid = tf.Fluid(name)
+
+#%% Properties to be calculated
+
+properties = {
+    'T': 'T',       # Temperature [K]
+    'D': 'D',       # Density [kg/m³]
+    'S': 'S',       # Entropy [J/kg/K]
+    'mu': 'V',      # Viscosity [Pa·s]
+    'k': 'L',       # Thermal conductivity [W/m/K]
+}
+
+# Create the grid of (h, P)
+h_vals = jnp.linspace(hmin, hmax, N)
+P_vals = jnp.linspace(Lmin, Lmax, M)
+h_grid, P_grid = jnp.meshgrid(h_vals, jnp.exp(P_vals), indexing='ij')
+
+# Evaluate each property on the grid using CoolProp
+raw_property_grids = {}
+
+print("⚙️ Generating raw property data using CoolProp...")
+for key, coolprop_symbol in properties.items():
+    prop_grid = np.zeros((N, M))
+    for i in range(N):
+        for j in range(M):
+            h = h_vals[i]
+            P = jnp.exp(P_vals[j])
+            try:
+                val = cp.PropsSI(coolprop_symbol, 'H', h, 'P', P, name)
+            except:
+                val = np.nan  # or use extrapolation/default value
+            prop_grid[i, j] = val
+    raw_property_grids[key] = prop_grid
+
+print("✅ Raw property grids created.")
+
+## Check raw_property_grids here to see if the values are proper or not ##
+
[…]
