### 2025-06-25 1e1d978
Update paper.md
Change Github to GitHub
