+++ paper/paper.md
@@ -41,7 +41,7 @@ TurboFlow provides access to advanced equations of state for real gas fluid prop
 TurboFlow comes with comprehensive documentation, including installation guides, tutorials, model descriptions, and a complete API reference. This extensive resource ensures that users can easily learn how to use the package and apply it effectively in their projects. For more details, visit the [documentation pages](https://turbo-sim.github.io/turboflow/).
 Additionally, the package includes preconfigured examples that demonstrate performance analysis and design optimization. These examples serve as practical guides for users to apply TurboFlow to their own projects. Additionally, these examples showcase the post-processing capabilities, including plotting, logging, and export utilities for result interpretation and analysis.
 
-The package source code is hosted in a [Github repository](https://github.com/turbo-sim/TurboFlow) [@turboflow]. Through GitHub Actions, an automated test suite is included, which checks the functionality of the performance analysis and design optimization, as well as all submodels. It enables continuous integration, ensuring that code changes are systematically tested and validated. This comprehensive testing framework provides confidence that the code works as expected, maintaining the reliability of the package with each update.
+The package source code is hosted in a [GitHub repository](https://github.com/turbo-sim/TurboFlow) [@turboflow]. Through GitHub Actions, an automated test suite is included, which checks the functionality of the performance analysis and design optimization, as well as all submodels. It enables continuous integration, ensuring that code changes are systematically tested and validated. This comprehensive testing framework provides confidence that the code works as expected, maintaining the reliability of the package with each update.
 
 With these features, TurboFlow should present a reliable and flexible tool for researchers and engineers within the field of turbomachinery. 
 
