### 2025-06-25 #7588 @lasseband
Hi @philipcardiff,

Great — thanks for thoroughly reviewing the paper! I’ve merged your pull request.
