### 2025-06-26 #7588 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
