### 2025-07-09 #7588 @lasseband
Hi @kyleniemeyer! Sorry for the late response. I have merged your changes regarding reference formatting now. Let me know if there is anything else I should do.
